% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{argument_distance}
\alias{argument_distance}
\title{Handles Distance Argument}
\usage{
argument_distance(distance)
}
\arguments{
\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}
}
\value{
Character vector with distance names
}
\description{
Handles Distance Argument
}
\examples{

distance_arg(
  distance = c(
    "euclidean",
    "euc"
   )
  )

}
