% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_by_similarity.R
\name{arrange_by_similarity}
\alias{arrange_by_similarity}
\title{Arrange Sequences by Similarity}
\usage{
arrange_by_similarity(
  x = NULL,
  distance = "euclidean",
  diagonal = FALSE,
  weighted = FALSE,
  ignore_blocks = FALSE,
  paired_samples = FALSE
)
}
\arguments{
\item{x}{(required, list of matrices) list of input matrices generated with [prepare_sequences()].}

\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}

\item{diagonal}{(optional, logical vector). If TRUE, diagonals are included in the computation of the cost matrix. Default: FALSE.}

\item{weighted}{(optional, logical vector) If TRUE, diagonal is set to TRUE, and diagonal cost is weighted by a factor of 1.414214. Default: FALSE.}

\item{ignore_blocks}{(optional, logical vector). If TRUE, blocks of consecutive path coordinates are trimmed to avoid inflating the psi distance. Default: FALSE.}

\item{paired_samples}{(optional, logical vector) If TRUE, time-series are compared row wise and no least-cost path is computed. Default: FALSE.}
}
\value{
Vector with names of `x` ordered from higher to lower similarity with the previous sequences. The first two sequences are the most similar ones.
}
\description{
Arranges sequences in a list by similarity. This is an internal function for the [combine()] function.
}
