% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_breaks.R
\name{auto_breaks}
\alias{auto_breaks}
\title{Auto Breaks for Matrix Plotting Functions}
\usage{
auto_breaks(m = NULL, n = 100)
}
\arguments{
\item{m}{(required, numeric matrix) distance or cost matrix generated by [distance_matrix()] or [cost_matrix()], but any numeric matrix will work. Default: NULL}

\item{n}{(required, integer) number of colors to compute the breaks for. Default: 100}
}
\value{
Numeric vector
}
\description{
Auto Breaks for Matrix Plotting Functions
}
