% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_distance.R
\name{auto_distance}
\alias{auto_distance}
\title{Sum of Auto Distances Between Adjacent Cases in a Sequence}
\usage{
auto_distance(x = NULL, path = NULL, distance = "euclidean")
}
\arguments{
\item{path}{(optional, data frame) result of [cost_path_trim()], only required if blocks in the least cost path were ignored.}

\item{distance}{(optional, character string) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}

\item{y}{(required, data frame or matrix) a sequence.}
}
\value{
Sum of auto distances
}
\description{
Sum of Auto Distances Between Adjacent Cases in a Sequence
}
\examples{
#TODO
}
