% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{auto_distance_cpp}
\alias{auto_distance_cpp}
\title{Auto Distance}
\usage{
auto_distance_cpp(m, distance = "euclidean")
}
\arguments{
\item{m}{(required, numeric matrix) multivariate time series.}

\item{distance}{(optional, character string) name or abbreviation of the
distance method. Valid values are in the columns "names" and "abbreviation"
of the dataset \code{distances}. Default: "euclidean".}
}
\value{
Matrix of distances between 'y' (rows) and 'x' (columns).
}
\description{
Computes the distance matrix between the rows of two matrices
\code{y} and \code{x} with the same number of columns and arbitrary numbers of rows.
NA values should be removed before using this function.
If the selected distance function is \link{distance_chi_cpp}, pairs of zeros should
be either removed or replaced with pseudo-zeros (i.e. 0.00001).
}
