% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_sum.R
\name{auto_sum}
\alias{auto_sum}
\title{Auto Sum}
\usage{
auto_sum(a = NULL, b = NULL, path = NULL, method = "euclidean")
}
\arguments{
\item{a}{(required, data frame) a time series. Default: NULL.}

\item{b}{(required, data frame) a time series. Default: NULL.}

\item{path}{(required, data frame) dataframe produced by [cost_path()]. Only required when [trim_blocks()] has been applied to `path`. Default: NULL.}

\item{method}{(optional, character string) name of the distance metric. Valid entries are:
\itemize{
  \item "euclidean" and "euc" (Default).
  \item "manhattan" and "man".
  \item "chi.
  \item "hellinger" and "hel".
  \item "chebyshev" and "che".
  \item "canberra" and "can".
  \item "cosine" and "cos".
  \item "russelrao" and "rus".
  \item "jaccard" and "jac".
}}
}
\value{
Named vector with the auto sums of `a` and `b`.
}
\description{
Computes the sum of distances between consecutive samples in two multivariate time-series under comparison. Required to compute the measure of dissimilarity \code{psi} (Birks and Gordon 1985). Distances can be computed via the methods "manhattan", "euclidean", "chi", and "hellinger.
}
\examples{
data(sequenceA, sequenceB)

a <- sequenceA |>
  na.omit() |>
  as.matrix()

b <- sequenceB |>
  na.omit() |>
  as.matrix()

ab_sum <- auto_sum(
  a = a,
  b = b,
  method = "manhattan"
)

ab_sum

}
