% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{auto_sum_cpp}
\alias{auto_sum_cpp}
\title{Auto-sum of Two Sequences}
\usage{
auto_sum_cpp(a, b, path, f)
}
\arguments{
\item{a}{(required, numeric matrix).}

\item{b}{(required, numeric matrix) of same number of columns as 'a'.}

\item{path}{(required, data frame) dataframe produced by [cost_path()]. Default: NULL}

\item{f}{(required, function) a distance function. One of:}
}
\value{
Numeric.
}
\description{
Sum of the the cumulative auto-sum of two sequences. This is
a key component of the psi computation.
}
