% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{auto_sum_no_path_cpp}
\alias{auto_sum_no_path_cpp}
\title{Auto-sum of Two Paired Sequences}
\usage{
auto_sum_no_path_cpp(a, b, distance = "euclidean")
}
\arguments{
\item{a}{(required, numeric matrix).}

\item{b}{(required, numeric matrix) of same number of columns as 'a'.}

\item{distance}{(optional, character string) name or abbreviation of the
distance method. Valid values are in the columns "names" and "abbreviation"
of the dataset `distances`. Default: "euclidean".}
}
\value{
Auto sum of distances.
}
\description{
Sum of the the cumulative auto-sum of two sequences with paired samples. This is
a key component of the psi computation.
}
