% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{auto_sum_path_cpp}
\alias{auto_sum_path_cpp}
\title{Auto-sum of Two Sequences}
\usage{
auto_sum_path_cpp(x, y, path, distance = "euclidean")
}
\arguments{
\item{x}{(required, numeric matrix) of same number of columns as 'y'.}

\item{y}{(required, numeric matrix).}

\item{path}{(required, data frame) dataframe produced by \code{\link[=cost_path_cpp]{cost_path_cpp()}}.
Default: NULL}

\item{distance}{(optional, character string) name or abbreviation of the
distance method. Valid values are in the columns "names" and "abbreviation"
of the dataset \code{distances}. Default: "euclidean".}
}
\value{
Numeric.
}
\description{
Sum of the the cumulative auto-sum of two sequences. This is
a key component of the psi computation.
}
