% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_args_distance}
\alias{check_args_distance}
\title{Check Distance Argument}
\usage{
check_args_distance(distance = NULL)
}
\arguments{
\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}
}
\value{
Character vector with distance names
}
\description{
Check Distance Argument
}
\examples{

check_args_distance(
  distance = c(
    "euclidean",
    "euc"
   )
  )

}
