% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_args_matrix}
\alias{check_args_matrix}
\title{Checks Input Matrix}
\usage{
check_args_matrix(m = NULL, arg_name = "m")
}
\arguments{
\item{m}{(required, matrix) distance or cost matrix resulting from [psi_dist_matrix()] or [psi_cost_matrix()]. Default: NULL}

\item{arg_name}{(optional, character string) name of the argument being checked. Default: NULL}
}
\value{
Argument m.
}
\description{
Checks Input Matrix
}
