% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_args_path}
\alias{check_args_path}
\title{Checks Least Cost Path}
\usage{
check_args_path(path = NULL, arg_name = "path")
}
\arguments{
\item{path}{(required, data frame) least cost path generated with [psi_cost_path()]. This data frame must have the attribute `type == "cost_path`, and must have been computed from the same sequences used to compute the matrix `m`. Default: NULL.}

\item{arg_name}{(optional, character string) name of the argument being checked. Default: NULL}
}
\value{
Least cost path
}
\description{
Checks Least Cost Path
}
