% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities_coordinates}
\alias{cities_coordinates}
\title{Coordinates of 100 Major Cities}
\format{
sf data frame with 5 columns and 100 rows.
}
\usage{
data(cities_coordinates)
}
\description{
City coordinates for the dataset `cities_temperature`.
}
\keyword{datasets}
