% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_names.R
\name{clean_names}
\alias{clean_names}
\title{Clean Names}
\usage{
clean_names(names = NULL, first_to_upper = FALSE, length = NULL)
}
\arguments{
\item{names}{(required, character vector) names to clean. Default: NULL}

\item{first_to_upper}{(optional, logical) If TRUE, the first letter is set to upper case. Default: FALSE}

\item{length}{(optional, integer) If provided, names are abbreviated to this length via [abbreviate()].}
}
\value{
Character vector
}
\description{
Generic function to clean time series names and column names.
}
\examples{

names(mis)

clean_names(
  names = names(mis),
  length = 3
)

}
