% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia.R
\docType{data}
\name{climateLong}
\alias{climateLong}
\title{Dataframe with palaeoclimatic data.}
\format{dataframe with 6 columns and 800 rows.}
\usage{
data(climateLong)
}
\description{
A dataframe containing 800 simulated samples of palaeoclimate data at 1 ky temporal resolution with the following columns:
}
\details{
\itemize{
  \item \emph{age} in kiloyears before present (ky BP).
  \item \emph{temperatureAverage} average annual temperature in Celsius degrees.
  \item \emph{rainfallAverage} average annual precipitation in milimetres per day (mm/day).
  \item \emph{temperatureWarmestMonth} average temperature of the warmest month, in Celsius degrees.
  \item \emph{temperatureColdestMonth} average temperature of the coldest month, in Celsius degrees.
  \item \emph{oxigenIsotope} delta O18, global ratio of stable isotopes in the sea floor, see \url{http://lorraine-lisiecki.com/stack.html} for further details.
}
}
\keyword{datasets}
