% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climateShort}
\alias{climateShort}
\title{Dataframe with palaeoclimatic data.}
\format{
dataframe with 5 columns and 11 rows.
}
\usage{
data(climateShort)
}
\description{
A dataframe containing 11 simulated samples of palaeoclimate data at 1 ky temporal resolution with the following columns:
}
\details{
\itemize{
  \item \emph{temperatureAverage} average annual temperature in Celsius degrees.
  \item \emph{rainfallAverage} average annual precipitation in milimetres per day (mm/day).
  \item \emph{temperatureWarmestMonth} average temperature of the warmest month, in Celsius degrees.
  \item \emph{temperatureColdestMonth} average temperature of the coldest month, in Celsius degrees.
  \item \emph{oxigenIsotope} delta O18, global ratio of stable isotopes in the sea floor, see \url{http://lorraine-lisiecki.com/stack.html} for further details.
}
}
\keyword{datasets}
