% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hclust_optimizer.R
\name{cluster_hclust_optimizer}
\alias{cluster_hclust_optimizer}
\title{Optimal Number of Clusters for Hierarchical Clustering}
\usage{
cluster_hclust_optimizer(d = NULL, method = NULL)
}
\arguments{
\item{d}{(required, matrix) distance matrix typically resulting from \code{\link[=distantia_matrix]{distantia_matrix()}}, but any other square matrix should work. Default: NULL}

\item{method}{(optional, character string) Argument of \code{\link[stats:hclust]{stats::hclust()}} defining the agglomerative method. One of: "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC). Unambiguous abbreviations are accepted as well.}
}
\value{
Data frame with number of clusters and their respective mean silhouette widths.
}
\description{
Performs a parallelized grid search to find the number of clusters maximizing the overall silhouette width of the clustering solution (see \code{\link[=cluster_silhouette]{cluster_silhouette()}}). When \code{method = NULL}, the optimization also includes all methods available in \code{\link[stats:hclust]{stats::hclust()}} in the grid search. This function supports parallelization via \code{\link[future:plan]{future::plan()}} and a progress bar generated by the \code{progressr} package (see Examples).
}
\examples{
#daily covid prevalence in California counties
data("covid_prevalence")

#load as tsl and aggregate to monthly data to accelerate example execution
tsl <- tsl_initialize(
  x = covid_prevalence,
  id_column = "county",
  time_column = "date"
) |>
  tsl_aggregate(
    new_time = "months",
    fun = sum
  )

if(interactive()){
  #plotting first three time series
  tsl_plot(
    tsl = tsl[1:3],
    guide_columns = 3
    )
}

#parallelization and progress bar
#for large datasets, parallelization accelerates cluster optimization
future::plan(
 future::multisession,
 workers = 2 #set to parallelly::availableWorkers() - 1
)

#progress bar
progressr::handlers(global = TRUE)

#compute dissimilarity matrix
psi_matrix <- distantia(
  tsl = tsl,
  lock_step = TRUE
) |>
  distantia_matrix()

#optimize hierarchical clustering
hclust_optimization <- cluster_hclust_optimizer(
  d = psi_matrix
)

#best solution in first row
head(hclust_optimization)

#disable parallelization
future::plan(
  future::sequential
)
}
