% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_silhouette.R
\name{cluster_silhouette}
\alias{cluster_silhouette}
\title{Silhouette Width of a Clustering Solution}
\usage{
cluster_silhouette(cluster = NULL, distance_matrix = NULL, median = FALSE)
}
\arguments{
\item{cluster}{(required, clustering object) Result of [stats::kmeans()]. Default: NULL}

\item{distance_matrix}{(required, matrix) Distance matrix typically resulting from [distantia_matrix()]. Default: NULL}

\item{median}{(optional, logical) If TRUE, the median of the silhouette widths is returned. Default: FALSE}
}
\value{
If median = FALSE, data frame with silhouette widths of the clustering, and numeric indicating the median silhouette width otherwise.
}
\description{
The silhouette width is a measure of how similar an object is to its own cluster (cohesion) compared to other clusters (separation). It provides a way to assess the quality of clustering results.

A silhouette width close to +1 indicates that the object is well matched to its own cluster and poorly matched to neighboring clusters.

A silhouette width close to 0 indicates that the object is between two neighboring clusters.

A silhouette width close to -1 indicates that the object is likely to be assigned to the wrong cluster. This suggests that there may be overlapping clusters or that the clustering algorithm has produced poor results.

This code was adapted from https://svn.r-project.org/R-packages/trunk/cluster/R/silhouette.R
}
