% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_continuous.R
\name{color_continuous}
\alias{color_continuous}
\title{Default Continuous Color Palette}
\usage{
color_continuous(n = 5, palette = "Zissou 1", rev = FALSE)
}
\arguments{
\item{n}{(required, integer) number of colors to generate. Default = NULL}

\item{palette}{(required, character string) Argument \code{palette} of \code{\link[grDevices:hcl.colors]{grDevices::hcl.colors()}}. Default: "Zissou 1"}

\item{rev}{(optional, logical) If TRUE, the color palette is reversed. Default: FALSE}
}
\value{
color vector
}
\description{
Uses the function \code{\link[grDevices:hcl.colors]{grDevices::hcl.colors()}} to generate a continuous color palette.
}
\examples{

color_continuous(n = 20)

}
\seealso{
Other internal_plotting:
\code{\link[=color_discrete]{color_discrete()}},
\code{\link[=utils_color_breaks]{utils_color_breaks()}},
\code{\link[=utils_line_color]{utils_line_color()}},
\code{\link[=utils_line_guide]{utils_line_guide()}},
\code{\link[=utils_matrix_guide]{utils_matrix_guide()}},
\code{\link[=utils_matrix_plot]{utils_matrix_plot()}}
}
\concept{internal_plotting}
