% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination_path.R
\name{combination_path}
\alias{combination_path}
\title{Least Cost Path to Combine Sequences}
\usage{
combination_path(x, distance = "euclidean")
}
\arguments{
\item{x}{(required, list of matrices) list of input matrices generated with [prepare_sequences()]. For this function, only the first two elements in `x` are used.}

\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}
}
\value{
A least cost path
}
\description{
This is an internal function of [combine()].
}
