% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combination of Sequences via Least Cost Path}
\usage{
combine(x = NULL, distance = "euclidean")
}
\arguments{
\item{x}{(required, list of matrices) list of input matrices generated with [ts_prepare()].}

\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}
}
\value{
Data frame with composite sequence and columns identifying the origin of each new row.
}
\description{
Combines two sequences by arranging their cases as defined in their orthogonal least cost path. When more than two sequences are provided in `x`, all sequences are first ordered according to their psi scores. Then, the first two sequences (the more similar ones) are combined first, and the remaining ones are added to the composite sequence one by one.
}
