% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cost_matrix_diagonal_weighted_cpp}
\alias{cost_matrix_diagonal_weighted_cpp}
\title{(C++) Compute Orthogonal and Weighted Diagonal Least Cost Matrix from a Distance Matrix}
\usage{
cost_matrix_diagonal_weighted_cpp(dist_matrix)
}
\arguments{
\item{dist_matrix}{(required, distance matrix). Distance matrix.}
}
\value{
Least cost matrix.
}
\description{
Computes the least cost matrix from a distance matrix.
Weights diagonals by a factor of 1.414214 with respect to orthogonal paths.
}
\seealso{
Other Rcpp_matrix: 
\code{\link{cost_matrix_diagonal_cpp}()},
\code{\link{cost_matrix_orthogonal_cpp}()},
\code{\link{distance_lock_step_cpp}()},
\code{\link{distance_matrix_cpp}()}
}
\concept{Rcpp_matrix}
