% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cost_matrix_weighted_diag_cpp}
\alias{cost_matrix_weighted_diag_cpp}
\title{Least Cost Matrix with Weighted Diagonals}
\usage{
cost_matrix_weighted_diag_cpp(dist_matrix)
}
\arguments{
\item{dist_matrix}{(required, distance matrix). Distance matrix.}
}
\value{
Least cost matrix.
}
\description{
Computes the least cost matrix from a distance matrix.
Weights diagonals by a factor of 1.414214 with respect to orthogonal paths.
}
