% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cost_path_diag_cpp}
\alias{cost_path_diag_cpp}
\title{Least Cost Path Considering Diagonals}
\usage{
cost_path_diag_cpp(dist_matrix, cost_matrix)
}
\arguments{
\item{dist_matrix}{(required, distance matrix). Distance matrix.}

\item{cost_matrix}{(required, numeric matrix) Cost matrix generated from the distance matrix.}
}
\value{
A data frame with least-cost path coordiantes.
}
\description{
Computes the least cost matrix from a distance matrix.
Considers diagonals during computation of least-costs. In case of ties,
diagonals are favored.
}
