% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cost_path_slotting_cpp}
\alias{cost_path_slotting_cpp}
\title{Least Cost Path for Sequence Slotting}
\usage{
cost_path_slotting_cpp(dist_matrix, cost_matrix)
}
\arguments{
\item{dist_matrix}{(required, distance matrix). Distance matrix.}

\item{cost_matrix}{(required, numeric matrix) Cost matrix generated from the distance matrix.}
}
\value{
A data frame with least-cost path coordiantes.
}
\description{
Computes the least cost matrix from a distance matrix.
Considers diagonals during computation of least-costs. This version differs
from cost_path_cpp() in the way it solves ties. In the case of a tie,
cost_path_cpp() uses the first neighbor satisfying the minimum distance condition,
while cost_path_slotting_cpp() selects the neighbor that changes the axis
of movement within the least cost matrix.
}
