% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_path_sum.R
\name{cost_path_sum}
\alias{cost_path_sum}
\title{Sum of Distances in Least-cost Path}
\usage{
cost_path_sum(path = NULL)
}
\arguments{
\item{path}{(required, data frame) dataframe produced by [cost_path()]. Default: NULL}
}
\value{
Sum of distances.
}
\description{
Sum of Distances in Least-cost Path
}
\examples{

data(sequenceA, sequenceB)

a <- sequenceA |>
  na.omit() |>
  as.matrix()

b <- sequenceB |>
  na.omit() |>
  as.matrix()

d <- distance_matrix(a, b, distance = "euclidean")

m <- cost_matrix(dist_matrix = d)

path <- cost_path(
  dist_matrix = d,
  cost_matrix = m
  )

cost_path_sum(path = path)

}
