% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_path_trim.R
\name{cost_path_trim}
\alias{cost_path_trim}
\title{Trims Blocks from a Least Cost Path}
\usage{
cost_path_trim(path = NULL)
}
\arguments{
\item{path}{(required, data frame) dataframe produced by [cost_path()]. Default: NULL}
}
\value{
A data frame with a least-cost path with no blocks.
}
\description{
Removes blocks (straight segments of the path that appear in highly dissimilar regions of the sequences) from least-cost paths. Blocks inflate psi values when two sequences are similar but have very different numbers of rows.
}
\examples{

a <- sequenceA |>
  na.omit() |>
  as.matrix()

b <- sequenceB |>
  na.omit() |>
  as.matrix()

dm <- distance_matrix(
  a = a,
  b = b
  )

cm <- cost_matrix(
  dist_matrix = dm
  )

path <- cost_path(
  dist_matrix = dm,
  cost_matrix = cm
)

nrow(path)

if(interactive()){
  plotMatrix(cm, path)
}

path <- cost_path_trim(path = path)

nrow(path)

if(interactive()){
  plotMatrix(cm, path)
}

}
