% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_coordinates}
\alias{covid_coordinates}
\title{Coordinates of the Covid Dataset.}
\format{
sf data frame with 4 columns and 36 rows
}
\usage{
data(covid_coordinates)
}
\description{
Coordinates of the Covid Dataset.
}
\keyword{datasets}
