% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_counties}
\alias{covid_counties}
\title{County Coordinates of the Covid Prevalence Dataset}
\format{
sf data frame with county polygons and census data from \url{https://www.counties.org/data-and-research}.
}
\usage{
data(covid_counties)
}
\description{
County Coordinates of the Covid Prevalence Dataset
}
\seealso{
Other example_data: 
\code{\link{albatross}},
\code{\link{cities_coordinates}},
\code{\link{cities_temperature}},
\code{\link{covid_prevalence}},
\code{\link{eemian_coordinates}},
\code{\link{eemian_pollen}},
\code{\link{fagus_coordinates}},
\code{\link{fagus_dynamics}},
\code{\link{honeycomb_climate}},
\code{\link{honeycomb_polygons}}
}
\concept{example_data}
\keyword{datasets}
