% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_prevalence}
\alias{covid_prevalence}
\title{Covid19 dataset.}
\format{
Dataframe with 3 columns and 51048 rows
}
\usage{
data(covid_prevalence)
}
\description{
Dataset with Covid19 prevalence in California counties between 2020 and 2024. [Source](https://healthdata.gov/State/COVID-19-Time-Series-Metrics-by-County-and-State-A/cr6j-rwfz/about_data).
}
\keyword{datasets}
