% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_bray_curtis_cpp}
\alias{distance_bray_curtis_cpp}
\title{(C++) Bray-Curtis Distance Between Two Vectors}
\usage{
distance_bray_curtis_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as \code{x}.}
}
\value{
numeric
}
\description{
Computes the Bray-Curtis distance, suitable for species abundance data.
}
\examples{
distance_bray_curtis_cpp(x = runif(100), y = runif(100))
}
\seealso{
Other Rcpp_distance_methods: 
\code{\link{distance_canberra_cpp}()},
\code{\link{distance_chebyshev_cpp}()},
\code{\link{distance_chi_cpp}()},
\code{\link{distance_cosine_cpp}()},
\code{\link{distance_euclidean_cpp}()},
\code{\link{distance_hamming_cpp}()},
\code{\link{distance_hellinger_cpp}()},
\code{\link{distance_jaccard_cpp}()},
\code{\link{distance_manhattan_cpp}()},
\code{\link{distance_russelrao_cpp}()},
\code{\link{distance_sorensen_cpp}()}
}
\concept{Rcpp_distance_methods}
