% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_canberra_cpp}
\alias{distance_canberra_cpp}
\title{Canberra Distance Between Two Binary Vectors}
\usage{
distance_canberra_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as `x`.//' @return Canberra distance between x and y.}
}
\description{
Computes the Canberra distance between two binary vectors.
}
\examples{
distance_canberra_cpp(c(0, 1, 0, 1), c(1, 1, 0, 0))
}
