% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_chebyshev_cpp}
\alias{distance_chebyshev_cpp}
\title{Chebyshev Distance Between Two Vectors}
\usage{
distance_chebyshev_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as `x`.}
}
\value{
Chebyshev distance between x and y.
}
\description{
Computed as: \code{max(abs(x - y))}. Cannot handle NA values.
}
\examples{
distance_chebyshev_cpp(x = runif(100), y = runif(100))
}
