% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_distance.R
\name{distance_chi}
\alias{distance_chi}
\title{Chi Distance Between Two Vectors}
\usage{
distance_chi(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as `x`.}
}
\description{
Chi Distance Between Two Vectors
}
\examples{

distance_chi(
  x = runif(100),
  y = runif(100)
)

}
