% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_chi_cpp}
\alias{distance_chi_cpp}
\title{Chi Distance Between Two Vectors}
\usage{
distance_chi_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as `x`.}
}
\value{
Chi distance between x and y.
}
\description{
Chi Distance Between Two Vectors
}
\details{
Cannot handle NA values.
}
\examples{
distance_chi_cpp(x = runif(100), y = runif(100))
}
