% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_cosine_cpp}
\alias{distance_cosine_cpp}
\title{Cosine Dissimilarity Between Two Vectors}
\usage{
distance_cosine_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as \code{x}.}
}
\value{
Cosine dissimilarity between x and y.
}
\description{
Computes the cosine dissimilarity between two numeric vectors.
}
\examples{
distance_cosine_cpp(c(0.2, 0.4, 0.5), c(0.1, 0.8, 0.2))
}
