% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance_euclidean}
\alias{distance_euclidean}
\title{Euclidean Distance Between Two Vectors}
\usage{
distance_euclidean(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as `x`.}
}
\description{
Euclidean Distance Between Two Vectors
}
\examples{

distance_euclidean(
  x = runif(100),
  y = runif(100)
)

}
