% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_euclidean_cpp}
\alias{distance_euclidean_cpp}
\title{Euclidean Distance Between Two Vectors}
\usage{
distance_euclidean_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as \code{x}.}
}
\value{
Euclidean distance between x and y.
}
\description{
Computed as: \code{sqrt(sum((x - y)^2)}. Cannot handle NA values.
}
\examples{
distance_euclidean_cpp(x = runif(100), y = runif(100))
}
