% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_hamming_cpp}
\alias{distance_hamming_cpp}
\title{(C++) Hamming Distance Between Two Binary Vectors}
\usage{
distance_hamming_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as \code{x}.}
}
\value{
numeric
}
\description{
Computes the Hamming distance between two binary vectors.
}
\examples{
distance_hamming_cpp(c(0, 1, 0, 1), c(1, 1, 0, 0))
}
\seealso{
Other Rcpp_distance_methods: 
\code{\link{distance_canberra_cpp}()},
\code{\link{distance_chebyshev_cpp}()},
\code{\link{distance_chi_cpp}()},
\code{\link{distance_cosine_cpp}()},
\code{\link{distance_euclidean_cpp}()},
\code{\link{distance_hellinger_cpp}()},
\code{\link{distance_jaccard_cpp}()},
\code{\link{distance_manhattan_cpp}()},
\code{\link{distance_russelrao_cpp}()}
}
\concept{Rcpp_distance_methods}
