% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance_hellinger}
\alias{distance_hellinger}
\title{Hellinger Distance Between Two Vectors}
\usage{
distance_hellinger(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as `x`.}
}
\description{
Hellinger Distance Between Two Vectors
}
\examples{

distance_hellinger(
  x = runif(100),
  y = runif(100)
)

}
