% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_hellinger_cpp}
\alias{distance_hellinger_cpp}
\title{(C++) Hellinger Distance Between Two Vectors}
\usage{
distance_hellinger_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as \code{x}.}
}
\value{
numeric
}
\description{
Computed as: \code{sqrt(1/2 * sum((sqrt(x) - sqrt(y))^2))}.
Cannot handle NA values.
}
\examples{
distance_hellinger_cpp(x = runif(100), y = runif(100))
}
\seealso{
Other Rcpp_distance_methods: 
\code{\link{distance_canberra_cpp}()},
\code{\link{distance_chebyshev_cpp}()},
\code{\link{distance_chi_cpp}()},
\code{\link{distance_cosine_cpp}()},
\code{\link{distance_euclidean_cpp}()},
\code{\link{distance_hamming_cpp}()},
\code{\link{distance_jaccard_cpp}()},
\code{\link{distance_manhattan_cpp}()},
\code{\link{distance_russelrao_cpp}()}
}
\concept{Rcpp_distance_methods}
