% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_lock_step_cpp}
\alias{distance_lock_step_cpp}
\title{Sum of Pairwise Distances Between Paired Sequences}
\usage{
distance_lock_step_cpp(y, x, distance = "euclidean")
}
\arguments{
\item{y}{(required, numeric matrix).}

\item{x}{(required, numeric matrix) of same number of columns and rows as 'y'.}

\item{distance}{(optional, character string) name or abbreviation of the
distance method. Valid values are in the columns "names" and "abbreviation"
of the dataset `distances`. Default: "euclidean".}
}
\value{
Vector of distances between 'y' (rows) and 'x' (columns).
}
\description{
Computes the lock-step sum of distances between two time series.
NA values should be removed before using this function.
If the selected distance function is "chi" or "cosine", pairs of zeros should
be either removed or replaced with pseudo-zeros (i.e. 0.00001).
}
