% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance_manhattan}
\alias{distance_manhattan}
\title{Manhattan Distance Between Two Vectors}
\usage{
distance_manhattan(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as `x`.}
}
\value{
Manhattan distance.
}
\description{
Manhattan Distance Between Two Vectors
}
\examples{

distance_manhattan(
  x = runif(100),
  y = runif(100)
)

}
