% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_manhattan_cpp}
\alias{distance_manhattan_cpp}
\title{Manhattan Distance Between Two Vectors}
\usage{
distance_manhattan_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as \code{x}.}
}
\value{
Manhattan distance between x and y.
}
\description{
Computed as: \code{sum(abs(x - y))}. Cannot handle NA values.
}
\examples{
distance_manhattan_cpp(x = runif(100), y = runif(100))
}
