% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_matrix_cpp}
\alias{distance_matrix_cpp}
\title{Distance Matrix}
\usage{
distance_matrix_cpp(x, y, distance = "euclidean")
}
\arguments{
\item{x}{(required, numeric matrix) of same number of columns as 'y'.}

\item{y}{(required, numeric matrix).}

\item{distance}{(optional, character string) name or abbreviation of the
distance method. Valid values are in the columns "names" and "abbreviation"
of the dataset \code{distances}. Default: "euclidean".}
}
\value{
Matrix of distances between 'y' (rows) and 'x' (columns).
}
\description{
Computes the distance matrix between the rows of two matrices
\code{y} and \code{x} with the same number of columns and arbitrary numbers
of rows.
NA values should be removed before using this function.
If the selected distance function is "chi" or "cosine", pairs of zeros should
be either removed or replaced with pseudo-zeros (i.e. 0.00001).
}
