% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_methods.R
\name{distance_methods}
\alias{distance_methods}
\title{Distance Between Two Numeric or Binary Vectors}
\usage{
distance_methods(x = NULL, y = NULL, method = "euclidean")
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as `x`.}

\item{method}{(optional, character string) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `methods`. Default: "euclidean".}
}
\value{
A distance value.
}
\description{
Computes the distance between two numeric or binary vectors.
}
\examples{

distance_methods(
  x = runif(100),
  y = runif(100),
  method = "euclidean"
)

}
\author{
Blas Benito <blasbenito@gmail.com>
}
