% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_pairwise_cpp}
\alias{distance_pairwise_cpp}
\title{Distance Matrix}
\usage{
distance_pairwise_cpp(a, b, method)
}
\arguments{
\item{a}{(required, numeric matrix).}

\item{b}{(required, numeric matrix) of same number of columns and rows as 'a'.}

\item{method}{(optional, character string) name of the distance metric. Valid entries are:
\itemize{
\item "euclidean" and "euc" (Default).
\item "manhattan" and "man".
\item "chi.
\item "hellinger" and "hel".
\item "chebyshev" and "che".
\item "canberra" and "can".
\item "cosine" and "cos".
\item "russelrao" and "rus".
\item "jaccard" and "jac".
}}
}
\value{
Vector of distances between 'a' (rows) and 'b' (columns).
}
\description{
Computes the vector of distances between paired rows in two sequences of the same length.
NA values should be removed before using this function.
If the selected distance function is "chi" or "cosine", pairs of zeros should
be either removed or replaced with pseudo-zeros (i.e. 0.00001).
}
