% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_russelrao_cpp}
\alias{distance_russelrao_cpp}
\title{(C++) Russell-Rao Distance Between Two Binary Vectors}
\usage{
distance_russelrao_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric). Binary vector of 1s and 0s.}

\item{y}{(required, numeric) Binary vector of 1s and 0s of same length as \code{x}.}
}
\value{
numeric
}
\description{
Computes the Russell-Rao distance between two binary vectors.
}
\examples{
distance_russelrao_cpp(c(0, 1, 0, 1), c(1, 1, 0, 0))
}
\seealso{
Other Rcpp_distance_methods: 
\code{\link{distance_bray_curtis_cpp}()},
\code{\link{distance_canberra_cpp}()},
\code{\link{distance_chebyshev_cpp}()},
\code{\link{distance_chi_cpp}()},
\code{\link{distance_cosine_cpp}()},
\code{\link{distance_euclidean_cpp}()},
\code{\link{distance_hamming_cpp}()},
\code{\link{distance_hellinger_cpp}()},
\code{\link{distance_jaccard_cpp}()},
\code{\link{distance_manhattan_cpp}()},
\code{\link{distance_sorensen_cpp}()}
}
\concept{Rcpp_distance_methods}
