% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceMatrix.R
\name{distanceMatrix}
\alias{distanceMatrix}
\title{Computes distance matrix among the samples of two multivariate time-series.}
\usage{
distanceMatrix <- function(
  sequence.A = NULL,
  sequence.B = NULL,
  time.column = NULL,
  exclude.columns = NULL,
  method = "manhattan"
  )
}
\arguments{
\item{sequence.A}{dataframe, a multivariate time-series.}

\item{sequence.B}{dataframe, a multivariate time-series.}

\item{exclude.columns}{character string or character vector with column names in \code{sequences}, or \code{squence.A} and \code{sequence.B} to be excluded from the analysis.}

\item{method}{character string naming a distance metric. Valid entries are: "manhattan", "euclidean", "chi", and "hellinger". Invalid entries will throw an error.}

\item{time.colum}{character string, name of the column with time/depth/rank data. The data in this column is not modified.}
}
\value{
A matrix with the distances among samples of both sequences.
}
\description{
Computes a distance matrix among the samples of two multivariate time-series. Distances can be computed through the methods "manhattan", "euclidean", "chi", and "hellinger", and are implemented in the function \code{\link{distance}}.
}
\details{
Distances are computed as:
\itemize{
\item \code{manhattan}: \code{d <- sum(abs(x - y))}
\item \code{euclidean}: \code{d <- sqrt(sum((x - y)^2))}
\item \code{chi}: \code{
    xy <- x + y
    y. <- y / sum(y)
    x. <- x / sum(x)
    d <- sqrt(sum(((x. - y.)^2) / (xy / sum(xy))))}
\item \code{hellinger}: \code{d <- sqrt(1/2 * sum(sqrt(x) - sqrt(y))^2)}
}
Note that zeroes are replaced by 0.00001 whem \code{method} equals "chi" or "hellinger".
}
\examples{

#loading data
data(sequenceA)
data(sequenceB)

#matching datasets
sequences <- prepareSequences(
 sequence.A = sequenceA,
 sequence.A.name = "A",
 sequence.B = sequenceB,
 sequence.B.name = "B",
 merge.mode = "complete",
 if.empty.cases = "zero",
 transformation = "hellinger"
 )

#separating them again
#(not actually required in a normal workflow)
A <- sequences[sequences$id == "A", ]
B <- sequences[sequences$id == "B", ]

#computing distance matrix
D <- distanceMatrix(
 sequence.A = A,
 sequence.B = B,
 time.column = NULL,
 exclude.columns = "id",
 method = "manhattan")

}
\seealso{
\code{\link{distance}}
}
\author{
Blas Benito <blasbenito@gmail.com>
}
