% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distances}
\alias{distances}
\title{Distance Methods}
\format{
A data frame with 5 columns and 10 rows.
}
\usage{
data(distances)
}
\description{
A data frame with the names, abbreviations, and expressions of the distance metrics implemented in the package.
}
\seealso{
Other distances:
\code{\link[=distance]{distance()}},
\code{\link[=distance_matrix]{distance_matrix()}}
}
\concept{distances}
\keyword{datasets}
