% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia.R
\name{distantia}
\alias{distantia}
\title{Time Series Dissimilarity Analysis}
\usage{
distantia(
  tsl = NULL,
  distance = "euclidean",
  diagonal = TRUE,
  weighted = TRUE,
  ignore_blocks = FALSE,
  lock_step = FALSE,
  repetitions = 0L,
  permutation = "restricted_by_row",
  block_size = NULL,
  seed = 1
)
}
\arguments{
\item{tsl}{(required, time series list) list of zoo time series. Default: NULL}

\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset \link{distances}. Default: "euclidean".}

\item{diagonal}{(optional, logical vector). If TRUE, diagonals are included in the dynamic time warping computation. Default: TRUE}

\item{weighted}{(optional, logical vector) If TRUE, diagonal is set to TRUE, and diagonal cost is weighted by a factor of 1.414214. Default: TRUE}

\item{ignore_blocks}{(optional, logical vector). If TRUE, blocks of consecutive least cost path coordinates are trimmed to avoid inflating the psi dissimilarity Irrelevant if \code{diagonal = TRUE}. Default: FALSE.}

\item{lock_step}{(optional, logical vector) If TRUE, time-series captured at the same times are compared sample wise (with no dynamic time warping). Requires time series in argument \code{tsl} to be fully aligned, or it will return an error. Default: FALSE.}

\item{repetitions}{(optional, integer vector) number of permutations to compute the p-value. If 0, p-values are not computed. Otherwise, the minimum is 2. Default: 0}

\item{permutation}{(optional, character vector) permutation method, only relevant when \code{repetitions} is higher than zero. Valid values are: "restricted_by_row", "restricted", "free_by_row", and "free". Default: "restricted_by_row".}

\item{block_size}{(optional, integer) Size of the row blocks for the restricted permutation test. Only relevant when permutation methods are "restricted" or "restricted_by_row" and \code{repetitions} is higher than zero. A block of size \code{n} indicates that a row can only be permuted within a block of \code{n} adjacent rows. If NULL, defaults to the rounded one tenth of the shortest sequence in \code{tsl}. Default: NULL.}

\item{seed}{(optional, integer) initial random seed to use for replicability when computing p-values. Default: 1}
}
\value{
Data frame with the attribute \code{type} set to \code{distantia_df} and the following columns:
\itemize{
\item \code{x}: name of the sequence \code{x}.
\item \code{y}: name of the sequence \code{y}.
\item \code{distance}: name of the distance metric.
\item \code{diagonal}: value of the argument \code{diagonal}.
\item \code{weighted}: value of the argument \code{weighted}.
\item \code{ignore_blocks}: value of the argument \code{ignore_blocks}.
\item \code{lock_step}: value of the argument \code{lock_step}.
\item \code{repetitions} (only if \code{repetitions > 0}): value of the argument \code{repetitions}.
\item \code{permutation} (only if \code{repetitions > 0}): name of the permutation method used to compute p-values.
\item \code{seed} (only if \code{repetitions > 0}): random seed used to in the permutations.
\item \code{null_mean} (only if \code{repetitions > 0}): mean of the null distribution of psi values computed from the permutaitons.
\item \code{null_sd} (only if \code{repetitions > 0}): standard deviation of the null distribution of psi values.
\item \code{p_value}  (only if \code{repetitions > 0}): proportion of scores smaller or equal than \code{psi} in the null distribution.
\item \code{psi}: psi dissimilarity of the sequences \code{a} and \code{b}.
}
}
\description{
This function combines dynamic time warping and lock-step comparison with restricted permutation methods to assess dissimilarity between pairs time series or any other sort of data composed of events ordered across a relevant dimension.

\strong{Dynamic time warping} (DTW for short) minimizes the distance between time series. This method ignores sampling time and relies on sample order only. This feature allows it to address different scenarios when comparing regular or irregular time series: phenological time series captured at different latitudes or elevations; time series captured over different years or past time periods; time series lacking precise times; pairs of coordinates representing movement over time, such as migration flight paths.

On the other hand, \strong{lock-step comparison} involves assessing the pairwise distance between samples in time series captured at the same times. This method is an alternative to dynamic time warping when the goal requires assessing the synchronicity of the compared time series.

The \strong{psi dissimilarity score} normalizes the multivariate distance between time series samples by the sum of distances between consecutive samples of each time series. This feature allows comparing psi dissimilarity scores computed for time series of different lengths. A psi score of zero indicates perfect similarity.

\strong{Restricted permutation} rearranges time series samples to help generate a null distribution of dissimilarity scores. The observed dissimilarity score computed on the original data is compared with this null distribution to obtain a p-value that helps assess the strength of the dissimilarity score between two time-series. In essence, restricted permutation is useful to answer the question "how robust is the similarity between two time series?"

Four different permutation methods are available:
\itemize{
\item "restricted": Restricted shuffling of rows and columns within blocks of rows. This method separates the data into blocks of rows of given sizes, and rearranges the data within blocks with no restrictions.
\item "restricted_by_row": Shuffling of complete rows within blocks of rows. This method assumes a dependency between data in the same row, and is useful for data representing percentages or proportions, or variables closely related with each other.
\item "free": Unrestricted shuffling of rows and columns across the complete time series. This method assumes independence between data in separate columns. It generates unrealistic randomized scenarios and may lead to an overestimation of the robustness of dissimilarity scores.
\item "free_by_row": Unrestricted shuffling of complete rows. This method is a useful alternative to "free" when there is a dependency between data in the same row, such as in data representing percentages or proportions.
}

This function supports progress bars generated by the \code{progressr} package. See examples.

This function also accepts a parallelization setup via \code{\link[future:plan]{future::plan()}}, but it might only be worth it for large time series lists.
}
\examples{
#parallelization (best option for large time series lists)
# future::plan(
#  multisession,
#  workers = 2
#)

#progress bar
progressr::handlers(global = TRUE)

}
