% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_aggregate.R
\name{distantia_aggregate}
\alias{distantia_aggregate}
\title{Aggregate Distantia Data Frame}
\usage{
distantia_aggregate(df = NULL, f = mean, ...)
}
\arguments{
\item{df}{(required, data frame) Output of [distantia()] or [distantia_importance()]. Default: NULL}

\item{f}{(optional, function) Function to summarize psi scores (for example, `mean`) when there are several combinations of parameters in `df`. Ignored when there is a single combination of [distantia()] arguments in the input. Default: `mean`}

\item{...}{(optional, arguments of `f`) Further arguments to pass to the function `f`.}
}
\value{
Data frame with attribute "type" equal to "distantia_df.
}
\description{
Applies a function to aggregate the output of [distantia()] when several combinations of arguments have been used.

The aggregation function is applied after scaling the dissimilarity scores `Psi` by group.

The original data frame is returned without modification if there is a single combination of arguments, or if an invalid `f` is provided.
}
