% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_boxplot.R
\name{distantia_boxplot}
\alias{distantia_boxplot}
\title{Boxplot of Distantia Data Frame}
\usage{
distantia_boxplot(df = NULL, color = NULL, f = mean, ...)
}
\arguments{
\item{df}{(required, data frame) output of \code{\link[=distantia]{distantia()}} or \code{\link[=distantia_importance]{distantia_importance()}}. Default: NULL}

\item{color}{(optional, character vector) boxplot fill color. Default: NULL}

\item{f}{(optional, function) function used to order the boxes. Can be one of \code{mean}, \code{median}, \code{min}, \code{max}, or \code{quantile}. Default: \code{mean}.}

\item{...}{(optional, additional arguments to \code{f}). If \code{f} is \code{quantile}, \code{probs = 0.75} can be used. Default: ...}
}
\value{
Boxplot and list with stats used to draw it
}
\description{
For data frames resulting from \code{\link[=distantia]{distantia()}} it plots the distribution of the Psi values of each time series against all others. Time series with higher values are the most dissimilar.

For data frames resulting from \code{\link[=distantia_importance]{distantia_importance()}} it plots the contribution to similarity/dissimilarity of each variable across all time-series. Variables with higher values contribute the most to the dissimilarity between time series.
}
\examples{
tsl <- tsl_simulate(
  n = 5,
  time_range = c(
    "2010-01-01",
    "2024-12-31"
  ),
  cols = 3
)

#distantia boxplot
df <- distantia(
  tsl = tsl,
  distance = c("euclidean", "manhattan")
)

distantia_boxplot(
  df = df,
  f = quantile,
  probs = 0.75
)

#importance boxplot
df_importance <- distantia_importance(
  tsl = tsl,
  distance = c("manhattan", "euclidean")
)

distantia_boxplot(
  df = df_importance,
  f = median
  )
}
