% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_cluster_kmeans.R
\name{distantia_cluster_kmeans}
\alias{distantia_cluster_kmeans}
\title{Kmeans Clustering of a Distantia Data Frame}
\usage{
distantia_cluster_kmeans(df = NULL, clusters = NULL, seed = 1)
}
\arguments{
\item{df}{(required, data frame) Output of \code{\link[=distantia]{distantia()}}. Default: NULL}

\item{clusters}{(required, integer) Number of groups to generate. If NULL (default), \code{\link[=cluster_kmeans_optimizer]{cluster_kmeans_optimizer()}} is used to find the number of clusters that maximizes the mean silhouette width (see \code{\link[=cluster_silhouette]{cluster_silhouette()}}). Default: NULL}

\item{seed}{(optional, integer) Random seed to be used during the K-means computation. Default: 1}
}
\value{
List with the following objects:
\itemize{
\item \code{kmeans}: kmeans object.
\item \code{clusters}: integer, number of clusters.
\item \code{silhouette_width}: mean silhouette width of the hierarchical clustering.
\item \code{df}: data frame with names of the time series, their group assignation, and their silhouette width scores.
\item \code{d}: psi distance matrix.
\item \code{optimization}: only if \code{clusters = NULL}, data frame with optimization results from \code{\link[=cluster_hclust_optimizer]{cluster_hclust_optimizer()}}.
}
}
\description{
Kmeans Clustering of a Distantia Data Frame
}
\details{
All relevant details about clustering methods and relevant references are available in the help file of \code{\link[stats:kmeans]{stats::kmeans()}}.
}
\examples{

tsl <- tsl_simulate(
  n = 10,
  time_range = c(
    "2010-01-01 12:00:25",
    "2024-12-31 11:15:45"
  )
)

df <- distantia(
  tsl = tsl
)

k <- distantia_cluster_kmeans(
  df = df
)

k$df

# #Optional: kmeans plot
# k_plot <- factoextra::fviz_cluster(
#   object = k$kmeans,
#   data = k$d,
#   repel = TRUE
# )

}
\seealso{
\code{\link[stats:kmeans]{stats::kmeans()}}
}
