% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_importance.R
\name{distantia_importance}
\alias{distantia_importance}
\title{Contribution of Individual Variables to Dissimilarity}
\usage{
distantia_importance(
  tsl = NULL,
  distance = "euclidean",
  diagonal = TRUE,
  weighted = TRUE,
  ignore_blocks = FALSE,
  lock_step = FALSE,
  robust = TRUE
)
}
\arguments{
\item{tsl}{(required, time series list) list of zoo time series. Default: NULL}

\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset \code{distances}. Default: "euclidean".}

\item{diagonal}{(optional, logical vector). If TRUE, diagonals are included in the computation of the cost matrix. Default: TRUE}

\item{weighted}{(optional, logical vector) If TRUE, diagonal is set to TRUE, and diagonal cost is weighted by a factor of 1.414214. Default: TRUE}

\item{ignore_blocks}{(optional, logical vector). If TRUE, blocks of consecutive path coordinates are trimmed to avoid inflating the psi distance. Ignored if \code{diagonal = TRUE}. Default: FALSE.}

\item{lock_step}{(optional, logical vector) If TRUE, time-series are compared row wise and no least-cost path optimization is performed. Requires time series in argument \code{tsl} to be fully aligned. Default: FALSE.}

\item{robust}{(required, logical vector). If TRUE (default), importance scores are computed using the least cost path of the complete time series as reference. Setting it to FALSE allows to replicate importance scores of the previous versions of this package. This option is irrelevant when \code{lock_step = TRUE}. Default: TRUE}
}
\value{
Data frame with the following columns:
\itemize{
\item \code{x}: name of the sequence \code{x}.
\item \code{y}: name of the sequence \code{y}.
\item \code{distance}: name of the distance metric.
\item \code{diagonal}: value of the argument \code{diagonal}.
\item \code{weighted}: value of the argument \code{weighted}.
\item \code{ignore_blocks}: value of the argument \code{ignore_blocks}.
\item \code{lock_step}: value of the argument \code{lock_step}.
\item \code{robust}: value of the argument \code{robust}.
\item \code{variable}: name of the individual variable.
\item \code{psi}: overall psi score of \code{a} and \code{b}.
\item \code{psi_only_with}: psi score of the variable.
\item \code{psi_without}: psi score without the variable.
\item \code{psi_difference}: difference between \code{psi_only_with} and \code{psi_without}.
\item \code{psi_drop}: change in psi (as a percentage of the overall psi) when the variable is removed.
\item \code{importance}: \code{psi_difference} as a precentage of \code{psi}. Only interpretable when \code{robust = TRUE}.
}
}
\description{
This function helps assess the contribution of individual variables to the dissimilarity between pairs of time series.

Importance scores are based on three values:
\itemize{
\item \strong{psi}: overall dissimilarity when all variables are considered.
\item \strong{psi_only_with}: dissimilarity when only the given variable is considered.
\item \strong{psi_without}: dissimilarity when all other variables are considered.
}

From these indicators, two importance scores are considered:

\itemize{
\item \strong{psi_drop}: Computed as \code{((psi - psi_without) * 100)/psi} and interpreted as "increase in similarity when a variable is removed". Positive values indicate that the variable contributes to dissimilarity, while negative values indicate otherwise. This is the original importance score introduced with the first version of the package.
\item \strong{importance}: Computed as \code{((psi_only_with - psi_without) * 100)/psi} and interpreted as "relative dissimilarity induced by the variable". Positive values indicate that the variable contributes to dissimilarity, while negative values indicate otherwise. This is a new metric introduced with the version 2.0 of the package, and it is more informative than \code{psi_drop}.
}

Additionally, these importance scores can be computed in two different ways defined by the argument \code{robust}. Please notice that this argument is irrelevant when \code{lock_step = TRUE}:
\itemize{
\item \code{robust = TRUE} (default): Normalizes \code{psi_only_with} and \code{psi_without} from the least cost path of the full set of variables, making importance scores of separate variables fully comparable. This a novel version of the importance algorithm that yields more stable and comparable solutions.
\item \code{robust = FALSE}: Normalizes \code{psi_only_with} and \code{psi_without} from the least cost path of the individual variables, which makes individual importance scores harder to compare. This option renders the "importance" column unreliable. However, the \code{psi_drop} score still remains as a useful metric of importance. This option can only be recommended when the goal is replicating or comparing with previous studies.
}

This function supports progress bars generated by the \code{progressr} package. See examples.

This function also accepts a parallelization setup via \code{\link[future:plan]{future::plan()}}, but it might only be worth it for very long time series.
}
\examples{
#three time series
#climate and ndvi in Fagus sylvatica stands in Spain, Germany, and Sweden
data("fagus_dynamics")

tsl <- tsl_initialize(
  x = fagus_dynamics,
  id_column = "site",
  time_column = "date"
)

if(interactive()){
  tsl_plot(
    tsl = tsl,
    guide_columns = 3
    )
}

#parallelization setup (not worth it for this data size)
future::plan(
 future::multisession,
 workers = 2 #set to parallelly::availableWorkers() - 1
)

#progress bar
progressr::handlers(global = TRUE)

#dissimilarity analysis
tsl_dtw <- distantia(
  tsl = tsl
)
tsl_dtw[, c("x", "y", "psi")]

#most similar cases are Germany and Sweden
#most different cases are Spain and Sweden

#importance scores using dynamic time warping (lock_step = FALSE)
tsl_dtw_importance <- distantia_importance(
  tsl = tsl
)

#focus on important columns
tsl_dtw_importance[, c("x", "y", "variable", "importance")]

#in column "importance":
#  positive values: contribution to dissimilarity
#  negative values: contribution to similarity

#variable making Germany and Sweden similar: rainfall
#variables making Spain and Sweden different: evi and temperature

#WARNING: importance scores are very sensitive to scale!
#---------------------------------------------------------

#scaling and centering tsl
tsl_scaled <- tsl_transform(
  tsl = tsl,
  f = f_scale
)

#importance scores on scaled values
tsl_dtw_importance_scaled <- distantia_importance(
  tsl = tsl_scaled
)

#very different results (scaling is the correct choice here)
tsl_dtw_importance_scaled[, c("x", "y", "variable", "importance")]

#variable making Germany and Sweden similar: temperature and rainfall
#variables making Spain and Sweden different: evi and rainfall

#importance computed with lock-step method
#NOTE: more sensitive to latitudinal or altitudinal differences
tsl_lock_step_importance_scaled <- distantia_importance(
  tsl = tsl_scaled,
  lock_step = TRUE
)

tsl_lock_step_importance_scaled[, c("x", "y", "variable", "importance")]

#variable making Germany and Sweden similar: rainfall and temperature
#variables making Spain and Sweden different: temperature and rainfall

#disable parallelization
future::plan(
  future::sequential
)
}
