% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_importance_wide.R
\name{distantia_importance_wide}
\alias{distantia_importance_wide}
\title{Distantia Importance Data Frame to Wide Format}
\usage{
distantia_importance_wide(df = NULL, sep = "__")
}
\arguments{
\item{df}{(required, data frame) Output of [distantia_importance()]. Default: NULL}

\item{sep}{(required, character string) Separator between the name of the importance metric and the time series variable. Default: "__".}
}
\value{
wide data frame
}
\description{
Distantia Importance Data Frame to Wide Format
}
\examples{

data("eemian_pollen")
data("eemian_coordinates")

#initialize tsl
tsl <- distantia::tsl_init(
  x = distantia::eemian_pollen,
  id_column = "site",
  time_column = "depth"
)

#compute importance
df <- distantia::distantia_importance(
  tsl = tsl
)

colnames(df)

#to wide format, one column per variable and importance metric
df_wide <- distantia::distantia_importance_wide(
  df = df
)

colnames(df_wide)


}
