% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_kmeans.R
\name{distantia_kmeans}
\alias{distantia_kmeans}
\title{Kmeans Clustering of a Distantia Data Frame}
\usage{
distantia_kmeans(
  df = NULL,
  clusters = NULL,
  output = c("kmeans", "df", "plot"),
  seed = 1
)
}
\arguments{
\item{df}{(required, data frame) Output of [distantia()]. Default: NULL}

\item{clusters}{(required, integer) Number of groups to generate. If NULL (default), a kmeans clustering with the number of groups having the highest median silhouette width is returned (see [cluster_silhouette()]). Default: NULL}

\item{output}{(required, character string) Options are:
\itemize{
  \item "df": data frame with the columns "time_series", "cluster" and "silhouette_width" (see [cluster_silhouette()]).
  \item "kmeans": object resulting from [stats::kmeans()]. This list includes the objects "silhouette_df" and "silhouette_median", both resulting from [cluster_silhouette()].
  \item "plot": clusters plot.

}
If "df", the argument 'df' with the columns "time_series", "cluster" and "silhouette_width" is returned. If "kmeans", then the kmeans clustering object is returned instead. Default: "kmeans"}

\item{seed}{(optional, integer) Random seed to be used during the K-means computation. Default: 1}
}
\value{
Object of class "kmeans" or "data.frame".
}
\description{
Kmeans Clustering of a Distantia Data Frame
}
\seealso{
[stats::kmeans()]
}
