% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_matrix.R
\name{distantia_matrix}
\alias{distantia_matrix}
\title{Distantia Data Frame to Distance Matrix}
\usage{
distantia_matrix(df = NULL)
}
\arguments{
\item{df}{(required, data frame) Output of [distantia()] or [distantia_aggregate()]. Default: NULL}
}
\value{
List of distance matrices.
}
\description{
Distantia Data Frame to Distance Matrix
}
