% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_model_frame.R
\name{distantia_model_frame}
\alias{distantia_model_frame}
\title{Dissimilarity Model Frame}
\usage{
distantia_model_frame(
  df = NULL,
  predictors_df = NULL,
  predictors_list = NULL,
  distance = "euclidean"
)
}
\arguments{
\item{df}{(required, data frame) Output of \code{\link[=distantia]{distantia()}} or \code{\link[=distantia_aggregate]{distantia_aggregate()}}. Default: NULL}

\item{predictors_df}{(required, data frame) data frame with numeric predictors to be added to the model frame. Must have a column with the names in \code{df$x} and \code{df$y}. If \code{sf} data frame, the predictor "distance" is added to the model frame. Default: NULL}

\item{predictors_list}{(optional, list) list defining predictors as combinations fo columns in \code{predictors_df}. For example, \code{predictors_list = list(distance = c("x", "y"))} uses the columns \code{"x"} and \code{"y"} from \code{predictors_df} to generate the predictor \code{distance} in the model frame. Can be used to combine groups of predictors with common themes. Any column from \code{predictors_df} not in this list is returned as an individual predictor in the model frame. Default: NULL}

\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset \link{distances}. Default: "euclidean".}
}
\value{
data frame
}
\description{
Generates a model frame from a dissimilarity data frame generated by \code{\link[=distantia]{distantia()}} and a data frame static descriptors of the time series.
}
\examples{
#covid prevalence in California counties
tsl <- tsl_initialize(
  x = covid_prevalence,
  name_column = "name",
  time_column = "time"
)

#dissimilarity analysis
df <- distantia(
  tsl = tsl,
  lock_step = TRUE
)

#combine predictors together
predictors_list <- list(
  economy = c(
    "poverty_percentage",
    "median_income",
    "domestic_product"
    )
)

#generate model frame
model_frame <- distantia_model_frame(
  df = df,
  predictors_df = covid_counties,
  predictors_list = predictors_list
)

head(model_frame)
}
\seealso{
Other dissimilarity_analysis_main: 
\code{\link{distantia}()},
\code{\link{momentum}()}
}
\concept{dissimilarity_analysis_main}
