% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_spatial_network.R
\name{distantia_spatial_network}
\alias{distantia_spatial_network}
\title{Spatial Network of Dissimilarities}
\usage{
distantia_spatial_network(df = NULL, xy = NULL)
}
\arguments{
\item{df}{(required, data frame) Output of [distantia()] or [distantia_aggregate()]. Default: NULL}

\item{xy}{(required, sf POINT data frame) Default: NULL}
}
\value{
sf data frame
}
\description{
Spatial Network of Dissimilarities
}
