% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_to_matrix.R
\name{distantia_to_matrix}
\alias{distantia_to_matrix}
\title{Distantia Data Frame to Distance Matrix}
\usage{
distantia_to_matrix(df = NULL, f = NULL, ...)
}
\arguments{
\item{df}{(required, data frame) Output of [distantia()]. Default: NULL}

\item{f}{(optional, function) Function to summarize psi scores (for example, `mean`) when there are several combinations of parameters in `df`. Ignored when there is a single combination of [distantia()] arguments in the input. Default: NULL}

\item{...}{(optional, arguments of `f`) Further arguments to pass to the function `f`.}
}
\value{
List of distance matrices.
}
\description{
The output of [distantia()] is a data frame of dissimilarity scores between pairs of time series in a time series list. This function transforms such data frame into a list of distance matrices.

The number of matrices the output list depends on how many combinations of the [distantia()] arguments `distance`, `diagonal`, `weighted`, and `ignore_blocks` are in `df`, and in whether the aggregation function `f` is provided or not.

If there is a single combination of arguments, or if a valid `f` is provided, a list with one distance matrix is returned. If there are several combinations of arguments and `f` is not provided, the list contains as many matrices as combinations of arguments.

If provided, the aggregation function `f` is applied after scaling dissimilarity scores separately by group.
}
