% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_to_sf.R
\name{distantia_to_sf}
\alias{distantia_to_sf}
\title{Distantia to Spatial Data Frame}
\usage{
distantia_to_sf(df = NULL, xy = NULL)
}
\arguments{
\item{df}{(required, data frame) Output of \code{\link[=distantia]{distantia()}} or \code{\link[=distantia_aggregate]{distantia_aggregate()}}. Default: NULL}

\item{xy}{(required, sf POINT data frame) Sf data frame with the coordinates of the time series in argument 'df'. It must have a column with all time series names in \code{df$x} and \code{df$y}. See \verb{[eemian_cooordinaes]} example. Default: NULL}
}
\value{
sf data frame with LINESTRING geometry.
}
\description{
Given an sf data frame with the coordinates of a time series list transforms a data frame resulting from \code{\link[=distantia]{distantia()}} to an sf data frame with lines connecting the time series locations. This can be used to visualize a geographic network of similarity/dissimilarity between locations. See example for further details.
}
\examples{
data("eemian_pollen")
data("eemian_coordinates")

#initialize tsl
tsl <- distantia::tsl_init(
  x = distantia::eemian_pollen,
  id_column = "site",
  time_column = "depth"
)

#compute dissimilarity
psi <- distantia::distantia(
  tsl = tsl
)

#transform to sf
psi_sf <- distantia::distantia_to_sf(
  df = psi,
  xy = distantia::eemian_coordinates
)

#mapping with tmap
# library(tmap)
# tmap::tmap_mode("view")
#
# tmap::tm_shape(psi_sf) +
#   tmap::tm_lines(
#     col = "psi",
#     lwd = 3
#   ) +
#   tmap::tm_shape(distantia::eemian_coordinates) +
#   tmap::tm_dots(size = 0.1, col = "gray50")
}
