% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eemian_coordinates}
\alias{eemian_coordinates}
\title{Coordinates of Nine Eemian Sites}
\format{
sf data frame with 4 columns and 9 rows.
}
\usage{
data(eemian_coordinates)
}
\description{
Coordinates of Nine Eemian Sites
}
\keyword{datasets}
