% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eemian_pollen}
\alias{eemian_pollen}
\title{Pollen Counts of Nine Eemian Sites}
\format{
dataframe with 24 columns and 376 rows.
}
\usage{
data(eemian_pollen)
}
\description{
Pollen Counts of Nine Eemian Sites
}
\keyword{datasets}
