% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evi}
\alias{evi}
\title{Enhanced Vegetation Index vs Climate in Three European Locations}
\format{
dataframe with 5 columns and 648 rows.
}
\usage{
data(evi)
}
\description{
A dataframe with 648 rows representing enhanced vegetation index, rainfall and teperature in three European locations.
}
\keyword{datasets}
