% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_detrend_gam}
\alias{f_detrend_gam}
\title{Linear Detrending of a Time Series}
\usage{
f_detrend_gam(x = NULL, select = FALSE, k = -1, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{select}{(optional, logical) Argument of [mgcv::gam()]. If TRUE then gam can add an extra penalty to each term so that it can be penalized to zero.}

\item{k}{(optional, integer) Argument of [mgcv::s()]. he dimension of the basis used to represent the smooth term. The default depends on the number of variables that the smooth is a function of. See [mgcv::choose.k()] for further information.}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
Transformed zoo object.
}
\description{
The function applies a GAM to each column (time series) of the input matrix x separately, modeling the relationship between the values and the corresponding time index using a smooth spline function (s()). The predicted values from the GAMs represent the trend component of each time series. After obtaining the trend components for each time series, the function subtracts these trends from the original time series data to obtain the detrended series.
}
