% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_percentage}
\alias{f_percentage}
\title{Transformation to Percentage}
\usage{
f_percentage(x = NULL, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
zoo object
}
\description{
Transformation to Percentage
}
\examples{
x <- zoo_simulate(cols = 2)

y <- f_percentage(
  x = x
)

if(interactive()){
  zoo_plot(x)
  zoo_plot(y)
}
}
\keyword{transformation}
