% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_proportion}
\alias{f_proportion}
\title{Transformation to Proportions}
\usage{
f_proportion(x = NULL, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
Transformed zoo object.
}
\description{
Transformation to Proportions
}
