% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_rescale}
\alias{f_rescale}
\title{Re-scaling of Zoo Object to a New Range}
\usage{
f_rescale(x = NULL, new_min = 0, new_max = 1, old_min = NULL, old_max = NULL)
}
\arguments{
\item{x}{(required, zoo object) Numeric vector. Default: \code{NULL}}

\item{new_min}{(optional, numeric) New minimum value. Default: \code{0}}

\item{new_max}{(optional_numeric) New maximum value. Default: \code{1}}

\item{old_min}{(optional, numeric) Old minimum value. Default: \code{NULL}}

\item{old_max}{(optional_numeric) Old maximum value. Default: \code{NULL}}
}
\value{
zoo object
}
\description{
Re-scaling of Zoo Object to a New Range
}
\examples{
x <- zoo_simulate(cols = 2)

y <- f_rescale(
  x = x,
  new_min = 0,
  new_max = 100
)

if(interactive()){
  zoo_plot(x)
  zoo_plot(y)
}
}
\seealso{
Other data_processing: 
\code{\link{f_center}()},
\code{\link{f_detrend_difference}()},
\code{\link{f_detrend_linear}()},
\code{\link{f_hellinger}()},
\code{\link{f_list}()},
\code{\link{f_pca}()},
\code{\link{f_percentage}()},
\code{\link{f_proportion}()},
\code{\link{f_scale}()},
\code{\link{f_slope}()},
\code{\link{f_smooth_window}()},
\code{\link{f_trend_linear}()},
\code{\link{tsl_aggregate}()},
\code{\link{tsl_resample}()},
\code{\link{tsl_split}()},
\code{\link{tsl_subset}()},
\code{\link{tsl_to_df}()},
\code{\link{tsl_transform}()},
\code{\link{zoo_aggregate}()},
\code{\link{zoo_permute}()},
\code{\link{zoo_resample}()}
}
\concept{data_processing}
