% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_rescale_global}
\alias{f_rescale_global}
\title{Data Transformation: Global Rescaling of to a New Range}
\usage{
f_rescale_global(
  x = NULL,
  new_min = 0,
  new_max = 1,
  old_min = NULL,
  old_max = NULL,
  .global,
  ...
)
}
\arguments{
\item{x}{(required, zoo object) Time Series. Default: \code{NULL}}

\item{new_min}{(optional, numeric) New minimum value. Default: \code{0}}

\item{new_max}{(optional_numeric) New maximum value. Default: \code{1}}

\item{old_min}{(optional, numeric) Old minimum value. Default: \code{NULL}}

\item{old_max}{(optional_numeric) Old maximum value. Default: \code{NULL}}

\item{.global}{(optional, logical) Used to trigger global scaling within \code{\link[=tsl_transform]{tsl_transform()}}.}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
zoo object
}
\description{
Data Transformation: Global Rescaling of to a New Range
}
\examples{
x <- zoo_simulate(cols = 2)

y <- f_rescale_global(
  x = x,
  new_min = 0,
  new_max = 100
)

if(interactive()){
  zoo_plot(x)
  zoo_plot(y)
}
}
\seealso{
Other tsl_transformation: 
\code{\link{f_detrend_difference}()},
\code{\link{f_detrend_linear}()},
\code{\link{f_hellinger}()},
\code{\link{f_list}()},
\code{\link{f_pca}()},
\code{\link{f_percentage}()},
\code{\link{f_proportion}()},
\code{\link{f_scale_global}()},
\code{\link{f_slope}()},
\code{\link{f_trend_linear}()}
}
\concept{tsl_transformation}
