% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_scale}
\alias{f_scale}
\title{Scaling and centering}
\usage{
f_scale(x = NULL, center = TRUE, scale = TRUE, ...)
}
\arguments{
\item{x}{Data frame or matrix.}
}
\value{
Data frame or matrix
}
\description{
Scaling and centering
}
