% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_scale_global}
\alias{f_scale_global}
\title{Data Transformation: Global Centering and Scaling}
\usage{
f_scale_global(x = NULL, center = TRUE, scale = TRUE, .global, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{center}{(optional, logical or numeric vector) Triggers centering if TRUE. Default: TRUE}

\item{scale}{(optional, logical or numeric vector) Triggers scaling if TRUE. Default: TRUE}

\item{.global}{(optional, logical) Used to trigger global scaling within \code{\link[=tsl_transform]{tsl_transform()}}.}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
zoo object
}
\description{
Scaling and/or centering by variable using the mean and standard deviation computed across all time series. Global scaling helps dynamic time warping take variable offsets between time series into account.
}
\examples{
x <- zoo_simulate()

y <- f_scale_global(
  x = x
)

if(interactive()){
  zoo_plot(x)
  zoo_plot(y)
}
}
\seealso{
Other tsl_transformation: 
\code{\link{f_detrend_difference}()},
\code{\link{f_detrend_linear}()},
\code{\link{f_hellinger}()},
\code{\link{f_list}()},
\code{\link{f_pca}()},
\code{\link{f_percentage}()},
\code{\link{f_proportion}()},
\code{\link{f_rescale_global}()},
\code{\link{f_slope}()},
\code{\link{f_smooth_window}()},
\code{\link{f_trend_linear}()}
}
\concept{tsl_transformation}
