% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_scale_local}
\alias{f_scale_local}
\title{Data Transformation: Local Centering and Scaling}
\usage{
f_scale_local(x = NULL, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{center}{(optional, logical or numeric vector) Triggers centering if TRUE. Default: TRUE}

\item{scale}{(optional, logical or numeric vector) Triggers scaling if TRUE. Default: TRUE}
}
\value{
zoo object
}
\description{
Scaling and/or centering by variable and time series. Local scaling helps dynamic time warping focus entirely on shape comparisons.
}
\examples{
x <- zoo_simulate()

y <- f_scale_global(
  x = x
)

if(interactive()){
  zoo_plot(x)
  zoo_plot(y)
}
}
\seealso{
Other tsl_transformation: 
\code{\link{f_detrend_difference}()},
\code{\link{f_detrend_linear}()},
\code{\link{f_hellinger}()},
\code{\link{f_list}()},
\code{\link{f_pca}()},
\code{\link{f_percentage}()},
\code{\link{f_proportion}()},
\code{\link{f_rescale_global}()},
\code{\link{f_slope}()},
\code{\link{f_smooth_window}()},
\code{\link{f_trend_linear}()}
}
