% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_stat}
\alias{f_stat}
\title{Rolling Mean Smoothing}
\usage{
f_stat(x = NULL, w = 3, statistic = mean, ...)
}
\arguments{
\item{x}{(required, zoo object).}

\item{w}{(required, window width) width of the window to compute the rolling mean of the time series.}

\item{statistic}{(required, function) name without quoites of a standard function to compute statistics. Typical examples are `mean`, `max`, `mean`, `median`, and `sd`.}
}
\value{
Data frame or matrix
}
\description{
Rolling Mean Smoothing
}
