% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fagus_coordinates}
\alias{fagus_coordinates}
\title{Coordinates of Fagus sylvatica stands}
\format{
sf data frame with 3 rows and 4 columns
}
\usage{
data(fagus_coordinates)
}
\description{
Coordinates of Fagus sylvatica stands
}
\keyword{datasets}
