% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPsi.R
\name{formatPsi}
\alias{formatPsi}
\title{Formats the output of \code{\link{psi}}.}
\usage{
formatPsi(
  psi.values = NULL,
  to = "dataframe")
}
\arguments{
\item{psi.values}{list produced by \code{\link{psi}}.}

\item{to}{character, either "dataframe" or "matrix".}
}
\description{
Parses a list produced by \code{\link{psi}} to generate either a dataframe or a matrix. Can also format a psi matrix into a dataframe and viceversa.
}
\details{
The function detects the type of input, and checks that it is different from the value of \code{to}. If that is the case, it throws a warning, and returns the input object. It uses the helper function .psiToDataframe, only intended for internal use.
}
\author{
Blas Benito <blasbenito@gmail.com>
}
