% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{honeycomb_climate}
\alias{honeycomb_climate}
\title{Rainfall and Temperature in The Americas}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9432 rows and 4 columns.
}
\usage{
data(honeycomb_climate)
}
\description{
Monthly temperature and rainfall between 2009 and 2019 in 72 hexagonal cells covering The Americas.
}
\seealso{
Other example_data: 
\code{\link{albatross}},
\code{\link{cities_coordinates}},
\code{\link{cities_temperature}},
\code{\link{covid_coordinates}},
\code{\link{covid_polygons}},
\code{\link{covid_prevalence}},
\code{\link{eemian_coordinates}},
\code{\link{eemian_pollen}},
\code{\link{fagus_coordinates}},
\code{\link{fagus_dynamics}},
\code{\link{honeycomb_polygons}}
}
\concept{example_data}
\keyword{datasets}
