% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{honeycomb_polygons}
\alias{honeycomb_polygons}
\title{Hexagonal Grid}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 72 rows and 2 columns.
}
\usage{
data(honeycomb_polygons)
}
\description{
Sf data frame with hexagonal grid of the dataset \link{honeycomb_climate}.
}
\seealso{
Other example_data: 
\code{\link{albatross}},
\code{\link{cities_coordinates}},
\code{\link{cities_temperature}},
\code{\link{covid_coordinates}},
\code{\link{covid_prevalence}},
\code{\link{eemian_coordinates}},
\code{\link{eemian_pollen}},
\code{\link{fagus_coordinates}},
\code{\link{fagus_dynamics}},
\code{\link{honeycomb_climate}}
}
\concept{example_data}
\keyword{datasets}
