% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{importance_full_cpp}
\alias{importance_full_cpp}
\title{Computes Per Column Psi Importance}
\usage{
importance_full_cpp(
  a,
  b,
  method,
  diagonal = FALSE,
  weighted = FALSE,
  trim_blocks = FALSE
)
}
\arguments{
\item{a}{(required, numeric matrix).}

\item{b}{(required, numeric matrix) of same number of columns as 'a'.}

\item{method}{(optional, character string) name or abbreviation of the
distance method. Valid values are in the columns "names" and "abbreviation"
of the dataset `methods`. Default: "euclidean".}

\item{diagonal}{(optional, logical). If TRUE, diagonals are included in the
computation of the cost matrix. Default: FALSE.}

\item{weighted}{(optional, logical). If TRUE, diagonal is set to TRUE, and
diagonal cost is weighted by a factor of 1.414214. Default: FALSE.}

\item{trim_blocks}{(optional, logical). If TRUE, blocks of consecutive path
coordinates are trimmed to avoid inflating the psi distance. Default: FALSE.}
}
\value{
Data frame with psi distances
}
\description{
Computes the psi distance between a and b using all columns,
with each column, and without each column
}
