% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{importance_vintage_cpp}
\alias{importance_vintage_cpp}
\title{Classic Computation of Variable Importance}
\usage{
importance_vintage_cpp(
  a,
  b,
  method = "euclidean",
  diagonal = FALSE,
  weighted = FALSE,
  ignore_blocks = FALSE
)
}
\arguments{
\item{a}{(required, numeric matrix).}

\item{b}{(required, numeric matrix) of same number of columns as 'a'.}

\item{method}{(optional, character string) name or abbreviation of the
distance method. Valid values are in the columns "names" and "abbreviation"
of the dataset `methods`. Default: "euclidean".}

\item{diagonal}{(optional, logical). If TRUE, diagonals are included in the
computation of the cost matrix. Default: FALSE.}

\item{weighted}{(optional, logical). If TRUE, diagonal is set to TRUE, and
diagonal cost is weighted by a factor of 1.414214. Default: FALSE.}

\item{ignore_blocks}{(optional, logical). If TRUE, blocks of consecutive path
coordinates are trimmed to avoid inflating the psi distance. Default: FALSE.}
}
\value{
Data frame with psi distances
}
\description{
Returns the contribution of each variable to the overall dissimilarity
of two sequences. In opposition to the robust version, least-cost paths for each combination
of variables are computed independently. To compute importance,
the function first computes the psi distance dissimilarity between seuqences
using all columns ("psi"), with each column separately ("psi_only_with"),
and without each column ("psi_without). Then, it computes "psi_difference" as
psi_only_with - psi_without. Highest positive values in this column represent the
variables that better explain the dissimilarity between the sequences.

Computes the psi distance between a and b using all columns,
with each column, and without each column
}
