% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{least_cost_matrix_diag_cpp}
\alias{least_cost_matrix_diag_cpp}
\title{Least Cost Matrix Considering Diagonals}
\usage{
least_cost_matrix_diag_cpp(d)
}
\arguments{
\item{d}{(required, distance matrix). Output of [distance_matrix_cpp()].}
}
\value{
Least cost matrix.
}
\description{
Computes the least cost matrix from a distance matrix.
Considers diagonals during computation of least-costs.
}
