% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leastCost.R
\name{leastCost}
\alias{leastCost}
\title{Extracts the least cost from a least-cost path.}
\usage{
leastCost(
  least.cost.path = NULL,
  parallel.execution = TRUE
  )
}
\arguments{
\item{least.cost.path}{dataframe produced by \code{\link{leastCostPath}}.}

\item{parallel.execution}{boolean, if \code{TRUE} (default), execution is parallelized, and serialized if \code{FALSE}.}
}
\value{
A named list with least-cost values.

A named list with least cost values.
}
\description{
Sums the the distances of the samples in a least-cost path.
}
