% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leastCostMatrix.R
\name{leastCostMatrix}
\alias{leastCostMatrix}
\title{Computes a least cost matrix from a distance matrix.}
\usage{
leastCostMatrix <- function(
  distance.matrix = NULL,
  diagonal = FALSE
  )
}
\arguments{
\item{distance.matrix}{numeric matrix or list of numeric matrices, a distance matrix produced by \code{\link{distanceMatrix}}.}

\item{diagonal}{boolean, if \code{TRUE}, diagonals are included in the computation of the least cost path. Defaults to \code{FALSE}, as the original algorithm did not include diagonals in the computation of the least cost path.}
}
\value{
A matrix with the same dimensions as \code{distance.matrix}.
}
\description{
Computes a least cost matrix from a distance matrix, constrained by the order of the samples in the two sequences used to compute the distance matrix.
}
\examples{

}
