% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methods}
\alias{methods}
\title{Distance Methods}
\format{
Dataframe with 5 columns and 10 rows
}
\usage{
data(methods)
}
\description{
Data frame with the names, abbreviations, and expressions of the distance metrics implemented in the package.
}
\keyword{datasets}
