% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{null_psi_cpp}
\alias{null_psi_cpp}
\title{Null Distribution of Psi Distances Between Two Time-Series}
\usage{
null_psi_cpp(
  a,
  b,
  distance = "euclidean",
  diagonal = FALSE,
  weighted = FALSE,
  ignore_blocks = FALSE,
  repetitions = 100L,
  permutation = "restricted_by_row",
  block_size = as.integer(c(2, 3, 4)),
  seed = 1L
)
}
\arguments{
\item{a}{(required, numeric matrix).}

\item{b}{(required, numeric matrix) of same number of columns as 'a'.}

\item{distance}{(optional, character string) name or abbreviation of the
distance method. Valid values are in the columns "names" and "abbreviation"
of the dataset `distances`. Default: "euclidean".}

\item{diagonal}{(optional, logical). If TRUE, diagonals are included in the
computation of the cost matrix. Default: FALSE.}

\item{weighted}{(optional, logical). If TRUE, diagonal is set to TRUE, and
diagonal cost is weighted by a factor of 1.414214. Default: FALSE.}

\item{ignore_blocks}{(optional, logical). If TRUE, blocks of consecutive path
coordinates are trimmed to avoid inflating the psi distance. This argument
has nothing to do with block_size!. Default: FALSE.}

\item{repetitions}{(optional, integer) number of null psi values to generate. Default: 100}

\item{permutation}{(optional, character) permutation method. Valid values are listed below from higher to lower randomness:
\itemize{
  \item "free": unrestricted shuffling of rows and columns. Ignores block_size.
  \item "free_by_row": unrestricted shuffling of complete rows. Ignores block size.
  \item "restricted": restricted shuffling of rows and columns within blocks.
  \item "restricted_by_row": restricted shuffling of rows within blocks.
}}

\item{block_size}{(optional, integer vector) vector with block sizes for
restricted permutation. A block size of 3 indicates that a row can only be permuted
within a block of 3 adjacent rows. Minimum value is 2. Default: c(2, 3, 4).}

\item{seed}{(optional, integer) initial random seed to use for replicability. Default: 1}
}
\value{
Numeric vector with null distribution of psi distances.
}
\description{
Computes null psi distances for permuted versions of the sequences
\code{a} and \code{b} with the same number of columns and arbitrary numbers of rows.
NA values should be removed before using this function.
If the selected distance function is "chi" or "cosine", pairs of zeros should
be either removed or replaced with pseudo-zeros (i.e. 0.00001).
}
