% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{null_psi_dynamic_time_warping_cpp}
\alias{null_psi_dynamic_time_warping_cpp}
\title{(C++) Null Distribution of Dissimilarity Scores of Two Time Series}
\usage{
null_psi_dynamic_time_warping_cpp(
  x,
  y,
  distance = "euclidean",
  diagonal = TRUE,
  weighted = TRUE,
  ignore_blocks = FALSE,
  bandwidth = 1,
  repetitions = 100L,
  permutation = "restricted_by_row",
  block_size = 3L,
  seed = 1L
)
}
\arguments{
\item{x}{(required, numeric matrix) of same number of columns as 'y'.}

\item{y}{(required, numeric matrix).}

\item{distance}{(optional, character string) distance name from the "names"
column of the dataset \code{distances} (see \code{distances$name}). Default: "euclidean".}

\item{diagonal}{(optional, logical). If TRUE, diagonals are included in the
computation of the cost matrix. Default: FALSE.}

\item{weighted}{(optional, logical). If TRUE, diagonal is set to TRUE, and
diagonal cost is weighted by a factor of 1.414214. Default: FALSE.}

\item{ignore_blocks}{(optional, logical). If TRUE, blocks of consecutive path
coordinates are trimmed to avoid inflating the psi distance. This argument
has nothing to do with block_size!. Default: FALSE.}

\item{bandwidth}{(required, numeric) Size of the Itakura parallelogram at
both sides of the diagonal used to constrain the least cost path. Expressed
as a fraction of the number of matrix rows and columns. Unrestricted by default.
Default: 1}

\item{repetitions}{(optional, integer) number of null psi values to generate. Default: 100}

\item{permutation}{(optional, character) permutation method. Valid values are listed below from higher to lower randomness:
\itemize{
\item "free": unrestricted shuffling of rows and columns. Ignores block_size.
\item "free_by_row": unrestricted shuffling of complete rows. Ignores block size.
\item "restricted": restricted shuffling of rows and columns within blocks.
\item "restricted_by_row": restricted shuffling of rows within blocks.
}}

\item{block_size}{(optional, integer) block size in rows for
restricted permutation. A block size of 3 indicates that a row can only be permuted
within a block of 3 adjacent rows. Minimum value is 2. Default: 3.}

\item{seed}{(optional, integer) initial random seed to use for replicability. Default: 1}
}
\value{
numeric vector
}
\description{
Applies permutation methods to compute null distributions for
the psi scores of two time series.
NA values should be removed before using this function.
If the selected distance function is "chi" or "cosine", pairs of zeros should
be either removed or replaced with pseudo-zeros (i.e. 0.00001).
}
\seealso{
Other Rcpp_dissimilarity_analysis: 
\code{\link{null_psi_lock_step_cpp}()},
\code{\link{psi_dynamic_time_warping_cpp}()},
\code{\link{psi_equation_cpp}()},
\code{\link{psi_lock_step_cpp}()}
}
\concept{Rcpp_dissimilarity_analysis}
