% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{null_psi_full_cpp}
\alias{null_psi_full_cpp}
\title{Computes Psi Distance Between Two Time-Series}
\usage{
null_psi_full_cpp(
  a,
  b,
  method,
  diagonal = FALSE,
  weighted = FALSE,
  trim_blocks = FALSE,
  block_size = as.integer(c(2, 3, 4)),
  seed = 1L,
  repetitions = 100L
)
}
\arguments{
\item{a}{(required, numeric matrix).}

\item{b}{(required, numeric matrix) of same number of columns as 'a'.}

\item{method}{(optional, character string) name or abbreviation of the
distance method. Valid values are in the columns "names" and "abbreviation"
of the dataset `methods`. Default: "euclidean".}

\item{diagonal}{(optional, logical). If TRUE, diagonals are included in the
computation of the cost matrix. Default: FALSE.}

\item{weighted}{(optional, logical). If TRUE, diagonal is set to TRUE, and
diagonal cost is weighted by a factor of 1.414214. Default: FALSE.}

\item{trim_blocks}{(optional, logical). If TRUE, blocks of consecutive path
coordinates are trimmed to avoid inflating the psi distance. This argument
has nothing to do with block_size!. Default: FALSE.}

\item{block_size}{(optional, integer vector) vector with block sizes for the
restricted permutation. A block size of 3 indicates that a row can only be permuted
within a block of 3 adjacent rows. Default: c(2, 3, 4).}

\item{seed}{(optional, integer) initial random seed to use for replicability. Default: 1}

\item{repetitions}{(optional, integer) number of null psi values to generate. Default: 100}
}
\value{
Psi distance
}
\description{
Computes the distance psi between two matrices
\code{a} and \code{b} with the same number of columns and arbitrary numbers of rows.
NA values should be removed before using this function.
If the selected distance function is "chi" or "cosine", pairs of zeros should
be either removed or replaced with pseudo-zeros (i.e. 0.00001).
}
