% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{null_psi_paired_cpp}
\alias{null_psi_paired_cpp}
\title{Computes Null Distribution of Psi Distances Between Two Time-Series With Paired Samples}
\usage{
null_psi_paired_cpp(
  a,
  b,
  method,
  block_size = as.integer(c(2, 3, 4)),
  seed = 1L,
  repetitions = 100L
)
}
\arguments{
\item{a}{(required, numeric matrix).}

\item{b}{(required, numeric matrix) of same number of columns as 'a'.}

\item{method}{(optional, character string) name or abbreviation of the
distance method. Valid values are in the columns "names" and "abbreviation"
of the dataset `methods`. Default: "euclidean".}

\item{block_size}{(optional, integer vector) vector with block sizes for the
restricted permutation. A block size of 3 indicates that a row can only be permuted
within a block of 3 adjacent rows. Default: c(2, 3, 4).}

\item{seed}{(optional, integer) initial random seed to use for replicability. Default: 1}

\item{repetitions}{(optional, integer) number of null psi values to generate. Default: 100}
}
\value{
Psi distance
}
\description{
Computes the distance psi between two matrices
\code{a} and \code{b} with the same number of columns and rows. Distances
between \code{a} and \code{b} are computed row wise rather than via distance
matrix and least-cost path computation.
NA values should be removed before using this function.
If the selected distance function is "chi" or "cosine", pairs of zeros should
be either removed or replaced with pseudo-zeros (i.e. 0.00001).
}
