% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permute}
\alias{permute}
\title{Restricted permutation}
\usage{
permute(x, block_size, seed = 1L)
}
\arguments{
\item{x}{(required, numeric matrix). Numeric matrix to permute.}

\item{block_size}{(optional, integer) block size in number of rows.
Minimum value is 2, and maximum value is nrow(x).}

\item{seed}{(optional, integer) random seed to use.}
}
\value{
Numeric matrix, permuted version of x.
}
\description{
Divides a sequence or time-series in blocks and permutes rows
within these blocks. Used to compute p-values for psi distances. Larger block
sizes disrupt the time-series structure in the data.
}
