% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permute_cpp}
\alias{permute_cpp}
\title{Restricted Permutation by Blocks}
\usage{
permute_cpp(x, block_size, seed = 1L)
}
\arguments{
\item{x}{(required, numeric matrix). Numeric matrix to permute.}

\item{block_size}{(optional, integer) block size in number of rows.
Minimum value is 2, and maximum value is nrow(x).}

\item{seed}{(optional, integer) random seed to use.}
}
\value{
Numeric matrix, permuted version of x.
}
\description{
Divides a matrix in blocks of a given size and permutes rows
within these blocks. Used to compute null distributions for psi distances.
Larger block sizes increasingly disrupt data structure over time.
}
