% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permute_independent_cpp}
\alias{permute_independent_cpp}
\title{Restricted Permutation by Blocks and Columns}
\usage{
permute_independent_cpp(x, block_size, seed = 1L)
}
\arguments{
\item{x}{(required, numeric matrix). Numeric matrix to permute.}

\item{block_size}{(optional, integer) block size in number of rows.
Minimum value is 2, and maximum value is nrow(x).}

\item{seed}{(optional, integer) random seed to use.}
}
\value{
Numeric matrix, permuted version of x.
}
\description{
Divides a sequence or time-series in blocks and permutes cases
within these blocks, but independently by column.
Used to compute p-values for psi distances when columns are independent.
Larger block sizes increasingly disrupt data structure over time.
}
