% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Plot Distance or Cost Matrix}
\usage{
plot_matrix(
  m = NULL,
  col = grDevices::hcl.colors(n = 100, palette = "Zissou 1"),
  main_title = NULL,
  xlab = NULL,
  ylab = NULL,
  guide_title = NULL,
  path = NULL,
  path_width = 1,
  path_color = "black"
)
}
\arguments{
\item{m}{(required, numeric matrix) distance or cost matrix generated by [distance_matrix()] or [cost_matrix()], but any numeric matrix will work. Default: NULL}

\item{col}{(optional, character vector) vector of colors. Default: `grDevices::hcl.colors(n = 100, palette = "Zissou 1")`.}

\item{main_title}{(optional, character string) plot title. Default: NULL}

\item{xlab}{(optional, character string) title of the x axis (matrix columns). Default: NULL}

\item{ylab}{(optional, character string) title of the y axis (matrix rows). Default: NULL}

\item{guide_title}{(optional, character string) title of the color guide. Default: NULL}

\item{path}{(optional, data frame) least cost path generated with [cost_path()]. This data frame must have the attribute `type == "cost_path`, or otherwise it will be ignored. Default: NULL.}

\item{path_width}{(optional, numeric) width of the least-cost path. Default: 1}

\item{path_color}{(optional, character string) color of the least-cost path. Default: "black"}
}
\description{
This function is a simplified version of [fields::imagePlot()], by [Douglas Nychka](https://dnychka.github.io/). The original version is recommended in case more customization than the provided here is needed.
}
