% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_panel.R
\name{plot_panel}
\alias{plot_panel}
\title{Plot Distance or Cost Matrix}
\usage{
plot_panel(
  m = NULL,
  matrix_color = NULL,
  main_title = NULL,
  xlab = NULL,
  ylab = NULL,
  guide_title = NULL,
  path = NULL,
  path_width = 1,
  path_color = "black",
  cex = 1
)
}
\arguments{
\item{m}{(required, numeric matrix) distance or cost matrix generated by [distance_matrix()] or [cost_matrix()], but any numeric matrix will work. Default: NULL}

\item{matrix_color}{(optional, character vector) vector of colors. Default: NULL}

\item{main_title}{(optional, character string) plot title. Default: NULL}

\item{xlab}{(optional, character string) title of the x axis (matrix columns). Default: NULL}

\item{ylab}{(optional, character string) title of the y axis (matrix rows). Default: NULL}

\item{guide_title}{(optional, character string) title of the color guide. Default: NULL}

\item{path}{(optional, data frame) least cost path generated with [cost_path()]. This data frame must have the attribute `type == "cost_path`, or otherwise it will be ignored. Default: NULL.}

\item{path_width}{(optional, numeric) width of the least-cost path. Default: 1}

\item{path_color}{(optional, character string) color of the least-cost path. Default: "black"}

\item{cex}{(optional, numeric) multiplicator for the text size. Default: 1}
}
\value{
A plot
}
\description{
This function is a simplified version of [fields::imagePlot()], by [Douglas Nychka](https://dnychka.github.io/). The original version is recommended in case more customization than the provided here is needed.
}
\examples{

#' library(distantia)

load("~/Dropbox/GITHUB/R_packages/distantia/data/sequencesMIS.RData")

#data frame with grouping column
###################################
data(sequencesMIS)

#prepare list of sequences
x <- prepare_sequences(
  x = sequencesMIS,
  id_column = "MIS",
  time_column = NULL,
  paired_samples = FALSE,
  pseudo_zero =  0.001,
  na_action = "to_zero"
)

#distance matrix of the first two sequences
dist_matrix <- distance_matrix(
  a = x[[1]],
  b = x[[2]],
  distance = "euclidean"
)

#cost matrix
cost_matrix <- cost_matrix(
  dist_matrix = dist_matrix
  )

#least cost path
path <- cost_path(
  dist_matrix = dist_matrix,
  cost_matrix = cost_matrix
)

#plot cost matrix and least cost path
plot_matrix(
  m = cost_matrix,
  path = path
  )

}
