% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sequence.R
\name{plot_sequence}
\alias{plot_sequence}
\title{Plot Multivariate Sequence}
\usage{
plot_sequence(
  x = NULL,
  center = FALSE,
  scale = FALSE,
  xlim = NULL,
  ylim = NULL,
  color = NULL,
  width = 1,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  text_cex = 1,
  box = FALSE,
  guide = TRUE,
  guide_position = "topright",
  guide_cex = 0.8,
  vertical = FALSE,
  subpanel = FALSE
)
}
\arguments{
\item{x}{(required, matrix) One element of a list generated by [prepare_sequences()]. Default: NULL}

\item{center}{(optional, logical) If TRUE, the columns of `x` are centered via [scale()]. Default: FALSE}

\item{scale}{(optional, logical) If TRUE, the columns of `x` are scaled via [scale()]. Default: FALSE}

\item{xlim}{(optional, numeric vector) Numeric vector with the limits of the x axis. Default: NULL}

\item{ylim}{(optional, numeric vector) Numeric vector with the limits of the x axis. Default: NULL}

\item{color}{(optional, character vector) vector of colors for the distance or cost matrix. If NULL, uses an appropriate palette generated with [grDevices::palette.colors()]. Default: NULL}

\item{width}{(optional, numeric vector) Widths of the sequence curves. Default: 1}

\item{title}{(optional, character string) Main title of the plot. If NULL, it's set to the name of the sequence. Default: NULL}

\item{xlab}{(optional, character string) Title of the x axis. Disabled if `subpanel` or `vertical` are TRUE. If NULL, the word "Time" is used. Default: NULL}

\item{ylab}{(optional, character string) Title of the x axis. Disabled if `subpanel` or `vertical` are TRUE. If NULL, it is left empty. Default: NULL}

\item{text_cex}{(optional, numeric) Multiplicator of the text size. Default: 1}

\item{guide}{(optional, logical) If TRUE, plots a legend. Default: TRUE}

\item{guide_position}{(optional, character string) Position of the guide. Accepted values are: "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center". Default: "topright"}

\item{guide_cex}{(optional, numeric) Size of the guide's text and separation between the guide's rows. Default: 0.7.}

\item{vertical}{(optional, logical) For internal use within the package in multipanel plots. Switches the plot axes. Disabled if `subpanel = FALSE`. Default: FALSE}

\item{subpanel}{(optional, logical) For internal use within the package in multipanel plots. Strips down the plot for a sub-panel. Default: FALSE}
}
\value{
A plot.
}
\description{
Plot Multivariate Sequence
}
