% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils_line_guide.R
\name{plot_utils_line_guide}
\alias{plot_utils_line_guide}
\title{Legend for Sequence Plots}
\usage{
plot_utils_line_guide(
  x,
  position = "topright",
  color = NULL,
  width = 1,
  text_cex = 0.7,
  ncol = 1,
  subpanel = FALSE
)
}
\arguments{
\item{x}{(required, sequence) a single sequence or set of sequences generated via [ts_prepare()]. Default: NULL}

\item{position}{(optional, vector of xy coordinates or character string). This is a condensed version of the `x` and `y` arguments of the [graphics::legend()] function. Coordinates (in the range 0 1) or keyword to position the legend. Accepted keywords are: "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center". Default: "topright".}

\item{color}{(optional, character vector) vector of colors for the sequence columns. If NULL, uses the palette "Zissou 1" provided by the function [grDevices::hcl.colors()]. Default: NULL}

\item{width}{(optional, numeric vector) Widths of the sequence curves. Default: 1}

\item{text_cex}{(optional, numeric) Multiplier of the text size. Default: 0.7}

\item{ncol}{(optional, integer) Number of columns in which to set the legend items. Default: 1.}

\item{subpanel}{(optional, logical) internal argument used when generating the multipanel plot produced by [plot_distantia()].}
}
\value{
A plot.
}
\description{
Legend for Sequence Plots
}
\examples{
data(sequencesMIS)

x <- ts_prepare(
  x = sequencesMIS,
  id_column = "MIS"
)

plot_utils_line_guide(
  x = x[["MIS-9"]],
  position = "center"
)
}
