% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia.R
\docType{data}
\name{pollenGP}
\alias{pollenGP}
\title{Pollen dataset.}
\format{
Dataframe with 42 columns and 200 rows
}
\usage{
data(sequenceA)
}
\description{
A subset of the Grande Pile dataset (\url{https://doi.pangaea.de/10.1594/PANGAEA.739275}). It contains a depth (cm) and age columns (ky BP), and 40 pollen types.
}
\keyword{datasets}
