% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ab.R
\name{prepare_ab}
\alias{prepare_ab}
\title{Prepare Two Sequences for Dissimilarity Analysis}
\usage{
prepare_ab(
  a = NULL,
  b = NULL,
  method = "euclidean",
  paired_samples = FALSE,
  pseudo_zero = 1e-04
)
}
\arguments{
\item{a}{(required, data frame or matrix) a time series.}

\item{b}{(required, data frame or matrix) a time series.}

\item{method}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `methods`. Default: "euclidean".}

\item{paired_samples}{(optional, logical vector) If TRUE, time-series are compared row wise and no least-cost path is computed. Default: FALSE.}

\item{pseudo_zero}{(optional, numeric) value to replace zero values if distance metrics "psi" and "cosine" are used. Default: 0.0001.}
}
\value{
A list containing the prepared sequences'a' and 'b'.
}
\description{
This function prepares two sequences 'a' and 'b' for dissimilarity analysis. Please make sure that 'a' and 'b' have matching column names, as the output will only contain their common columns.
}
\details{
This function checks and preprocesses the input data 'a' and 'b' for compatibility and consistency in preparation for dissimilarity analysis.
}
\examples{

ab <- prepare_ab(
 a = na.omit(sequenceA),
 b = na.omit(sequenceB),
 method = "euclidean",
 paired_samples = FALSE
 )

 a <- ab$a
 b <- ab$b

}
