% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_column_names.R
\name{prepare_column_names}
\alias{prepare_column_names}
\title{Checks Column Names in a List of Data Frames}
\usage{
prepare_column_names(x = NULL)
}
\arguments{
\item{x}{(required, list of data frames) Default: NULL}
}
\value{
Used for it's side effects: message with names of exclusive column names
}
\description{
Checks Column Names in a List of Data Frames
}
