% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_list_class.R
\name{prepare_list_class}
\alias{prepare_list_class}
\title{Checks Classes of List Elements Against Expectation}
\usage{
prepare_list_class(x = NULL, expected_class = "data.frame")
}
\arguments{
\item{x}{(required, list) Default: NULL}

\item{expected_class}{(required, class name). One of "data.frame", "matrix", or "vector". Default: "data.frame".}
}
\value{
Used for its side effects. Error message if the observed classes do not match the expectation.
}
\description{
Checks Classes of List Elements Against Expectation
}
