% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_matrix.R
\name{prepare_matrix}
\alias{prepare_matrix}
\title{Matrix to Data Frame}
\usage{
prepare_matrix(x = NULL)
}
\arguments{
\item{x}{(required, matrix) Default: NULL}
}
\value{
A data frame
}
\description{
Matrix to Data Frame
}
