% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_na.R
\name{prepare_na}
\alias{prepare_na}
\title{Handles NA in List of Data Frames}
\usage{
prepare_na(x = NULL, pseudo_zero = NULL, na_action = NULL)
}
\arguments{
\item{x}{(required, list of data frames) A named list with data frames. Default: NULL.}

\item{pseudo_zero}{(optional, numeric) Value used to replace zeros in the data. Default: NULL.}

\item{na_action}{(optional, character string) Action to handle missing values. default: NULL.
\itemize{
  \item NULL: returns the input without changes.
  \item "omit": applies [na.omit()] to each sequence.
  \item "to_zero": replaces NA values with zero or `pseudo_zero`, if provided.
  \item "impute": not implemented yet.
}}
}
\value{
Named list of data frames
}
\description{
Handles NA in List of Data Frames
}
