% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sequences.R
\name{prepare_sequences}
\alias{prepare_sequences}
\title{Prepare Data for Dissimilarity Analysis}
\usage{
prepare_sequences(
  x = NULL,
  id_column = NULL,
  time_column = NULL,
  paired_samples = FALSE,
  pseudo_zero = NULL,
  na_action = "omit",
  transformation = NULL
)
}
\arguments{
\item{x}{(required, list or data frame) A named list with ordered sequences, or a long data frame with a grouping column. Default: NULL.}

\item{id_column}{(optional, column name) Column name used for splitting a 'x' data frame into a list.}

\item{time_column}{(optional if `paired_samples = FALSE`, and required otherwise, column name) Name of the column representing time, if any. Default: NULL.}

\item{paired_samples}{(optional, logical) If TRUE, all input sequences are subset to their common times according to the values in the `time_column`.}

\item{pseudo_zero}{(optional, numeric) Value used to replace zeros in the data. Default: NULL.}

\item{na_action}{(optional, character string) Action to handle missing values. default: NULL.
\itemize{
  \item NULL: returns the input without changes.
  \item "omit": applies [na.omit()] to each sequence.
  \item "to_zero": replaces NA values with zero or `pseudo_zero`, if provided.
  \item "impute": not implemented yet.
}}

\item{transformation}{(optional, function) A function to transform the data within each sequence. A few options are:
\itemize{
  \item [f_proportion()]: to transform counts into proportions.
  \item [f_percentage()]: to transform counts into percentages.
  \item [f_hellinger()]: to apply a Hellinger transformation.
  \item [f_scale()]: to center and scale the data.
  }}
}
\value{
A named list of data frames, matrices, or vectors.
}
\description{
This function performs the following steps:
- Converts a data frame to a list by 'id_column'.
- Checks the validity of input x.
- Orders x by time if a time column is PROVIDED
- Handles missing values according to the specified action.
- Handles zeros by replacing them with a pseudo-zero value, if provided.
- Transforms the data if a transformation function is provided.
}
\examples{
data(sequencesMIS)
x <- prepare_sequences(
  x = sequencesMIS,
  id_column = "MIS"
)
}
