% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_time.R
\name{prepare_time}
\alias{prepare_time}
\title{Handles Time Column in List of Data Frames}
\usage{
prepare_time(x = NULL, time_column = NULL, paired_samples = NULL)
}
\arguments{
\item{x}{(required, named list of data frames). List with named data frames. Default: NULL.}

\item{time_column}{(optional if `paired_samples = FALSE`, and required otherwise, column name) Name of numeric column representing time. Default: NULL.}

\item{paired_samples}{(optional, logical) If TRUE, all input sequences are subset to their common times according to the values in the `time_column`. Default: FALSE.}
}
\value{
List of data frames
}
\description{
Handles Time Column in List of Data Frames
}
