% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_transformation.R
\name{prepare_transformation}
\alias{prepare_transformation}
\title{Handles Transformation in List of Data Frames}
\usage{
prepare_transformation(
  x = NULL,
  id_column = NULL,
  time_column = NULL,
  transformation = NULL
)
}
\arguments{
\item{x}{(required, data frame) Default: NULL.}

\item{transformation}{(optional, function) A function to transform the data within each sequence. A few options are:
\itemize{
  \item [f_proportion()]: to transform counts into proportions.
  \item [f_percentage()]: to transform counts into percentages.
  \item [f_hellinger()]: to apply a Hellinger transformation.
  \item [f_scale()]: to center and scale the data.
  }}
}
\value{
Named list of data frames
}
\description{
Handles Transformation in List of Data Frames
}
