% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_xy.R
\name{prepare_xy}
\alias{prepare_xy}
\title{Prepare Two Sequences for Dissimilarity Analysis}
\usage{
prepare_xy(x = NULL, y = NULL, distance = "euclidean", lock_step = FALSE)
}
\arguments{
\item{x}{(required, data frame, matrix, or numeric vector) a sequence. Default: NULL}

\item{y}{(required, data frame, matrix, or numeric vector) a sequence. Default: NULL}

\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}

\item{lock_step}{(optional, logical vector) If TRUE, time-series are compared row wise and no least-cost path is computed. Default: FALSE.}
}
\value{
A list containing the prepared sequences'x' and 'y'.
}
\description{
Prepares two sequences 'x' and 'y' for dissimilarity analysis. Please make sure that 'x' and 'y' have matching column names, as the output will only contain their common columns.
}
\details{
This function checks and preprocesses the input data 'x' and 'y' for compatibility and consistency in preparation for dissimilarity analysis.
}
