% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_zeros.R
\name{prepare_zeros}
\alias{prepare_zeros}
\title{Handles Zeros in List of Data Frames}
\usage{
prepare_zeros(x = NULL, pseudo_zero = NULL)
}
\arguments{
\item{x}{(required, list of data frames) A named list with data frames. Default: NULL.}

\item{pseudo_zero}{(optional, numeric) Value used to replace zeros in the data. Default: NULL.}
}
\value{
Named list of data frames
}
\description{
Handles Zeros in List of Data Frames
}
