% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSequences.R
\name{prepareSequences}
\alias{prepareSequences}
\title{Prepare input sequences.}
\usage{
prepareSequences(
sequence.A = NULL,
sequence.A.name = "A",
sequence.B = NULL,
sequence.B.name = "B",
merge.mode = "complete",
sequences = NULL,
grouping.column = NULL,
time.column = NULL,
exclude.columns = NULL,
if.empty.cases = "zero",
transformation = "none",
silent = FALSE
)
}
\arguments{
\item{sequence.A}{dataframe containing a multivariate time-series.}

\item{sequence.A.name}{character string with the name of the sequence.}

\item{sequence.B}{dataframe containing a multivariate time-series. Must have same column names and units as \code{sequence.A}.}

\item{sequence.B.name}{character string with the name of the sequence.}

\item{merge.mode}{character string, one of: "overlap", "complete" (default optiion). If "overlap", \code{sequence.A} and \code{sequence.B} are merged by their common columns, and non-common columns are removed. If "complete", columns absent in one dataset but present in the other are added, with values equal to 0. This argument is ignored if \code{sequence.A} and \code{sequence.B} are not provided.}

\item{sequences}{dataframe with multiple sequences identified by a grouping column.}

\item{grouping.column}{character string, name of the column in \code{sequences} to be used to identify separates sequences within the file. This argument is ignored if \code{sequence.A} and \code{sequence.B} are provided.}

\item{exclude.columns}{character string or character vector with column names in \code{sequences}, or \code{squence.A} and \code{sequence.B} to be excluded from the analysis.}

\item{if.empty.cases}{character string with two possible values: "omit", or "zero". If "zero" (default), \code{NA} values are replaced by zeroes. If "omit", rows with \code{NA} data are removed.}

\item{transformation}{character string. Defines what data transformations to apply to the sequences. One of: "none" (default), "percentage", "proportion", and "hellinger".}

\item{silent}{boolean, set to TRUE to hide all messages, and set to FALSE otherwise.}

\item{time.colum}{character string, name of the column with time/depth/rank data. The data in this column is not modified.}
}
\value{
A list with four slots:
\itemize{
\item \emph{taxa} Common column names of the sequences, listing the taxa included in them.
\item \emph{metadata} A dataframe with details about the processing of the sequences.
\item \emph{sequence.A} Dataframe, a processed pollen sequence.
\item \emph{sequence.B} Dataframe, a processed pollen sequence.
}
}
\description{
Takes two dataframes containing multivariate time-series (\code{sequence.A} and \code{sequence.B}), checks the integrity of both datasets, and tries to adapt \code{sequence.B} to the characteristics of\code{sequence.A} by matching number of columns, column names, and missing data. Missing data can be handled in two ways: 1) setting NA to zero; 2) deleting rows with NA or empty cases.
}
\examples{

#two sequences as inputs
data(sequenceA)
data(sequenceB)

sequences <- prepareSequences(
 sequence.A = sequenceA,
 sequence.A.name = "A",
 sequence.B = sequenceB,
 sequence.B.name = "B",
 merge.mode = "complete",
 if.empty.cases = "zero",
 transformation = "hellinger"
 )


#several sequences in a single dataframe
data(sequencesMIS)
sequences <- prepareSequences(
 sequences = sequencesMIS,
 grouping.column = "MIS",
 if.empty.cases = "zero",
 transformation = "hellinger"
 )

}
\author{
Blas Benito <blasbenito@gmail.com>
}
