% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_auto_distance.R
\name{psi_auto_distance}
\alias{psi_auto_distance}
\title{Sum of Auto Distances Between Adjacent Cases in a Sequence}
\usage{
psi_auto_distance(x = NULL, path = NULL, distance = "euclidean")
}
\arguments{
\item{x}{(required, data frame or matrix) a sequence.}

\item{path}{(optional, data frame) result of \code{\link[=psi_cost_path_trim_blocks]{psi_cost_path_trim_blocks()}}, only required if blocks in the least cost path were ignored.}

\item{distance}{(optional, character string) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset \code{distances}. Default: "euclidean".}
}
\value{
Sum of auto distances
}
\description{
Sum of Auto Distances Between Adjacent Cases in a Sequence
}
\examples{
#TODO
}
