% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_auto_sum.R
\name{psi_auto_sum}
\alias{psi_auto_sum}
\title{Auto Sum}
\usage{
psi_auto_sum(x = NULL, y = NULL, path = NULL, distance = "euclidean")
}
\arguments{
\item{x}{(required, zoo object or numeric matrix) univariate or multivariate time series with no NAs. Default: NULL.}

\item{y}{(required, zoo object or numeric matrix) a time series with the same number of columns as \code{x} and no NAs. Default: NULL.}

\item{path}{(required, data frame) least cost path produced by \code{\link[=psi_cost_path]{psi_cost_path()}}. Only required when \code{\link[=psi_cost_path_ignore_blocks]{psi_cost_path_ignore_blocks()}} has been applied to \code{path}. Default: NULL.}

\item{distance}{(optional, character string) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset \code{distances}. Default: "euclidean".}
}
\value{
numeric vector
}
\description{
Demonstration function to computes the sum of distances between consecutive samples in two time series.
}
\examples{
#simulate two zoo objects
tsl <- tsl_simulate(
  n = 2,
  seed = 1
)

if(interactive()){
  tsl_plot(tsl = tsl)
}

#auto sum of distances
psi_auto_sum(
  x = tsl[[1]],
  y = tsl[[2]],
  distance = "euclidean"
)

#same as:
x_sum <- psi_auto_distance(
  x = tsl[[1]],
  distance = "euclidean"
)

y_sum <- psi_auto_distance(
  x = tsl[[2]],
  distance = "euclidean"
)

x_sum + y_sum
}
\seealso{
Other psi_demo: 
\code{\link{distance}()},
\code{\link{distances}},
\code{\link{psi}()},
\code{\link{psi_auto_distance}()},
\code{\link{psi_cost_matrix}()},
\code{\link{psi_cost_path}()},
\code{\link{psi_cost_path_ignore_blocks}()},
\code{\link{psi_cost_path_sum}()},
\code{\link{psi_dist_lock_step}()},
\code{\link{psi_dist_matrix}()}
}
\concept{psi_demo}
