% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psi_auto_sum_cpp}
\alias{psi_auto_sum_cpp}
\title{Auto Sum of Two Time Series}
\usage{
psi_auto_sum_cpp(x, y, path, distance = "euclidean", ignore_blocks = FALSE)
}
\arguments{
\item{x}{(required, numeric matrix) of same number of columns as 'y'.}

\item{y}{(required, numeric matrix) of same number of columns as 'x'.}

\item{path}{(required, data frame) dataframe produced by \code{\link[=cost_path_orthogonal_cpp]{cost_path_orthogonal_cpp()}}.
Default: NULL}

\item{distance}{(optional, character string) name or abbreviation of the
distance method. Valid values are in the columns "names" and "abbreviation"
of the dataset \code{distances}. Default: "euclidean".}

\item{ignore_blocks}{(optional, logical). If TRUE, blocks of consecutive path
coordinates are trimmed to avoid inflating the psi distance. Default: FALSE.}
}
\value{
Auto sum of matrices y and x.
}
\description{
Cumulative sum of two time series \code{x} and \code{y}
to compute the normalization factor required for psi dissimilarity scores.
}
