% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_cost_path.R
\name{psi_cost_path}
\alias{psi_cost_path}
\title{Least Cost Path}
\usage{
psi_cost_path(dist_matrix = NULL, cost_matrix = NULL, diagonal = FALSE)
}
\arguments{
\item{dist_matrix}{(required, numeric matrix) Distance matrix.}

\item{cost_matrix}{(required, numeric matrix) Cost matrix generated from the distance matrix.}

\item{diagonal}{(optional, logical) If TRUE, diagonals are used during the least-cost path search. Default: FALSE}
}
\value{
A data frame with least-cost path coordiantes.
}
\description{
Least Cost Path
}
\examples{

x <- sequenceB |>
  na.omit() |>
  as.matrix()

y <- sequenceA |>
  na.omit() |>
  as.matrix()

d <- distance_matrix(x, y, distance = "euclidean")

m <- cost_matrix(dist_matrix = d)

path <- psi_cost_path(
  dist_matrix = d,
  cost_matrix = m
  )

}
