% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_cost_path_trim_blocks.R
\name{psi_cost_path_trim_blocks}
\alias{psi_cost_path_trim_blocks}
\title{Trims Blocks from a Least Cost Path}
\usage{
psi_cost_path_trim_blocks(path = NULL)
}
\arguments{
\item{path}{(required, data frame) dataframe produced by \code{\link[=psi_cost_path]{psi_cost_path()}}. Default: NULL}
}
\value{
A data frame with a least-cost path with no blocks.
}
\description{
Removes blocks (straight segments of the path that appear in highly dissimilar regions of the sequences) from least-cost paths. Blocks inflate psi values when two sequences are similar but have very different numbers of rows.
}
\examples{
#TODO: complete example
}
\seealso{
Other psi_demo: 
\code{\link{psi_auto_distance}()},
\code{\link{psi_auto_sum}()},
\code{\link{psi_cost_matrix}()},
\code{\link{psi_cost_path}()},
\code{\link{psi_cost_path_sum}()},
\code{\link{psi_dist_matrix}()}
}
\concept{psi_demo}
