% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_dist_lock_step.R
\name{psi_dist_lock_step}
\alias{psi_dist_lock_step}
\title{Lock-Step Distance}
\usage{
psi_dist_lock_step(x = NULL, y = NULL, distance = "euclidean")
}
\arguments{
\item{x}{(required, zoo object or numeric matrix) a time series with no NAs. Default: NULL}

\item{y}{(zoo object or numeric matrix) a time series with the same columns as \code{x} and no NAs. Default: NULL}

\item{distance}{(optional, character string) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset \code{distances}. Default: "euclidean".}
}
\value{
numeric
}
\description{
Demonstration function to compute the lock-step distance between two univariate or multivariate time series.

This function does not accept NA data in the matrices \code{x} and \code{y}.
}
\examples{
#simulate two time series
tsl <- tsl_simulate(
  n = 2,
  seed = 1
)

if(interactive()){
  tsl_plot(tsl = tsl)
}

#distance matrix
d <- psi_dist_lock_step(
  x = tsl[[1]],
  y = tsl[[2]],
  distance = "euclidean"
)

d
}
\seealso{
Other psi_demo: 
\code{\link{distance}()},
\code{\link{distances}},
\code{\link{psi}()},
\code{\link{psi_auto_distance}()},
\code{\link{psi_auto_sum}()},
\code{\link{psi_cost_matrix}()},
\code{\link{psi_cost_path}()},
\code{\link{psi_cost_path_ignore_blocks}()},
\code{\link{psi_cost_path_sum}()},
\code{\link{psi_dist_matrix}()}
}
\concept{psi_demo}
