% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_dist_matrix.R
\name{psi_dist_matrix}
\alias{psi_dist_matrix}
\title{Distance Matrix}
\usage{
psi_dist_matrix(x = NULL, y = NULL, distance = "euclidean")
}
\arguments{
\item{x}{(required, data frame or matrix) a time series.}

\item{y}{(required, data frame or matrix) a time series.}

\item{distance}{(optional, character string) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}
}
\value{
A distance matrix.
}
\description{
Distance Matrix
}
