% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_distance_matrix.R
\name{psi_distance_matrix}
\alias{psi_distance_matrix}
\title{Distance Matrix}
\usage{
psi_distance_matrix(x = NULL, y = NULL, distance = "euclidean")
}
\arguments{
\item{x}{(required, zoo object or numeric matrix) a time series with no NAs. Default: NULL}

\item{y}{(zoo object or numeric matrix) a time series with the same columns as \code{x} and no NAs. Default: NULL}

\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset \link{distances}. Default: "euclidean".}
}
\value{
numeric matrix
}
\description{
Demonstration function to compute the distance matrix between two univariate or multivariate time series.

This function does not accept NA data in the matrices \code{x} and \code{y}.
}
\examples{
#distance metric
d <- "euclidean"

#simulate two irregular time series
x <- zoo_simulate(
  name = "x",
  rows = 100,
  seasons = 2,
  seed = 1
)

y <- zoo_simulate(
  name = "y",
  rows = 80,
  seasons = 2,
  seed = 2
)

if(interactive()){
  zoo_plot(x = x)
  zoo_plot(x = y)
}

#distance matrix
dist_matrix <- psi_distance_matrix(
  x = x,
  y = y,
  distance = d
)

if(interactive()){
  utils_matrix_plot(
    m = dist_matrix
    )
}
}
\seealso{
Other psi_demo: 
\code{\link{d}()},
\code{\link{distances}},
\code{\link{psi_auto_distance}()},
\code{\link{psi_auto_sum}()},
\code{\link{psi_cost_matrix}()},
\code{\link{psi_cost_path}()},
\code{\link{psi_cost_path_sum}()},
\code{\link{psi_distance_lock_step}()},
\code{\link{psi_equation}()}
}
\concept{psi_demo}
