% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psi_paired_cpp}
\alias{psi_paired_cpp}
\title{Computes Psi Distance Between Two Time-Series With Paired Samples}
\usage{
psi_paired_cpp(a, b, method = "euclidean")
}
\arguments{
\item{a}{(required, numeric matrix).}

\item{b}{(required, numeric matrix) of same number of columns as 'a'.}

\item{method}{(optional, character string) name or abbreviation of the
distance method. Valid values are in the columns "names" and "abbreviation"
of the dataset `methods`. Default: "euclidean".}
}
\value{
Psi distance
}
\description{
Computes the distance psi between two matrices
\code{a} and \code{b} with the same number of columns and rows. Distances
between \code{a} and \code{b} are computed row wise rather than via distance
matrix and least-cost path computation.
NA values should be removed before using this function.
If the selected distance function is "chi" or "cosine", pairs of zeros should
be either removed or replaced with pseudo-zeros (i.e. 0.00001).
}
