% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_plot.R
\name{psi_plot}
\alias{psi_plot}
\title{Multipanel Plot Comparing Two Sequences}
\usage{
psi_plot(
  x = NULL,
  y = NULL,
  distance = "euclidean",
  diagonal = FALSE,
  weighted = FALSE,
  ignore_blocks = FALSE,
  matrix_type = "cost",
  matrix_color = NULL,
  path_width = 1,
  path_color = "black",
  line_center = FALSE,
  line_scale = FALSE,
  line_color = NULL,
  line_width = 1,
  text_cex = 1
)
}
\arguments{
\item{x}{(required, sequence) a sequence generated via [tsl_prepare()] for the X axis of the multi-panel plot. Default: NULL}

\item{y}{(required, sequence) a sequence generated via [tsl_prepare()] for the Y axis of the multi-panel plot. Default: NULL}

\item{distance}{(optional, character STRING) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}

\item{diagonal}{(optional, logical). If TRUE, diagonals are included in the computation of the cost matrix. Default: FALSE.}

\item{weighted}{@param weighted (optional, logical) If TRUE, diagonal is set to TRUE, and diagonal cost is weighted by a factor of 1.414214. Default: FALSE.}

\item{ignore_blocks}{(optional, logical). If TRUE, blocks of consecutive path coordinates are trimmed to avoid inflating the psi distance. Default: FALSE.}

\item{matrix_type}{(optional, character string): one of "cost" or "distance" (the abbreviation "dist" is accepted as well). Default: "cost".}

\item{matrix_color}{(optional, character vector) vector of colors for the distance or cost matrix. If NULL, uses the palette "Zissou 1" provided by the function [grDevices::hcl.colors()]. Default: NULL}

\item{path_width}{(optional, numeric) width of the least cost path. Default: 1}

\item{path_color}{(optional, character string) color of the least-cost path. Default: "black"}

\item{line_center}{(optional, logical) If TRUE, the sequences are centered via [scale()]. This centering is only for plotting purposes, and it is ignored when computing the dissimilarity between the sequences. Default: FALSE}

\item{line_scale}{(optional, logical) If TRUE, the sequences are scaled via [scale()]. This scaling is only for plotting purposes, and it is ignored when computing the dissimilarity between the sequences. Default: FALSE}

\item{line_color}{(optional, character vector) Vector of colors for the sequence curves. If not provided, defaults to a subset of `matrix_color`.}

\item{line_width}{(optional, numeric vector) Widths of the sequence curves. Default: 1}

\item{text_cex}{(optional, numeric) Multiplier of the text size. Default: 1}
}
\value{
A plot.
}
\description{
Plots two sequences, their distance or cost matrix, and their least cost path.

Unlike [distantia()], this function does not accept vectors as inputs for the arguments to compute dissimilarity (these are `distance`, `diagonal`, `weighted`, and `ignore_blocks`).

The argument `paired_samples` is not available because this multipanel plot does not make sense in such a case.
}
\examples{
data(sequencesMIS)

x <- tsl_prepare(
  x = sequencesMIS,
  id_column = "MIS"
)

plot_distantia(
  y = x[["MIS-9"]],
  x = x[["MIS-11"]]
)
}
