% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_value.R
\name{psi_value}
\alias{psi_value}
\title{Normalized Dissimilarity Score}
\usage{
psi_value(cost_path_sum = NULL, auto_sum = NULL, diagonal = FALSE)
}
\arguments{
\item{auto_sum}{(required, numeric) Result of [auto_sum()] on the sequences to compare. Default: NULL}

\item{diagonal}{(optional, logical) If the cost matrix and least cost path were computed using `diagonal = TRUE`, this argument should be `TRUE` as well. Used to correct the computation of `psi` when diagonals are used.}

\item{path_sum}{(required, numeric) Result of [cost_path_sum()] on the least cost path between the sequences to compare.}
}
\value{
Psi value.
}
\description{
Computes the dissimilarity metric \code{psi} (Birks and Gordon 1985).
}
\details{
The measure of dissimilarity \code{psi} is computed as: \code{least.cost - (autosum of sequences)) / autosum of sequences}. It has a lower limit at 0, while there is no upper limit.
}
\author{
Blas Benito <blasbenito@gmail.com>

\itemize{
\item Birks, H.J.B.  and Gordon, A.D. (1985) Numerical Methods in Quaternary Pollen Analysis. Academic Press.
}
}
