% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiToDataframe.R
\name{psiToDataframe}
\alias{psiToDataframe}
\title{Turns the output of \code{\link{psi}} into a dataframe.}
\usage{
psiToDataframe(psi.values = NULL)
}
\arguments{
\item{psi.values}{list produced by \code{\link{psi}}.}
}
\description{
Parses a list produced by \code{\link{psi}} to generate a dataframe with psi values for every pair of sequences available in the data. This dataframe allows an easy plotting of the dissimilarity pattern.
}
\author{
Blas Benito <blasbenito@gmail.com>
}
