% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiToMatrix.R
\name{psiToMatrix}
\alias{psiToMatrix}
\title{Turns the output of \code{\link{psi}} into a matrix.}
\usage{
psiToMatrix(psi.values = NULL)
}
\arguments{
\item{psi.values}{list produced by \code{\link{psi}}.}
}
\description{
Parses a list produced by \code{\link{psi}} to generate a matrix of psi values among every pair of sequences available in the data. This matrix allows an easy plotting of the dissimilarity pattern, either as a heatmap, or as a network.
}
\author{
Blas Benito <blasbenito@gmail.com>
}
