% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{scaling_parameters}
\alias{scaling_parameters}
\title{Handles Centering and Scaling within tsl_transform()}
\usage{
scaling_parameters(tsl = NULL, f = NULL, ...)
}
\arguments{
\item{tsl}{(required, list of zoo objects) List of time series. Default: NULL}

\item{f}{(required, transformation function) name of a function taking a matrix as input.}

\item{...}{(optional, arguments of `f`) Optional arguments for the transformation function.}
}
\value{
Named list with center and scale parameters
}
\description{
Handles Centering and Scaling within tsl_transform()
}
