% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia.R
\docType{data}
\name{sequenceA}
\alias{sequenceA}
\title{Multivariate and irregular time series with pollen counts.}
\format{Dataframe with 9 columns and 49 rows}
\usage{
data(sequenceA)
}
\description{
A dataframe with 9 columns representing pollen types (betula, pinus, corylus, empetrum, cypera, artemisia, rumex) and 49 rows representing increasing depths with pollen counts taken from the Abernethy dataset (Birks and Mathewes (1978).
}
\references{
Birks, H.H. and Mathewes, R.W. (1978) Studies in the vegetational history of Scotland. \emph{New Phytologist} \strong{80}, 455-484.
}
\keyword{datasets}
