% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia.R
\docType{data}
\name{sequenceB}
\alias{sequenceB}
\title{Multivariate and irregular time series with pollen counts.}
\format{
Dataframe with 9 columns and 41 rows
}
\usage{
data(sequenceB)
}
\description{
A dataframe with 8 columns (the column \code{empetr} is missing with respect to \code{\link{sequenceA}}) representing pollen types (betula, pinus, corylus, cypera, artemisia, rumex) and 41 rows representing increasing depths with pollen counts taken from the Abernethy dataset (Birks and Mathewes (1978). Several NA values have been introduced in the dataset to demonstrate the data-handling capabilities of \code{\link{prepareSequences}}.
}
\references{
Birks, H.H. and Mathewes, R.W. (1978) Studies in the vegetational history of Scotland. \emph{New Phytologist} \strong{80}, 455-484.
}
\keyword{datasets}
