% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sequencesMIS}
\alias{sequencesMIS}
\title{Dataframe with pollen counts for different MIS stages.}
\format{
dataframe with 7 columns and 427 rows.
}
\usage{
data(sequencesMIS)
}
\description{
A dataframe with 427 rows representing pollen counts for 12 marine isotope stages and 6 pollen types
}
\keyword{datasets}
