% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slotting.R, R/slotting_multiple.R
\name{slotting}
\alias{slotting}
\title{Combination of Sequences via Least Cost Paths}
\usage{
slotting(x = NULL, distance = "euclidean")

slotting(x = NULL, distance = "euclidean")
}
\arguments{
\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}

\item{a}{(required, data frame, matrix, or numeric vector) a time series.}

\item{b}{(required, data frame, matrix, or numeric vector) a time series.}
}
\value{
Data frame with composite sequence and columns identifying the origin of each new row.

Data frame with composite sequence and columns identifying the origin of each new row.
}
\description{
Combination of Sequences via Least Cost Paths

Combination of Sequences via Least Cost Paths
}
