% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subset_matrix_by_rows_cpp}
\alias{subset_matrix_by_rows_cpp}
\title{Subset Matrix by Rows}
\usage{
subset_matrix_by_rows_cpp(m, rows)
}
\arguments{
\item{m}{(required, numeric matrix) a sequence.}

\item{rows}{(required, integer vector) vector of rows to subset.}
}
\value{
Numeric matrix of length rows.
}
\description{
Subsets the matrix \code{m} to the indices given in the vector \code{rows}.
Used to subset a sequence to the coordinates of a trimmed least-cost path.
}
