% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_count_NA.R
\name{ts_count_NA}
\alias{ts_count_NA}
\title{Counts NA Cases in Time Series}
\usage{
ts_count_NA(x = NULL, verbose = TRUE)
}
\arguments{
\item{x}{(required, list of zoo objects) List of time series. Default: NULL}

\item{verbose}{(optional, logical) If FALSE, all messages are suppressed. Default: TRUE}
}
\value{
Integer, sum of NA cases
}
\description{
Counts NA Cases in Time Series
}
