% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_names.R
\name{ts_get_names}
\alias{ts_get_names}
\title{Get Names of a Time Series List}
\usage{
ts_get_names(x = NULL)
}
\arguments{
\item{x}{(required, zoo time series or time series list) Individual time series or time series list created with [ts_initialize]. Default: NULL}
}
\value{
Character vector of names.
}
\description{
A list of zoo objects created with [ts_initialize()] has two sets of names that should ideally be the same: the list names, accessed with `names(x)`, and the names of the individual zoo objects (stored in their attribute "name"), accessed with `ts_get_names(x)`.
}
