% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_initialize.R
\name{ts_initialize}
\alias{ts_initialize}
\alias{ts_init}
\title{Prepare Data for Dissimilarity Analysis}
\usage{
ts_initialize(
  x = NULL,
  id_column = NULL,
  time_column = NULL,
  paired_samples = FALSE,
  na_action = "impute",
  na_fill = 1e-04,
  transformation = NULL
)

ts_init(
  x = NULL,
  id_column = NULL,
  time_column = NULL,
  paired_samples = FALSE,
  na_action = "impute",
  na_fill = 1e-04,
  transformation = NULL
)
}
\arguments{
\item{x}{(required, list or data frame) A named list with ordered sequences, or a long data frame with a grouping column. Default: NULL.}

\item{id_column}{(optional, column name) Column name used for splitting a 'x' data frame into a list.}

\item{time_column}{(optional if `paired_samples = FALSE`, and required otherwise, column name) Name of the column representing time, if any. Default: NULL.}

\item{paired_samples}{(optional, logical) If TRUE, all input sequences are subset to their common times according to the values in the `time_column`.}

\item{na_action}{(required, character) Action to handle NA data in `x`. Current options are:
\itemize{
  \item "impute" (default): NA cases are interpolated against time via splines with [zoo::na.spline()].
  \item "omit": rows with NA cases are removed.
  \item "fill": NA cases are replaced with the value in `na_fill`.
}}

\item{na_fill}{(optional, numeric) Only relevant when `na_action = "fill"`, defines the value used to replace NAs with. Ideally, a small number different from zero (pseudo-zero). Default: 0.0001}

\item{transformation}{(optional, function) A function to transform the data within each sequence. A few options are:
\itemize{
  \item [f_proportion()]: to transform counts into proportions.
  \item [f_percentage()]: to transform counts into percentages.
  \item [f_hellinger()]: to apply a Hellinger transformation.
  \item [f_scale()]: to center and scale the data.
  }}
}
\value{
A named list of matrices.
}
\description{
This function performs the following steps:
- Converts a data frame to a list by 'id_column'.
- Checks the validity of input x.
- Orders x by time if a time column is PROVIDED
- Handles missing values according to the specified action.
- Handles zeros by replacing them with a pseudo-zero value, if provided.
- Transforms the data if a transformation function is provided.
}
\examples{
data(sequencesMIS)
x <- ts_prepare(
  x = sequencesMIS,
  id_column = "MIS"
)
}
