% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_is_valid.R
\name{ts_is_valid}
\alias{ts_is_valid}
\title{Checks Structure Validity of a Time Series List}
\usage{
ts_is_valid(x = NULL)
}
\arguments{
\item{x}{(required, list of zoo time series) Default NULL}
}
\value{
Logical
}
\description{
Checks that a time series list is in fact a named list containing zoo time series with matching attributes "name".
}
