% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_names.R
\name{ts_set_names}
\alias{ts_set_names}
\title{Set Names of a Time Series List}
\usage{
ts_set_names(x = NULL, names = NULL)
}
\arguments{
\item{x}{(required, list of zoo objects) List of time series. Default: NULL}

\item{names}{(optional, character vector) names to set. Must be of the same length of `x`. If NULL, and the list `x` has names, then the names of the zoo objects inside of the list are taken from the names of the list elements.}
}
\description{
Sets the names of a time series list created with [ts_initialize()] and the internal names of the zoo objects inside, stored in their attribute "name".
}
