% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_subset.R
\name{ts_subset}
\alias{ts_subset}
\title{Subset Sequences}
\usage{
ts_subset(x = NULL, sequences = NULL, variables = NULL, time = NULL)
}
\arguments{
\item{x}{(required, matrix) List of sequences generated by [ts_prepare()]. Default: NULL}

\item{sequences}{(optional, character or numeric vector) Character vector of sequence names or numeric vector of sequence indices in `x`. If NULL, all sequences are kept. Default: NULL}

\item{variables}{(optional, character vector) Column names of the sequences in `x`. If NULL, all variables are returned. Default: NULL}

\item{time}{(optional, numeric vector) numeric vector with range of times to keep from the sequences in `x`. Rows of sequences in `x` outside such range are removed. If NULL, all rows in `x` are preserved. Default: NULL}
}
\value{
#' @return A named list of matrices.
}
\description{
Manipulates the dimension of a set of sequences.
}
