% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_aggregate.R
\name{tsl_aggregate}
\alias{tsl_aggregate}
\title{Aggregates Time Series List}
\usage{
tsl_aggregate(tsl = NULL, breaks = NULL, f = mean, ...)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{breaks}{(required, numeric, numeric vector, Date vector, POSIXct vector, or keyword) breakpoints defining aggregation groups. Options are:
\itemize{
  \item numeric vector: only for the "numeric" time class, defines the breakpoints for time series aggregation.
  \item "Date" or "POSIXct" vector: as above, but for the time classes "Date" and "POSIXct." In any case, the input vector is coerced to the time class of the `tsl` argument.
  \item numeric: defines fixed with time intervals for time series aggregation. Used as is when the time class is "numeric", and coerced to integer and interpreted as days for the time classes "Date" and "POSIXct".
  \item keyword (see [utils_time_units()]): the common options for the time classes "Date" and "POSIXct" are: "millennia", "centuries", "decades", "years", "quarters", "months", and "weeks". Exclusive keywords for the "POSIXct" time class are: "days", "hours", "minutes", and "seconds". The time class "numeric" accepts keywords coded as scientific numbers, from "1e8" to "1e-8".
}}

\item{f}{(required, function) name without quotes and parenthesis of a standard or user-defined function to aggregate a vector series. Typical examples are `mean`, `max`,`min`, `median`, and `sd`, but others such as [f_slope()] are available as well. Default: `mean`.}

\item{...}{(optional) further arguments for `f`.}
}
\value{
time series list
}
\description{
Aggregates Time Series List
}
