% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_colnames.R
\name{tsl_clean_colnames}
\alias{tsl_clean_colnames}
\title{Cleans Column Names of a Time Series List}
\usage{
tsl_clean_colnames(
  tsl = NULL,
  as_title = FALSE,
  length = NULL,
  test_valid = TRUE
)
}
\arguments{
\item{tsl}{(required, time series list) Individual time series or time series list created with [tsl_initialize]. Default: NULL}

\item{test_valid}{(optional, logical) If FALSE, the validity test performed by [tsl_is_valid()] is ignored. Useful to improve performance when the user knows that `tsl` is valid. Default: TRUE}
}
\value{
time series list
}
\description{
Cleans Column Names of a Time Series List
}
