% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_names.R
\name{tsl_colnames}
\alias{tsl_colnames}
\title{Column Names of a Time Series List}
\usage{
tsl_colnames(tsl = NULL, common = FALSE)
}
\arguments{
\item{tsl}{(required, time series list) Default: NULL}

\item{common}{(optional, logical). If TRUE, only common column names to all time-series in `tsl` are returned as a character vector. If FALSE, a list with the column names of each time series is returned. Default: FALSE}
}
\value{
List if `common = FALSE`, character vector if `common = TRUE`.
}
\description{
Column Names of a Time Series List
}
