% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_colnames.R
\name{tsl_colnames_prefix}
\alias{tsl_colnames_prefix}
\title{Append Prefix to Column Names of Time Series List}
\usage{
tsl_colnames_prefix(tsl = NULL, prefix = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{prefix}{(optional, character string)  String to prepend to the column names. Default: NULL.}
}
\value{
time series list
}
\description{
Append Prefix to Column Names of Time Series List
}
\examples{
tsl <- tsl_simulate()

tsl_colnames_get(tsl = tsl)

tsl <- tsl_colnames_prefix(
  tsl = tsl,
  prefix = "my_prefix_"
)

tsl_colnames_get(tsl = tsl)
}
