% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_colnames.R
\name{tsl_colnames_suffix}
\alias{tsl_colnames_suffix}
\title{Append Suffix to Column Names of Time Series List}
\usage{
tsl_colnames_suffix(tsl = NULL, suffix = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{suffix}{(optional, character string) String to append to the column names. Default: NULL.}
}
\value{
time series list
}
\description{
Append Suffix to Column Names of Time Series List
}
\examples{
tsl <- tsl_simulate()

tsl_colnames_get(tsl = tsl)

tsl <- tsl_colnames_suffix(
  tsl = tsl,
  suffix = "_my_suffix"
)

tsl_colnames_get(tsl = tsl)
}
