% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_cols.R
\name{tsl_cols}
\alias{tsl_cols}
\title{Number of Columns of a Time Series List}
\usage{
tsl_cols(tsl = NULL)
}
\arguments{
\item{tsl}{(required, list of zoo objects) List of time series. Default: NULL}
}
\value{
A list if each time series has a different number of columns, and a integer otherwise.
}
\description{
Number of Columns of a Time Series List
}
