% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_initialize.R
\name{tsl_initialize}
\alias{tsl_initialize}
\alias{tsl_init}
\title{Prepare Time Series List}
\usage{
tsl_initialize(
  x = NULL,
  id_column = NULL,
  time_column = NULL,
  lock_step = FALSE,
  verbose = TRUE
)

tsl_init(
  x = NULL,
  id_column = NULL,
  time_column = NULL,
  lock_step = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{(required, list or data frame) A named list with ordered sequences, or a long data frame with a grouping column. Default: NULL.}

\item{id_column}{(optional, column name) Column name used for splitting a 'x' data frame into a list.}

\item{time_column}{(optional if \code{lock_step = FALSE}, and required otherwise, column name) Name of the column representing time, if any. Default: NULL.}

\item{lock_step}{(optional, logical) If TRUE, all input sequences are subset to their common times according to the values in \code{time_column}.}

\item{verbose}{(optional, logical) If FALSE, all messages are suppressed. Default: TRUE}
}
\value{
A named list of matrices.
}
\description{
Most functions in this package take a \strong{time series list} (or \strong{tsl} for short) as main input. A \code{tsl} is a list of zoo time-series objects (see \code{\link[zoo:zoo]{zoo::zoo()}}). There is not a formal class for \code{tsl} objects, but there are requirements these objects must follow to ensure the stability of the package functionalities (see \code{\link[=tsl_is_valid]{tsl_is_valid()}}). These requirements are:
\itemize{
\item There are no NA, Inf, -Inf, or NaN cases in the zoo objects (see \code{\link[=tsl_count_NA]{tsl_count_NA()}} and \code{\link[=tsl_handle_NA]{tsl_handle_NA()}}).
\item All zoo objects must have at least one common column name to allow time series comparison (see \code{\link[=tsl_colnames]{tsl_colnames()}}).
\item All zoo objects have a character attribute "name" identifying the object. This attribute is not part of the zoo class, but the package ensures that this attribute is not lost during data manipulations.
\item Each element of the time series list is named after the zoo object it contains (see \code{\link[=tsl_names]{tsl_names()}}, \code{\link[=tsl_names_set]{tsl_names_set()}} and \code{\link[=tsl_names_clean]{tsl_names_clean()}}).
\item The time series list contains two zoo objects or more.
}

The function \code{\link[=tsl_initialize]{tsl_initialize()}} (also \code{\link[=tsl_init]{tsl_init()}}) is designed to help the user turn their data into a valid time series list.

TODO: list valid input objects.
}
\examples{
data(mis)
x <- tsl_initialize(
  x = mis,
  id_column = "MIS"
)
}
