% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_names.R
\name{tsl_names}
\alias{tsl_names}
\title{Names of Time Series List}
\usage{
tsl_names(tsl = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}
}
\value{
character vector
}
\description{
Just a wrapper to \link[base:names]{base::names}.
}
\examples{
TODO: complete example
}
