% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_names.R
\name{tsl_names_get}
\alias{tsl_names_get}
\title{Names of Time Series List}
\usage{
tsl_names_get(tsl = NULL, colnames = FALSE)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{colnames}{(optional, logial) If TRUE, all column names are shown as well. Default: FALSE}
}
\value{
list
}
\description{
A time series list has three sets of names: the names of the list items (as returned by \code{names(tsl)}), the names of the contained zoo objects, as stored in their attribute "name", and the column names of teh zoo objects. This function extracts these names as a list mimicking the structure of the original time series list.
}
\examples{
#simulate a time series list
tsl <- tsl_initialize(
  x = fagus_dynamics,
  id_column = "site",
  time_column = "date"
)

#get list and zoo names
tsl_names_get(
  tsl = tsl
  )

#get column names
tsl_names_get(
  tsl = tsl,
  colnames = TRUE
)
TODO: think of a more flexible approach to get names? The current list output is a bit ugly
}
