% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_ncol.R
\name{tsl_ncol}
\alias{tsl_ncol}
\title{Number of Columns in a Time Series Lists}
\usage{
tsl_ncol(tsl = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}
}
\value{
list
}
\description{
Number of Columns in a Time Series Lists
}
\examples{
#initialize time series list
tsl <- tsl_simulate(
  n = 2,
  cols = 6
)

#number of columns per zoo object
tsl_ncol(tsl = tsl)
}
