% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_remove_exclusive_cols.R
\name{tsl_remove_exclusive_cols}
\alias{tsl_remove_exclusive_cols}
\title{Remove Exclusive Columns from Time Series List}
\usage{
tsl_remove_exclusive_cols(tsl = NULL, tsl_test = FALSE)
}
\arguments{
\item{tsl}{(required, list of zoo objects) List of time series. Default: NULL}

\item{tsl_test}{(optional, logical) If TRUE, a validity test on the argument `tsl` is performed by [tsl_is_valid()]. It might be useful to set it to TRUE if something goes wrong while executing this function. Default: FALSE}
}
\value{
time series list
}
\description{
Removes columns that are not shared across all zoo objects within a time series list.
}
