% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_resample.R
\name{tsl_resample}
\alias{tsl_resample}
\title{Resamples Time Series List to a New Time}
\usage{
tsl_resample(tsl = NULL, time = NULL, method = "loess")
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{time}{(required, zoo object, or vector of the classes numeric, Date, or POSIXct) New time of the resampled zoo object. Must be of the same class of the time in `x`. If one of them is "numeric" and the other is not, an error is returned. If a zoo object is provided, its time is used as a template to resample `x`. If NULL, irregular time series are predicted into a regular version of their own time, and regular time series are returned as is. Default: NULL}

\item{method}{(optional, string) Name of the method to resample the time series. One of "gam" or "loess". Default: "loess".}
}
\value{
time series list
}
\description{
Given time series list and a time vector, this function uses Generalized Additive Models (GAM) to resample the time series to a new time. This operation is useful to align time series with different frequencies, or to transform irregular time series into regular.

The new time vector must be of the same class as the time of the zoo object (either  numeric, Date, or POSIXct). This new time vector can be either created from scratch, or can be extracted from another zoo object with [zoo::index()].

Each time series is modelled as a function of the original time of the zoo object with [mgcv::gam()], and then predicted to the new time.

Please use this operation with care, as there are limits to the amount of resampling that can be done without distorting the data. The safest option is to keep the distance between new time points within the same magnitude of the distance between the old time points.

This function supports progress bars generated by the `progressr` package. See examples.

This function also accepts a parallelization setup via [future::plan()], but it might only be worth it for very long time series.
}
\examples{
#simulate time series list
x <- tsl_simulate(
  cols = 2,
  rows = 50
)

#plot zoo object
tsl_plot(x)

#check time
tsl_time(x)

#new time to first day of each month
new_time <- seq.Date(
  from = as.Date("2010-01-01"),
  to = as.Date("2020-01-01"),
  by = "1 month"
)

#resampling
x_new <- tsl_resample(
  tsl = x,
  time = new_time
)

#check new time
tsl_time(x_new)

#plot resampled zoo object
tsl_plot(x_new)
}
