% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_rows.R
\name{tsl_rows}
\alias{tsl_rows}
\title{Number of rows of a Time Series List}
\usage{
tsl_rows(tsl = NULL)
}
\arguments{
\item{tsl}{(required, list of zoo objects) List of time series. Default: NULL}
}
\value{
A list if each time series has a different number of rows, and a integer otherwise.
}
\description{
Number of rows of a Time Series List
}
