% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_select_shared_cols.R
\name{tsl_select_shared_cols}
\alias{tsl_select_shared_cols}
\title{Select Shared Columns from a Time Series List}
\usage{
tsl_select_shared_cols(tsl = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}
}
\value{
time series list
}
\description{
Returns a time series list with only the columns shared across all zoo time series.
}
\examples{
TODO: complete example
}
