% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_colnames.R
\name{tsl_set_colnames}
\alias{tsl_set_colnames}
\title{Renames Columns of Zoo Objects in a Time Series List}
\usage{
tsl_set_colnames(tsl = NULL, new_names = NULL, test_valid = TRUE)
}
\arguments{
\item{tsl}{(required, time series list) Individual time series or time series list created with [tsl_initialize]. Default: NULL}

\item{new_names}{(required, list) Named list. List names should match old column names in `tsl`, and each named item should contain a character string with the new name. For example, `colnames = list(old_name = "new_name")` changes the name of the column "old_name" to "new_name".}

\item{test_valid}{(optional, logical) If FALSE, the validity test performed by [tsl_is_valid()] is ignored. Useful to improve performance when the user knows that `tsl` is valid. Default: TRUE}
}
\description{
Renames Columns of Zoo Objects in a Time Series List
}
