% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_names.R
\name{tsl_set_names}
\alias{tsl_set_names}
\title{Set Names of a Time Series List}
\usage{
tsl_set_names(tsl = NULL, names = NULL)
}
\arguments{
\item{tsl}{(required, list of zoo objects) List of time series. Default: NULL}

\item{names}{(optional, character vector) names to set. Must be of the same length of `x`. If NULL, and the list `x` has names, then the names of the zoo objects inside of the list are taken from the names of the list elements.}

\item{clean}{(optional, logical) If TRUE, the names of `tsl` and the ones in `names` (if provided) are cleaned via [clean_names()].}

\item{abbreviate}{(optional, integer)}
}
\description{
Sets the names of a time series list created with [tsl_initialize()] and the internal names of the zoo objects inside, stored in their attribute "name".
}
