% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_smoothing.R
\name{tsl_smoothing}
\alias{tsl_smoothing}
\title{Rolling Window Smoothing of Time Series Lists}
\usage{
tsl_smoothing(tsl = NULL, window = 3, f = NULL, ...)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{window}{(optional, integer) Smoothing window width, in number of cases. Default: 3}

\item{f}{(optional, quoted or unquoted function name) Name of a standard or custom function to aggregate numeric vectors. Typical examples are \code{mean}, \code{max},\code{min}, \code{median}, and \code{quantile}. Default: \code{mean}.}

\item{...}{(optional, additional arguments) additional arguments to \code{f}.}
}
\value{
time series list
}
\description{
Rolling Window Smoothing of Time Series Lists
}
\examples{

future::plan(
  future::multisession,
  workers = 2 #set to parallelly::availableCores() - 1
)

# progress bar (does not work in examples)
# progressr::handlers(global = TRUE)

tsl <- tsl_simulate(n = 2)

tsl_smooth <- tsl_smoothing(
  tsl = tsl,
  window = 5,
  f = mean
)

if(interactive()){
  tsl_plot(tsl)
  tsl_plot(tsl_smooth)
}

future::plan(
  future::sequential
)
}
\seealso{
Other tsl_processing: 
\code{\link{tsl_aggregate}()},
\code{\link{tsl_resample}()},
\code{\link{tsl_stats}()},
\code{\link{tsl_transform}()}
}
\concept{tsl_processing}
