% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_stats.R
\name{tsl_stats}
\alias{tsl_stats}
\title{Summary Statistics of Time Series Lists}
\usage{
tsl_stats(tsl = NULL, ac_lags = 1)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{ac_lags}{(optional, integer) An integer specifying the number of autocorrelation lags to compute. Default: 1.}
}
\value{
data frame:
\itemize{
\item name: The name of the time series (derived from the list names).
\item rows: Number of rows in the time series.
\item columns: Number of columns in the time series.
\item time_units: Character string describing the time units.
\item time_begin: Earliest time point in the time series.
\item time_end: Latest time point in the time series.
\item time_length: Length of the time series, experssed in time_units.
\item time_resolution: Average distance between consecutive samples.
\item variable: Name of the variable (column) for which statistics are computed.
\item min: Minimum value of the variable.
\item q1: First quartile (25th percentile) of the variable.
\item median: Median value of the variable.
\item q3: Third quartile (75th percentile) of the variable.
\item max: Maximum value of the variable.
\item mean: Mean value of the variable.
\item sd: Standard deviation of the variable.
\item range: Range of the variable, computed as max - min.
\item iq_range: Interquartile range (IQR) of the variable, computed as q3 - q1.
\item skewness: Skewness of the variable, a measure of asymmetry in the distribution.
\item kurtosis: Kurtosis of the variable, a measure of the "tailedness" of the distribution.
\item ac_lag_1, ac_lag_2, ...: Autocorrelation values for the specified number of lags. The column names will correspond to the lag values (e.g., \code{ac_lag_1}, \code{ac_lag_2}).
}
}
\description{
This function computes a variety of summary statistics for each time series and numeric column within time series list. The statistics include common metrics  such as minimum, maximum, quartiles, mean, standard deviation, range, interquartile range, skewness, kurtosis, and autocorrelation for specified lags.

This function supports progress bars generated by the \code{progressr} package. See examples.

This function also accepts a parallelization setup via \code{\link[future:plan]{future::plan()}}, but it might only be worth it for large time series lists.
}
\examples{
TODO: add example

}
