% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_time_class.R
\name{tsl_time_class_set}
\alias{tsl_time_class_set}
\title{Coerce Elements of Time Series List to same Time Class}
\usage{
tsl_time_class_set(tsl = NULL, to = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{to}{(required, class name) class to coerce \code{x} to. Either "Date", "POSIXct", "integer" or "numeric". If NULL, coerces to the majority time class. Default: NULL}
}
\value{
time series list
}
\description{
Coerce Elements of Time Series List to same Time Class
}
\examples{
#TODO missing example
}
\seealso{
Other tsl_management: 
\code{\link{tsl_colnames_clean}()},
\code{\link{tsl_colnames_get}()},
\code{\link{tsl_colnames_set}()},
\code{\link{tsl_count_NA}()},
\code{\link{tsl_diagnose}()},
\code{\link{tsl_handle_NA}()},
\code{\link{tsl_names_clean}()},
\code{\link{tsl_names_get}()},
\code{\link{tsl_names_set}()},
\code{\link{tsl_names_test}()},
\code{\link{tsl_ncol}()},
\code{\link{tsl_nrow}()},
\code{\link{tsl_repair}()},
\code{\link{tsl_split}()},
\code{\link{tsl_subset}()},
\code{\link{tsl_time}()},
\code{\link{tsl_to_df}()}
}
\concept{tsl_management}
