% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_time_summary.R
\name{tsl_time_summary}
\alias{tsl_time_summary}
\title{Time Summary of a Time Series List}
\usage{
tsl_time_summary(tsl = NULL)
}
\arguments{
\item{tsl}{(required, list of zoo objects) Time series list. Default: NULL}
}
\value{
List with class, range, units, and resolution of the time in `tsl`.
}
\description{
Time Summary of a Time Series List
}
