% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_time_units.R
\name{tsl_time_units}
\alias{tsl_time_units}
\title{Time Units of Time Series List}
\usage{
tsl_time_units(tsl = NULL, tsl_test = FALSE)
}
\arguments{
\item{tsl}{(required, list of zoo objects) List of time series. Default: NULL}

\item{tsl_test}{(optional, logical) If TRUE, a validity test on the argument `tsl` is performed by [tsl_is_valid()]. It might be useful to set it to TRUE if something goes wrong while executing this function. Default: FALSE}
}
\value{
List of character vector or character vector with the time units.
}
\description{
Returns keywords representing the time units available in the 'index' attribute of the zoo time series within a time series list.

If the 'index' attribute is of the class "Date", then this function searches for the highest and lowest units of variability in the range of dates, and returns them along with all the units in between. All possible unit keywords are: "millennium", "century", "decade", "year", "quarter", "month", "week", "day", "hour", "minute", "second". For example, if the argument 'tsl' has a range of dates c("2020-02-01" "2023-12-19"), since the difference between years is higher than 1 and the difference between days is higher than one, then the returned units are c("year", "quarter", "month", "week", and "day").

If the 'index' attribute is of the class 'numeric', then the then this function searches for the highest and lowest units of variability in the range of dates, and returns them along with all the units in between. All possible unit keywords are: "1e6" (million), "1e5" (one hundred thousand), "1e3" (ten thousand), "1e3" (thousand), "1e2" (hundred), "1e1" (ten). For example, for the range of numeric times c(-100000, 2000), the units returned are c("1e5", "1e4", "1e3", "1e2", "1e1").
}
