% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_to_df.R
\name{tsl_to_df}
\alias{tsl_to_df}
\title{Time Series List to Data Frame}
\usage{
tsl_to_df(tsl = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}
}
\value{
data frame
}
\description{
Time Series List to Data Frame
}
\examples{

tsl <- tsl_simulate(
  n = 3,
  rows = 10,
  time_range = c(
    "2010-01-01",
    "2020-01-01"
  )
)

df <- tsl_to_df(
  tsl = tsl
)

names(df)
nrow(df)
head(df)
}
\seealso{
Other data_processing: 
\code{\link{tsl_aggregate}()},
\code{\link{tsl_resample}()},
\code{\link{tsl_split}()},
\code{\link{tsl_subset}()},
\code{\link{tsl_transform}()},
\code{\link{zoo_aggregate}()},
\code{\link{zoo_permute}()},
\code{\link{zoo_resample}()}
}
\concept{data_processing}
