% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_validate.R
\name{tsl_validate}
\alias{tsl_validate}
\alias{tsl_is_valid}
\title{Validate Time Series List}
\usage{
tsl_validate(tsl = NULL)

tsl_is_valid(tsl = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}
}
\value{
time series list
}
\description{
Validity assessment of Time Series Lists. The required features for a valid time series list are:
\itemize{
\item Argument \code{tsl} is a list.
\item List \code{tsl} has unique names.
\item List \code{tsl} has more than one element.
\item All elements in \code{tsl} are \code{zoo} objects.
\item All \code{zoo} objects have the attribute "name".
\item All \code{zoo} objects have unique values in the attribute "name".
\item All \code{zoo} objects have at least one common column name.
\item All columns in the \code{zoo} objects are numeric.
\item All \code{zoo} objects have zero NA cases.
}

If a time series list is valid, it is returned without any messages or warnings, and the value of the attribute "valid" is set to \code{TRUE}.
}
\examples{
TODO: complete example
}
