% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_args.R
\name{utils_check_args_tsl}
\alias{utils_check_args_tsl}
\title{Structural Check for Time Series Lists}
\usage{
utils_check_args_tsl(tsl = NULL, min_length = 2)
}
\arguments{
\item{tsl}{(required, list) list of zoo objects. Default: NULL}

\item{min_length}{(required, positive integer) minimum number of zoo objects in \code{tsl}. Default: 2}
}
\value{
error messages (if any)
}
\description{
Internal function to check that a time series list is a list of zoo objects and has a minimum number of objects. For a more comprehensive test, use \code{\link[=tsl_diagnose]{tsl_diagnose()}}.
}
