% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_arguments.R
\name{utils_check_distance_args}
\alias{utils_check_distance_args}
\title{Check Distance Argument}
\usage{
utils_check_distance_args(distance = NULL)
}
\arguments{
\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}
}
\value{
Character vector with distance names
}
\description{
Check Distance Argument
}
\examples{

utils_check_distance_args(
  distance = c(
    "euclidean",
    "euc"
   )
  )

}
