% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_arguments.R
\name{utils_check_distantia_df}
\alias{utils_check_distantia_df}
\title{Checks Distantia Data Frame}
\usage{
utils_check_distantia_df(df = NULL)
}
\arguments{
\item{df}{(required, data frame) output of [distantia()]. Default: NULL}
}
\value{
distantia data frame
}
\description{
Checks Distantia Data Frame
}
