% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_list_class.R
\name{utils_check_list_class}
\alias{utils_check_list_class}
\title{Checks Classes of List Elements Against Expectation}
\usage{
utils_check_list_class(x = NULL, expected_class = "data.frame")
}
\arguments{
\item{x}{(required, list) Default: NULL}

\item{expected_class}{(required, class name). One of "data.frame", "matrix", or "vector". Default: "data.frame".}
}
\value{
side effects
}
\description{
Checks Classes of List Elements Against Expectation
}
\seealso{
Other internal: 
\code{\link{utils_check_distance_args}()},
\code{\link{utils_check_distantia_args}()},
\code{\link{utils_check_matrix_args}()},
\code{\link{utils_check_path_args}()},
\code{\link{utils_check_zoo_args}()},
\code{\link{utils_digits}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()}
}
\concept{internal}
