% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_arguments.R
\name{utils_check_matrix_args}
\alias{utils_check_matrix_args}
\title{Checks Input Matrix}
\usage{
utils_check_matrix_args(m = NULL, arg_name = "m")
}
\arguments{
\item{m}{(required, matrix) distance or cost matrix resulting from \code{\link[=psi_dist_matrix]{psi_dist_matrix()}} or \code{\link[=psi_cost_matrix]{psi_cost_matrix()}}. Default: NULL}

\item{arg_name}{(optional, character string) name of the argument being checked. Default: NULL}
}
\value{
matrix
}
\description{
Checks Input Matrix
}
