% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_arguments.R
\name{utils_check_zoo_args}
\alias{utils_check_zoo_args}
\title{Checks Argument x}
\usage{
utils_check_zoo_args(x = NULL, arg_name = "x")
}
\arguments{
\item{x}{(required, list of matrices) list of input matrices generated with \code{\link[=tsl_initialize]{tsl_initialize()}}. Default: NULL}

\item{arg_name}{(optional, character string) name of the argument being checked. Default: NULL}
}
\value{
zoo object
}
\description{
Checks Argument x
}
\seealso{
Other internal: 
\code{\link{utils_check_distance_args}()},
\code{\link{utils_check_distantia_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_check_matrix_args}()},
\code{\link{utils_check_path_args}()},
\code{\link{utils_digits}()},
\code{\link{utils_distantia_df_split}()},
\code{\link{utils_distantia_df_to_matrix}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()},
\code{\link{utils_tsl_pairs}()}
}
\concept{internal}
