% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color_continuous_default.R
\name{utils_color_continuous_default}
\alias{utils_color_continuous_default}
\title{Default Continuous Color Palette}
\usage{
utils_color_continuous_default(n = NULL, palette = "Zissou 1")
}
\arguments{
\item{n}{(required, integer) number of colors to generate. Default = NULL}

\item{palette}{(required, character string) Argument \code{palette} of \code{\link[grDevices:palettes]{grDevices::hcl.colors()}}. Default: "Zissou 1"}
}
\value{
Color palette
}
\description{
Uses the function \code{\link[grDevices:palettes]{grDevices::hcl.colors()}} to generate a continuous color palette.
}
\examples{

utils_color_continuous_default(n = 20)

}
