% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color_discrete_default.R
\name{utils_color_discrete_default}
\alias{utils_color_discrete_default}
\title{Default Discrete Color Palettes}
\usage{
utils_color_discrete_default(n = NULL)
}
\arguments{
\item{n}{(required, integer) number of colors (up to 36) to generate. Default: NULL}
}
\value{
Color palette
}
\description{
Uses the function \code{\link[grDevices:palette]{grDevices::palette.colors()}} to generate discrete color palettes using the following rules:
\itemize{
\item \code{n <= 9}: "Okabe-Ito".
\item \code{n == 10}: "Tableau 10"
\item \code{n > 10 && n <= 12}: "Paired"
\item \code{n > 12 && n <= 26}: "Alphabet"
\item \code{n > 26 && n <= 36}: "Polychrome 36"
}
}
\examples{

utils_color_discrete_default(n = 9)
}
\seealso{
Other internal: 
\code{\link{distantia-package}},
\code{\link{utils_block_size}()},
\code{\link{utils_check_distantia_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_clean_names}()},
\code{\link{utils_cluster_hclust_optimizer}()},
\code{\link{utils_cluster_kmeans_optimizer}()},
\code{\link{utils_cluster_silhouette}()},
\code{\link{utils_coerce_time_class}()},
\code{\link{utils_color_breaks}()},
\code{\link{utils_color_continuous_default}()},
\code{\link{utils_df_to_matrix}()},
\code{\link{utils_digits}()},
\code{\link{utils_distantia_df_split}()},
\code{\link{utils_global_scaling_params}()},
\code{\link{utils_importance_df_to_wide}()},
\code{\link{utils_is_time}()},
\code{\link{utils_line_color}()},
\code{\link{utils_line_guide}()},
\code{\link{utils_matrix_guide}()},
\code{\link{utils_matrix_plot}()},
\code{\link{utils_new_time}()},
\code{\link{utils_optimize_loess}()},
\code{\link{utils_optimize_spline}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()},
\code{\link{utils_rescale_vector}()},
\code{\link{utils_time_keywords}()},
\code{\link{utils_time_keywords_dictionary}()},
\code{\link{utils_time_keywords_translate}()},
\code{\link{utils_time_units}()},
\code{\link{utils_tsl_pairs}()}
}
\concept{internal}
