% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_digits.R
\name{utils_digits}
\alias{utils_digits}
\title{Number of Decimal Places}
\usage{
utils_digits(x = NULL)
}
\arguments{
\item{x}{(required, numeric) Default: NULL}
}
\value{
Integer, number of decimal places
}
\description{
Number of Decimal Places
}
\examples{
utils_digits(x = 0.234)
}
\keyword{internal}
