% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_distantia_df_to_matrix.R
\name{utils_distantia_df_to_matrix}
\alias{utils_distantia_df_to_matrix}
\title{Data Frame to Matrix}
\usage{
utils_distantia_df_to_matrix(df = NULL, x = "x", y = "y", value = "psi")
}
\arguments{
\item{df}{(required, data frame) data frame typically resulting from \code{\link[=distantia]{distantia()}}, but others are accepted as long as the columns \code{x}, \code{y}, and \code{value} are available. Default: NULL}

\item{x}{(required, character string) Name of the column with the names of the matrix columns. Default: "x"}

\item{y}{(required, character string) Name of the column with the names of the matrix rows. Default: "y"}

\item{value}{(required, character string) Name of the column with values.}
}
\value{
dist matrix
}
\description{
Internal function to convert a distantia data frame into a matrix.
}
\examples{
tsl <- tsl_simulate(
  n = 3,
  time_range = c(
    "2010-01-01 12:00:25",
    "2024-12-31 11:15:45"
  )
)

df <- distantia(
  tsl = tsl
)

m <- utils_distantia_df_to_matrix(
  df = df
)

m
}
\seealso{
Other internal: 
\code{\link{utils_boxplot_common}()},
\code{\link{utils_check_args_distantia}()},
\code{\link{utils_check_args_matrix}()},
\code{\link{utils_check_args_path}()},
\code{\link{utils_check_args_tsl}()},
\code{\link{utils_check_args_zoo}()},
\code{\link{utils_check_distance_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_clean_names}()},
\code{\link{utils_digits}()},
\code{\link{utils_distantia_df_split}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()},
\code{\link{utils_tsl_pairs}()}
}
\concept{internal}
