% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_drop_geometry.R
\name{utils_drop_geometry}
\alias{utils_drop_geometry}
\title{Removes Geometry Column from SF Data Frames}
\usage{
utils_drop_geometry(df = NULL)
}
\arguments{
\item{df}{(required, data frame) Input data frame. Default: NULL.}
}
\value{
data frame
}
\description{
Replicates the functionality of  \code{sf::st_drop_geometry()} without depending on the \code{sf} package.
}
\seealso{
Other tsl_processing_internal: 
\code{\link{utils_global_scaling_params}()},
\code{\link{utils_optimize_loess}()},
\code{\link{utils_optimize_spline}()},
\code{\link{utils_rescale_vector}()}
}
\concept{tsl_processing_internal}
