% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_importance_df_to_wide.R
\name{utils_importance_df_to_wide}
\alias{utils_importance_df_to_wide}
\title{Importance Data Frame to Wide Format}
\usage{
utils_importance_df_to_wide(df = NULL, sep = "__")
}
\arguments{
\item{df}{(required, data frame) Output of \code{\link[=distantia_importance]{distantia_importance()}}. Default: NULL}

\item{sep}{(required, character string) Separator between the name of the importance metric and the time series variable. Default: "__".}
}
\value{
data frame
}
\description{
Takes as input a data frame returned by \code{\link[=distantia_importance]{distantia_importance()}} to return a data frame with one pair of time series per row, and the following columns:
\itemize{
\item \code{most_similar}: name of the variable with the highest contribution to similarity (most negative value in the \code{importance} column) for each pair of time series.
\item \code{most_dissimilar}: name of the variable with the highest contribution to dissimilarity (most positive value in the \code{importance} column) for each pair of time series.
\item \code{importance__variable_name}: contribution to similarity (negative values) or dissimilarity (positive values) of the given variable.
\item \code{psi_only_with__variable_name}: dissimilarity of the two time series when only using the given variable.
\item \code{psi_without__variable_name}: dissimilarity of the two time series when removing the given variable.
}
}
\examples{
#prepare time series
data("fagus_dynamics")

tsl <- tsl_initialize(
  x = fagus_dynamics,
  id_column = "site",
  time_column = "date"
) |>
  tsl_transform(
    f = f_scale
  )

#importance data frame
df <- distantia_importance(
  tsl = tsl
)
df

#to wide format
df_wide <- utils_importance_df_to_wide(
  df = df
)
df_wide

}
\seealso{
Other dissimilarity_analysis: 
\code{\link{distance}()},
\code{\link{distantia}()},
\code{\link{distantia_aggregate}()},
\code{\link{distantia_boxplot}()},
\code{\link{distantia_cluster_hclust}()},
\code{\link{distantia_cluster_kmeans}()},
\code{\link{distantia_importance}()},
\code{\link{distantia_matrix}()},
\code{\link{distantia_plot}()},
\code{\link{distantia_to_sf}()},
\code{\link{utils_block_size}()}

Other internal: 
\code{\link{distantia-package}},
\code{\link{utils_block_size}()},
\code{\link{utils_check_distantia_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_clean_names}()},
\code{\link{utils_cluster_hclust_optimizer}()},
\code{\link{utils_cluster_kmeans_optimizer}()},
\code{\link{utils_cluster_silhouette}()},
\code{\link{utils_coerce_time_class}()},
\code{\link{utils_color_breaks}()},
\code{\link{utils_color_continuous_default}()},
\code{\link{utils_color_discrete_default}()},
\code{\link{utils_df_to_matrix}()},
\code{\link{utils_digits}()},
\code{\link{utils_distantia_df_split}()},
\code{\link{utils_global_scaling_params}()},
\code{\link{utils_is_time}()},
\code{\link{utils_line_color}()},
\code{\link{utils_line_guide}()},
\code{\link{utils_matrix_guide}()},
\code{\link{utils_matrix_plot}()},
\code{\link{utils_new_time}()},
\code{\link{utils_optimize_loess}()},
\code{\link{utils_optimize_spline}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()},
\code{\link{utils_rescale_vector}()},
\code{\link{utils_time_keywords}()},
\code{\link{utils_time_keywords_dictionary}()},
\code{\link{utils_time_keywords_translate}()},
\code{\link{utils_time_units}()},
\code{\link{utils_tsl_pairs}()}
}
\concept{dissimilarity_analysis}
\concept{internal}
