% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_is_time.R
\name{utils_is_time}
\alias{utils_is_time}
\title{Title}
\usage{
utils_is_time(x = NULL)
}
\arguments{
\item{x}{(required, vector) Vector to test. If the class of the vector elements is 'numeric', 'POSIXct', or 'Date', the function returns TRUE. Default: NULL.}
}
\value{
logical
}
\description{
Title
}
\examples{

utils_is_time(
  x = c("2024-01-01", "2024-02-01")
)

utils_is_time(
  x = utils_as_time(
    x = c("2024-01-01", "2024-02-01")
    )
)
}
\seealso{
Other internal_time_handling: 
\code{\link{utils_as_time}()},
\code{\link{utils_coerce_time_class}()},
\code{\link{utils_new_time}()},
\code{\link{utils_time_keywords}()},
\code{\link{utils_time_keywords_dictionary}()},
\code{\link{utils_time_keywords_translate}()},
\code{\link{utils_time_units}()}
}
\concept{internal_time_handling}
