% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prepare_matrix.R
\name{utils_prepare_matrix}
\alias{utils_prepare_matrix}
\title{Matrix to Data Frame}
\usage{
utils_prepare_matrix(x = NULL)
}
\arguments{
\item{x}{(required, matrix) Default: NULL}
}
\value{
A data frame
}
\description{
Matrix to Data Frame
}
\seealso{
Other internal_data_preparation: 
\code{\link{utils_clean_names}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()}
}
\concept{internal_data_preparation}
