% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prepare_matrix_list.R
\name{utils_prepare_matrix_list}
\alias{utils_prepare_matrix_list}
\title{Convert List of Matrices to List of Data Frames}
\usage{
utils_prepare_matrix_list(x = NULL)
}
\arguments{
\item{x}{(required, list of matrices) Default: NULL}
}
\value{
List of Data Frames
}
\description{
Convert List of Matrices to List of Data Frames
}
\seealso{
Other internal: 
\code{\link{utils_check_args_distantia}()},
\code{\link{utils_check_args_matrix}()},
\code{\link{utils_check_args_path}()},
\code{\link{utils_check_args_tsl}()},
\code{\link{utils_check_args_zoo}()},
\code{\link{utils_check_distance_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_clean_names}()},
\code{\link{utils_digits}()},
\code{\link{utils_distantia_df_split}()},
\code{\link{utils_distantia_df_to_matrix}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()},
\code{\link{utils_tsl_pairs}()}
}
\concept{internal}
