% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rescale_vector.R
\name{utils_rescale_vector}
\alias{utils_rescale_vector}
\title{Rescales Numeric Vector to a New Range}
\usage{
utils_rescale_vector(
  x = NULL,
  new_min = 0,
  new_max = 1,
  old_min = NULL,
  old_max = NULL
)
}
\arguments{
\item{x}{(required, numeric vector) Numeric vector. Default: \code{NULL}}

\item{new_min}{(optional, numeric) New minimum value. Default: \code{0}}

\item{new_max}{(optional_numeric) New maximum value. Default: \code{1}}

\item{old_min}{(optional, numeric) Old minimum value. Default: \code{NULL}}

\item{old_max}{(optional_numeric) Old maximum value. Default: \code{NULL}}
}
\value{
Numeric vector
}
\description{
Rescales Numeric Vector to a New Range
}
\examples{

 out <- utils_rescale_vector(
   x = stats::rnorm(100),
   new_min = 0,
   new_max = 100,
   integer = TRUE
   )
   out

}
