% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_simplify_list.R
\name{utils_simplify_list}
\alias{utils_simplify_list}
\title{Simmplify List With Repeated Elements}
\usage{
utils_simplify_list(x)
}
\arguments{
\item{x}{(required, list)}
}
\value{
simplified list
}
\description{
Simmplify List With Repeated Elements
}
