% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_breaks.R
\name{utils_time_breaks}
\alias{utils_time_breaks}
\title{Time Breaks for Time Series Aggregation}
\usage{
utils_time_breaks(tsl = NULL, breaks = NULL)
}
\arguments{
\item{tsl}{(required, list of zoo objects) List of time series. Default: NULL}

\item{breaks}{(required, numeric, numeric vector, or keyword) definition of the aggregation groups. There are several options:
\itemize{
  \item keyword: Only when time in tsl is either a date "YYYY-MM-DD" or a datetime "YYYY-MM-DD hh-mm-ss". Valid options are "year", "quarter", "month", and "week" for date, and, "day", "hour", "minute", and "second" for datetime.
}}
}
\value{
Vector of class numeric, Date, or POSIXct
}
\description{
Time Breaks for Time Series Aggregation
}
