% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_class.R
\name{utils_time_class}
\alias{utils_time_class}
\title{Ensures Correct Class for Time Arguments}
\usage{
utils_time_class(x = NULL, quiet = TRUE)
}
\arguments{
\item{x}{(required, vector) Vectors of the classes 'numeric', 'Date', and 'POSIXct' are valid and returned without any transformation. Character vectors are analyzed to determine their more probable type, and are coerced to 'Date' or 'POSIXct' depending on their number of elements. Generally, any character vector representing an ISO 8601 standard, like "YYYY-MM-DD" or "YYYY-MM-DD HH:MM:SS" will be converted to a valid class. Additionally, character vectors representing a year (i.e. `c("2010", "2011")`) or a year and a month (i.e. `c("2010-02", "2011-03")`) will be added month and day ("-01" in either case) and coerced to 'Date'.}

\item{quiet}{(optional, logical) If TRUE, all messages are suppressed. Default is TRUE}
}
\value{
Input vector if its class is 'numeric', 'Date', or 'POSIXct', or input vector coerced to one of these classes otherwise.
}
\description{
This function guesses the class of a vector based on its elements. It can handle numeric vectors, character vectors that can be coerced to either "Date" or "POSIXct" classes, and vectors already in "Date" or "POSIXct" classes.
}
\examples{
# numeric
utils_time_class(
  x = c(-123120, 1200)
  )

# character string to Date
utils_time_class(
  x = c("2022-03-17", "2024-02-05")
  )

# incomplete character strings to Date
utils_time_class(
  x = c("2022", "2024")
  )

utils_time_class(
  x = c("2022-02", "2024-03")
  )

# character string to POSIXct
utils_time_class(
  x = c("2022-03-17 12:30:45", "2024-02-05 11:15:45")
  )

# Date vector (returns the input)
utils_time_class(
  x = as.Date(c("2022-03-17", "2024-02-05"))
  )

# POSIXct vector (returns the input)
utils_time_class(
  x = as.POSIXct(c("2022-03-17 12:30:45", "2024-02-05 11:15:45"))
  )

}
