% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_units.R
\name{utils_time_units}
\alias{utils_time_units}
\title{Supported Time Units}
\usage{
utils_time_units()
}
\value{
data frame
}
\description{
Returns a data frame with the names of the supported time units, the classes that can handle each time unit, and a the R expression used to identify what time units can be used when aggregating a time series.
}
