% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_units_df.R
\name{utils_time_units_df}
\alias{utils_time_units_df}
\title{Data Frame with Supported Time Units}
\usage{
utils_time_units_df(expression = FALSE)
}
\arguments{
\item{expression}{(optional, logical) if TRUE, the column "expression" is returned. Default: FALSE}
}
\value{
data frame
}
\description{
Returns a data frame with the names of the supported time units, the classes that can handle each time unit, and a the R expression used to identify what time units can be used when aggregating a time series.
}
\examples{

df <- utils_time_units()
head(df)

}
