% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_zoo_to_tsl.R
\name{utils_zoo_to_tsl}
\alias{utils_zoo_to_tsl}
\title{Converts Single Zoo Object to a Time Series List}
\usage{
utils_zoo_to_tsl(x = NULL)
}
\arguments{
\item{x}{(required, zoo object) Time series. Default: NULL}
}
\value{
time series list of length one.
}
\description{
Internal function to wrap a zoo object into a time series list.
}
\examples{
#create zoo object
x <- zoo_simulate()
class(x)

#to time series list
tsl <- utils_zoo_to_tsl(
  x = x
)

class(tsl)
class(tsl[[1]])
names(tsl)
attributes(tsl[[1]])$name
}
