% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowPsiPairedSamples.R
\name{workflowPsiPairedSamples}
\alias{workflowPsiPairedSamples}
\title{Computes the dissimilarity measure \emph{psi} on two or more sequences with paired samples.}
\usage{
workflowPsiPairedSamples(
  sequences = NULL,
  grouping.column = NULL,
  time.column = NULL,
  exclude.columns = NULL,
  method = "manhattan",
  diagonal = FALSE,
  format = "dataframe"
  )
}
\arguments{
\item{sequences}{dataframe with multiple sequences identified by a grouping column generated by \code{\link{prepareSequences}}.}

\item{grouping.column}{character string, name of the column in \code{sequences} to be used to identify separates sequences within the file.}

\item{time.column}{character string, name of the column with time/depth/rank data.}

\item{exclude.columns}{character string or character vector with column names in \code{sequences} to be excluded from the analysis.}

\item{method}{character string naming a distance metric. Valid entries are: "manhattan", "euclidean", "chi", and "hellinger". Invalid entries will throw an error.}

\item{output}{string, type of output. One of: "list" (default), "data.frame", "matrix"}
}
\value{
A list, matrix, or dataframe, with sequence names and psi values.
}
\description{
This function executes the following steps:
\itemize{
\item Computes the autosum of the sequences with \code{\link{autoSum}}.
\item Computes the distance between paired samples and sums it with \code{\link{distancePairedSamples}}.
\item Computes the dissimilarity measure \emph{psi} with the function \code{\link{psi}}.
\item Delivers an output of type "list" (default), "data.frame" or "matrix", depending on the user input, through either \code{\link{psiToDataframe}} or \code{\link{psiToMatrix}}.
}
}
\examples{

\dontrun{
}

}
\author{
Blas Benito <blasbenito@gmail.com>
}
