% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_aggregate.R
\name{zoo_aggregate}
\alias{zoo_aggregate}
\title{Aggregates Zoo Time Series}
\usage{
zoo_aggregate(x = NULL, breaks = NULL, f = mean, ...)
}
\arguments{
\item{x}{(required, zoo object) Time series to aggregate. Default: NULL}

\item{breaks}{(required, vector) Vector to define time breaks for aggregation. Must be in the same time units as `zoo::index(x)`.}

\item{stat}{(required, function) name without quotes and parenthesis of a standard function to smooth a time series. Typical examples are `mean` (default), `max`,`min`, `median`, and `sd`. Default: `mean`.}
}
\value{
zoo object
}
\description{
Aggregates Zoo Time Series
}
