% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_permute.R
\name{zoo_permute}
\alias{zoo_permute}
\title{Permutation of Zoo Time Series}
\usage{
zoo_permute(
  x = NULL,
  repetitions = 1L,
  permutation = "restricted_by_row",
  block_size = NULL,
  seed = 1
)
}
\arguments{
\item{x}{(required, zoo object) zoo time series. Default: NULL}

\item{repetitions}{(optional, integer vector) number of permutations to compute the p-value. If 0, p-values are not computed. Otherwise, the minimum is 2. Default: 1}

\item{permutation}{(optional, character string) permutation method. Valid values are listed below from higher to lower induced randomness:
\itemize{
  \item "free": unrestricted shuffling of rows and columns. Ignores block_size.
  \item "free_by_row": unrestricted shuffling of complete rows. Ignores block size.
  \item "restricted": restricted shuffling of rows and columns within blocks.
  \item "restricted_by_row": restricted shuffling of rows within blocks.
}}

\item{block_size}{(optional, integer) Row block size for the restricted permutation test. Only relevant when permutation methods are "restricted" or "restricted_by_row" and `repetitions` is higher than zero. A block of size `n` indicates that a row can only be permuted within a block of `n` adjacent rows. If NULL, defaults to the rounded one tenth of the shortest sequence in `tsl`. Default: NULL.}

\item{seed}{(optional, integer) initial random seed to use for replicability when computing p-values. Default: 1}
}
\value{
Time Series List
}
\description{
Permutation of Zoo Time Series
}
\examples{
x <- zoo_simulate()

y <- zoo_permute(
  x = x,
  repetitions = 3
)

tsl_plot(
  x = y,
  guide = FALSE
  )
}
