% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_time.R
\name{zoo_time}
\alias{zoo_time}
\title{Time Features of a Zoo Object}
\usage{
zoo_time(x = NULL)
}
\arguments{
\item{x}{(required, zoo object) Default: NULL.}
}
\value{
List with time features of the zoo object.
}
\description{
Time Features of a Zoo Object
}
