% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_time_units.R
\name{zoo_time_units}
\alias{zoo_time_units}
\title{Time Units of a Zoo Time Series}
\usage{
zoo_time_units(x)
}
\arguments{
\item{x}{(required, zoo object) time series with an index of the classes "Date". "POSIXct". or "numeric".}
}
\value{
Character vector of unit names.
}
\description{
Time Units of a Zoo Time Series
}
