% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_vector_to_matrix.R
\name{zoo_vector_to_matrix}
\alias{zoo_vector_to_matrix}
\title{Coerce Coredata of Univariate Zoo Time Series to Matrix}
\usage{
zoo_vector_to_matrix(x = NULL, name = "a")
}
\arguments{
\item{x}{(required, zoo object) zoo time series. Default: NULL}

\item{name}{(required, character string) name of the matrix column. Default: "a"}
}
\value{
zoo time series
}
\description{
Transforms vector coredata of univariate zoo time series to class matrix. If the input zoo time series has the attribute "name", the output inherits the value of such attribute.

Multivariate zoo objects are returned without changes.
}
\examples{
#create zoo object from vector
x <- zoo::zoo(
  x = runif(100)
)

#coredata is not a matrix
is.matrix(zoo::coredata(x))

#convert to matrix
y <- zoo_vector_to_matrix(
  x = x
)

#coredata is now a matrix
is.matrix(zoo::coredata(y))
}
\seealso{
Other data_preparation: 
\code{\link{tsl_colnames_clean}()},
\code{\link{tsl_colnames_set}()},
\code{\link{tsl_diagnose}()},
\code{\link{tsl_handle_NA}()},
\code{\link{tsl_initialize}()},
\code{\link{tsl_names_clean}()},
\code{\link{tsl_names_set}()},
\code{\link{tsl_repair}()},
\code{\link{tsl_time_class_set}()},
\code{\link{zoo_name_get}()},
\code{\link{zoo_name_set}()},
\code{\link{zoo_to_tsl}()}
}
\concept{data_preparation}
