% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{auto_distance_cpp}
\alias{auto_distance_cpp}
\title{Distance Matrix}
\usage{
auto_distance_cpp(m, f)
}
\arguments{
\item{f}{(required, function) a distance function. One of:
\itemize{
  \item distance_manhattan_cpp.
  \item distance_euclidean_cpp.
  \item distance_hellinger_cpp.
  \item distance_chi_cpp.
}}

\item{a}{(required, numeric matrix).}

\item{b}{(required, numeric matrix) of same number of columns as 'a'.}
}
\value{
Matrix of distances between 'a' (rows) and 'b' (columns).
}
\description{
Computes the distance matrix between the rows of two matrices
\code{a} and \code{b} with the same number of columns and arbitrary numbers of rows.
NA values should be removed before using this function.
If the selected distance function is [distance_chi_cpp], pairs of zeros should
be either removed or replaced with pseudo-zeros (i.e. 0.00001).
}
