% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_sum.R
\name{auto_sum}
\alias{auto_sum}
\title{Auto Sum}
\usage{
auto_sum(a = NULL, b = NULL, path = NULL, distance = "euclidean")
}
\arguments{
\item{a}{(required, data frame) a time series. Default: NULL.}

\item{b}{(required, data frame) a time series. Default: NULL.}

\item{path}{(required, data frame) dataframe produced by [cost_path()]. Only required when [cost_path_trim()] has been applied to `path`. Default: NULL.}

\item{distance}{(optional, character string) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}
}
\value{
Named vector with the auto sums of `a` and `b`.
}
\description{
Computes the sum of distances between consecutive samples in two multivariate time-series under comparison. Required to compute the measure of dissimilarity \code{psi} (Birks and Gordon 1985).
}
\examples{
data(sequenceA, sequenceB)

a <- sequenceA |>
  na.omit() |>
  as.matrix()

b <- sequenceB |>
  na.omit() |>
  as.matrix()

ab_sum <- auto_sum(
  a = a,
  b = b,
  distance = "manhattan"
)

ab_sum

}
