% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{auto_sum_cpp}
\alias{auto_sum_cpp}
\title{Auto-sum of Two Sequences}
\usage{
auto_sum_cpp(a, b, path, method)
}
\arguments{
\item{a}{(required, numeric matrix).}

\item{b}{(required, numeric matrix) of same number of columns as 'a'.}

\item{path}{(required, data frame) dataframe produced by [cost_path()]. Default: NULL}

\item{method}{(optional, character string) name of the distance metric. Valid entries are:
\itemize{
\item "euclidean" and "euc" (Default).
\item "manhattan" and "man".
\item "chi.
\item "hellinger" and "hel".
\item "chebyshev" and "che".
\item "canberra" and "can".
\item "cosine" and "cos".
\item "russelrao" and "rus".
\item "jaccard" and "jac".
}}
}
\value{
Numeric.
}
\description{
Sum of the the cumulative auto-sum of two sequences. This is
a key component of the psi computation.
}
