% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{auto_sum_no_path_cpp}
\alias{auto_sum_no_path_cpp}
\title{Auto Sum of Two Time Series}
\usage{
auto_sum_no_path_cpp(x, y, distance = "euclidean")
}
\arguments{
\item{x}{(required, numeric matrix) of same number of columns as 'y'.}

\item{y}{(required, numeric matrix).}

\item{distance}{(optional, character string) name or abbreviation of the
distance method. Valid values are in the columns "names" and "abbreviation"
of the dataset `distances`. Default: "euclidean".}
}
\value{
Auto sum of distances.
}
\description{
Sum of the the cumulative auto-sum of two sequences with paired samples. This is
a key component of the psi computation.
}
