% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosum.R
\name{autosum}
\alias{autosum}
\title{Computes sum of distances between samples in a multivariate time-series.}
\usage{
autosum <- function(
  sequences = NULL,
  grouping.column = NULL,
  time.column = NULL,
  exclude.columns = NULL,
  method = "manhattan"
  )
}
\arguments{
\item{sequences}{dataframe with one or several multivariate time-series identified by a grouping column.}

\item{grouping.column}{character string, name of the column in \code{sequences} to be used to identify separates sequences within the file. This argument is ignored if \code{sequence.A} and \code{sequence.B} are provided.}

\item{exclude.columns}{character string or character vector with column names in \code{sequences}, or \code{squence.A} and \code{sequence.B} to be excluded from the analysis.}

\item{method}{character string naming a distance metric. Valid entries are: "manhattan", "euclidean", "chi", and "hellinger". Invalid entries will throw an error.}

\item{time.colum}{character string, name of the column with time/depth/rank data. The data in this column is not modified.}
}
\value{
A list with slots named according \code{grouping.column} if there are several sequences in \code{sequences} or a number if there is only one sequence.
}
\description{
Computes the autosum of distances between consecutive samples in a multivariate time-series. Required to compute the measure of dissimilarity \code{psi}. Distances can be computed through the methods "manhattan", "euclidean", "chi", and "hellinger", and are implemented in the function \code{\link{distance}}.
}
\details{
Distances are computed as:
\itemize{
\item \code{manhattan}: \code{d <- sum(abs(x - y))}
\item \code{euclidean}: \code{d <- sqrt(sum((x - y)^2))}
\item \code{chi}: \code{
    xy <- x + y
    y. <- y / sum(y)
    x. <- x / sum(x)
    d <- sqrt(sum(((x. - y.)^2) / (xy / sum(xy))))}
\item \code{hellinger}: \code{d <- sqrt(1/2 * sum(sqrt(x) - sqrt(y))^2)}
}
Note that zeroes are replaced by 0.00001 whem \code{method} equals "chi" or "hellinger".
}
\examples{

#loading data
data(sequenceA)

#computing sum of distances between samples
autosum(sequences = sequenceA, method = "manhattan")

}
\seealso{
\code{\link{distance}}
}
\author{
Blas Benito <blasbenito@gmail.com>
}
