% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_args_x}
\alias{check_args_x}
\title{Checks Argument x}
\usage{
check_args_x(x = NULL, arg_name = "x")
}
\arguments{
\item{x}{(required, list of matrices) list of input matrices generated with [prepare_sequences()]. Default: NULL}

\item{arg_name}{(optional, character string) name of the argument being checked. Default: NULL}
}
\value{
Argument x.
}
\description{
Checks Argument x
}
