% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities_coordinates}
\alias{cities_coordinates}
\title{Coordinates of 100 Major Cities}
\format{
sf data frame with 5 columns and 100 rows.
}
\usage{
data(cities_coordinates)
}
\description{
City coordinates for the dataset \code{cities_temperature}.
}
\seealso{
Other example_data: 
\code{\link{albatross}},
\code{\link{cities_temperature}},
\code{\link{covid_coordinates}},
\code{\link{covid_prevalence}},
\code{\link{eemian_coordinates}},
\code{\link{eemian_pollen}},
\code{\link{fagus_coordinates}},
\code{\link{fagus_dynamics}},
\code{\link{mis}},
\code{\link{tsl_simulate}()},
\code{\link{zoo_simulate}()}

Other example_data: 
\code{\link{albatross}},
\code{\link{cities_temperature}},
\code{\link{covid_coordinates}},
\code{\link{covid_prevalence}},
\code{\link{eemian_coordinates}},
\code{\link{eemian_pollen}},
\code{\link{fagus_coordinates}},
\code{\link{fagus_dynamics}},
\code{\link{mis}},
\code{\link{tsl_simulate}()},
\code{\link{zoo_simulate}()}
}
\concept{example_data}
\keyword{datasets}
