% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities_temperature}
\alias{cities_temperature}
\title{Average Temperature of 100 Major Cities}
\format{
data frame with 3 columns and 52100 rows.
}
\usage{
data(cities_temperature)
}
\description{
Average temperatures between 1975 and 2010 of 100 major cities of the world. [Source](https://www.kaggle.com/datasets/berkeleyearth/climate-change-earth-surface-temperature-data?resource=download&select=GlobalLandTemperaturesByMajorCity.csv).
}
\keyword{datasets}
