% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hclust_optimizer.R
\name{cluster_hclust_optimizer}
\alias{cluster_hclust_optimizer}
\title{Optimal Number of Clusters for Hierarchical Clustering}
\usage{
cluster_hclust_optimizer(d = NULL, method = NULL)
}
\arguments{
\item{d}{(required, matrix) distance matrix typically resulting from [distantia_matrix()], but any other square matrix should work. Default: NULL}
}
\value{
Data frame with number of clusters and their respective mean silhouette widths.
}
\description{
Generates hierarchical clustering solutions from 2 to `nrow(d) - 1` number of clusters and returns the number of clusters with a higher silhouette width median. See [cluster_silhouette()] for more details.
}
