% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_kmeans_optimizer.R
\name{cluster_kmeans_optimizer}
\alias{cluster_kmeans_optimizer}
\title{Optimal Number of Clusters for Kmeans}
\usage{
cluster_kmeans_optimizer(d = NULL, seed = 1)
}
\arguments{
\item{d}{(required, matrix) distance matrix typically resulting from \code{\link[=distantia_matrix]{distantia_matrix()}}, but any other square matrix should work. Default: NULL}

\item{seed}{(optional, integer) Random seed to be used during the K-means computation. Default: 1}
}
\value{
Data frame with number of clusters and their respective mean silhouette widths.
}
\description{
Generates k-means solutions from 2 to \code{nrow(d) - 1} number of clusters and returns the number of clusters with a higher silhouette width median. See \code{\link[=cluster_silhouette]{cluster_silhouette()}} for more details.
}
