% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_silhouette.R
\name{cluster_silhouette}
\alias{cluster_silhouette}
\title{Silhouette Width of a Clustering Solution}
\usage{
cluster_silhouette(cluster = NULL, distance_matrix = NULL, median = FALSE)
}
\arguments{
\item{cluster}{(required, clustering object) Result of [stats::kmeans()]. Default: NULL}

\item{distance_matrix}{(required, matrix) Distance matrix typically resulting from [distantia_matrix()]. Default: NULL}

\item{median}{(optional, logical) If TRUE, the median of the silhouette widths is returned. Default: FALSE}
}
\value{
If median = FALSE, data frame with silhouette widths of the clustering, and numeric indicating the median silhouette width otherwise.
}
\description{
Adapted from https://svn.r-project.org/R-packages/trunk/cluster/R/silhouette.R
}
