% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cost_matrix_diag_cpp}
\alias{cost_matrix_diag_cpp}
\title{Least Cost Matrix Considering Diagonals}
\usage{
cost_matrix_diag_cpp(dist_matrix)
}
\arguments{
\item{dist_matrix}{(required, distance matrix). Square distance matrix, output of [distance_matrix_cpp()].}
}
\value{
Least cost matrix.
}
\description{
Computes the least cost matrix from a distance matrix.
Considers diagonals during computation of least-costs.
}
