% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cost_path_sum_cpp}
\alias{cost_path_sum_cpp}
\title{Sum of Least Cost Distance Times Two}
\usage{
cost_path_sum_cpp(path)
}
\arguments{
\item{path}{(required, data frame) dataframe produced by \code{\link[=cost_path_cpp]{cost_path_cpp()}}. Default: NULL}
}
\value{
Sum of distances
}
\description{
Sum of Least Cost Distance Times Two
}
