% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cost_path_trim_cpp}
\alias{cost_path_trim_cpp}
\title{Trims Blocks from Least Cost Path}
\usage{
cost_path_trim_cpp(path)
}
\arguments{
\item{path}{(required, data frame) dataframe produced by \code{\link[=cost_path]{cost_path()}}. Default: NULL}
}
\value{
A data frame with least-cost path coordinates.
}
\description{
Trims Blocks from Least Cost Path
}
