% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_coordinates}
\alias{covid_coordinates}
\title{Coordinates of the Covid Dataset.}
\format{
sf data frame with 4 columns and 36 rows
}
\usage{
data(covid_coordinates)
}
\description{
Coordinates of the Covid Dataset.
}
\seealso{
Other example_data: 
\code{\link{albatross}},
\code{\link{cities_coordinates}},
\code{\link{covid_prevalence}},
\code{\link{eemian_coordinates}},
\code{\link{eemian_pollen}},
\code{\link{fagus_coordinates}},
\code{\link{fagus_dynamics}},
\code{\link{mis}},
\code{\link{tsl_simulate}()}
}
\concept{example_data}
\keyword{datasets}
