% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_prevalence}
\alias{covid_prevalence}
\title{Covid19 Prevalence Time Series in California}
\format{
data frame with 3 columns and 51048 rows
}
\usage{
data(covid_prevalence)
}
\description{
Dataset with Covid19 prevalence time series in California counties between 2020 and 2024. \href{https://healthdata.gov/State/COVID-19-Time-Series-Metrics-by-County-and-State-A/cr6j-rwfz/about_data}{Source}.
}
\details{
Site coordinates for this dataset are in \link{covid_coordinates}.
}
\examples{
#to time series list
tsl <- tsl_initialize(
  x = covid_prevalence,
  id_column = "county",
  time_column = "date"
)

#time series plot
if(interactive()){

 #subset to avoid margin errors
 tsl_plot(
  tsl = tsl_subset(
    tsl = tsl,
    names = 1:4
    ),
  guide = FALSE
  )

}
}
\seealso{
Other example_data: 
\code{\link{albatross}},
\code{\link{cities_coordinates}},
\code{\link{cities_temperature}},
\code{\link{covid_coordinates}},
\code{\link{eemian_coordinates}},
\code{\link{eemian_pollen}},
\code{\link{fagus_coordinates}},
\code{\link{fagus_dynamics}},
\code{\link{mis}},
\code{\link{tsl_simulate}()},
\code{\link{zoo_simulate}()}
}
\concept{example_data}
\keyword{datasets}
