% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_prevalence}
\alias{covid_prevalence}
\title{Covid19 dataset.}
\format{
data frame with 3 columns and 51048 rows
}
\usage{
data(covid_prevalence)
}
\description{
Dataset with Covid19 prevalence in California counties between 2020 and 2024. \href{https://healthdata.gov/State/COVID-19-Time-Series-Metrics-by-County-and-State-A/cr6j-rwfz/about_data}{Source}.
}
\details{
Site coordinates for this dataset are in \link{covid_coordinates}.
}
\examples{
#to time series list
tsl <- tsl_initialize(
  x = covid_prevalence,
  id_column = "county",
  time_column = "date"
)

#time series plot
if(interactive()){

 #subset to avoid margin errors
 tsl_plot(
  tsl = tsl_subset(
    tsl = tsl,
    names = 1:4
    ),
  guide = FALSE
  )

}
}
\keyword{datasets}
