% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance Between Two Vectors}
\usage{
distance(x = NULL, y = NULL, method = "euclidean")
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as `x`.}

\item{method}{(required, character string) name of a distance metric. Valid entries are: "manhattan", "euclidean", "chi", and "hellinger".}
}
\value{
A distance value.
}
\description{
Computes Manhattan, Euclidean, Chi, or Hellinger distances between numeric vectors of the same length. The underlying C++ functions (see [distance_manhattan_cpp()], [distance_euclidean_cpp()], [distance_hellinger_cpp()], and [distance_chi_cpp()]) do not accept NA as input, and as such, positions with NA in `x` and `y` are ignored.
}
\details{
Distances are computed as:
\itemize{
\item "manhattan" or "man: `sum(abs(x - y))`
\item "euclidean" or "euc": `sqrt(sum((x - y)^2))`.
\item "chi":
    `xy <- x + y`
    `y. <- y / sum(y)`
    `x. <- x / sum(x)`
    `sqrt(sum(((x. - y.)^2) / (xy / sum(xy))))`
\item "hellinger" or "hel": `sqrt(1/2 * sum((sqrt(x) - sqrt(y))^2))`
\item "canberra" or "can": `sum(abs(x - y) / (abs(x) + abs(y)))`
\item "cosine" or "cos": `1 - sum(x * y) / (sqrt(sum(x^2)) * sqrt(sum(y^2)))`.
\item "jaccard" or "jac": `1 - (sum(x & y) / sum(x | y))`.
\item "chebyshev" or "che": `max(abs(x - y))`
\item "russelrao" or "rus": `1 - sum(x == y) / length(x)`
}
Note that zeroes are replaced by `mean(c(x, y)) * 0.0001` whem `method` is "chi".
}
\examples{

distance(
  x = runif(100),
  y = runif(100),
  method = "manhattan"
)

}
\author{
Blas Benito <blasbenito@gmail.com>
}
