% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance Between Numeric Vectors}
\usage{
distance(x = NULL, y = NULL, distance = "euclidean")
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as \code{x}.}

\item{distance}{(optional, character string) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset \code{distances}. Default: "euclidean".}
}
\value{
numeric value
}
\description{
Computes the distance between two numeric or binary vectors.
}
\examples{

distance(
  x = runif(100),
  y = runif(100),
  distance = "euclidean"
)

}
\seealso{
Other psi_demo: 
\code{\link{distances}},
\code{\link{psi_auto_distance}()},
\code{\link{psi_auto_sum}()},
\code{\link{psi_cost_matrix}()},
\code{\link{psi_cost_path}()},
\code{\link{psi_cost_path_sum}()},
\code{\link{psi_distance_lock_step}()},
\code{\link{psi_distance_matrix}()},
\code{\link{psi_equation}()}
}
\concept{psi_demo}
