% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_distance.R
\name{distance}
\alias{distance}
\title{Distance Between Two Vectors}
\usage{
distance(x, y, method = c("manhattan", "chi", "hellinger", "euclidean"))
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as `x`.}

\item{method}{(required, character string) name of a distance metric. Valid entries are: "manhattan", "euclidean", "chi", and "hellinger".}
}
\value{
A distance value.
}
\description{
Computes Manhattan, Euclidean, Chi, or Hellinger distances between numeric vectors of the same length. Used internally within [distanceMatrix()] and [autoSum()].
}
\details{
Vectors \code{x} and \code{y} are not checked to speed-up execution time. Distances are computed as:
\itemize{
\item "manhattan": `sum(abs(x - y))`
\item "euclidean": `qrt(sum((x - y)^2))`.
\item "chi":
    `xy <- x + y`
    `y. <- y / sum(y)`
    `x. <- x / sum(x)`
    `sqrt(sum(((x. - y.)^2) / (xy / sum(xy))))`
\item "hellinger": `sqrt(1/2 * sum((sqrt(x) - sqrt(y))^2))`
}
Note that zeroes are replaced by `mean(c(x, y)) * 0.001`  whem `method` is "chi" or "hellinger".
}
\examples{

distance(
  x = runif(100),
  y = runif(100),
  method = "manhattan"
)

}
\author{
Blas Benito <blasbenito@gmail.com>
}
