% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_chi_cpp}
\alias{distance_chi_cpp}
\title{Normalized Chi Distance Between Two Vectors}
\usage{
distance_chi_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as \code{x}.}
}
\value{
Chi distance between x and y.
}
\description{
Computed as:
\code{xy <- x + y}
\code{y. <- y / sum(y)}
\code{x. <- x / sum(x)}
\code{sqrt(sum(((x. - y.)^2) / (xy / sum(xy))))}.
Cannot handle NA values. When \code{x} and \code{y} have zeros in the same
position, \code{NaNs} are produced. Please replace these zeros with
pseudo-zeros (i.e. 0.0001) if you wish to use this distance metric.
}
\examples{
distance_chi_cpp(x = runif(100), y = runif(100))
}
