% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_cosine_cpp}
\alias{distance_cosine_cpp}
\title{(C++) Cosine Dissimilarity Between Two Vectors}
\usage{
distance_cosine_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as \code{x}.}
}
\value{
numeric
}
\description{
Computes the cosine dissimilarity between two numeric vectors.
}
\examples{
distance_cosine_cpp(c(0.2, 0.4, 0.5), c(0.1, 0.8, 0.2))
}
\seealso{
Other Rcpp_distance_methods: 
\code{\link{distance_bray_curtis_cpp}()},
\code{\link{distance_canberra_cpp}()},
\code{\link{distance_chebyshev_cpp}()},
\code{\link{distance_chi_cpp}()},
\code{\link{distance_euclidean_cpp}()},
\code{\link{distance_hamming_cpp}()},
\code{\link{distance_hellinger_cpp}()},
\code{\link{distance_jaccard_cpp}()},
\code{\link{distance_manhattan_cpp}()},
\code{\link{distance_russelrao_cpp}()},
\code{\link{distance_sorensen_cpp}()}
}
\concept{Rcpp_distance_methods}
