% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_hamming_cpp}
\alias{distance_hamming_cpp}
\title{Hamming Distance Between Two Binary Vectors}
\usage{
distance_hamming_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as \code{x}.}
}
\value{
Hamming distance between x and y.
}
\description{
Computes the Hamming distance between two binary vectors.
}
\examples{
distance_hamming_cpp(c(0, 1, 0, 1), c(1, 1, 0, 0))
}
