% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_hellinger_cpp}
\alias{distance_hellinger_cpp}
\title{Hellinger Distance Between Two Vectors}
\usage{
distance_hellinger_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as `x`.}
}
\value{
Hellinger distance between x and y.
}
\description{
Hellinger Distance Between Two Vectors
}
\details{
Cannot handle NA values.
}
\examples{
distance_hellinger_cpp(x = runif(100), y = runif(100))
}
