% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_jaccard_cpp}
\alias{distance_jaccard_cpp}
\title{Jaccard Distance Between Two Binary Vectors}
\usage{
distance_jaccard_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric vector).}

\item{y}{(required, numeric vector) of same length as \code{x}.}
}
\value{
Jaccard distance between x and y.
}
\description{
Computes the Jaccard distance between two binary vectors.
}
\examples{
distance_jaccard_cpp(c(0, 1, 0, 1), c(1, 1, 0, 0))
}
