% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Distance Matrix}
\usage{
distance_matrix(a = NULL, b = NULL, method = "euclidean")
}
\arguments{
\item{a}{(required, data frame or matrix) a time series.}

\item{b}{(required, data frame or matrix) a time series.}

\item{method}{(optional, character string) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `methods`. Default: "euclidean".}
}
\value{
A distance matrix.
}
\description{
Distance Matrix
}
\examples{

data(
  sequenceA,
  sequenceB
)

distance.matrix <- distance_matrix(
  a = sequenceA,
  b = sequenceB,
  method = "manhattan"
)

if(interactive()){
 plotMatrix(distance.matrix = distance.matrix)
}

}
