% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Distance Matrix}
\usage{
distance_matrix(a, b, method = c("manhattan", "chi", "hellinger", "euclidean"))
}
\arguments{
\item{a}{(required, data frame) a time series.}

\item{b}{(required, data frame) a time series.}

\item{method}{(optional, character string) name of the distance metric. Valid entries are: "manhattan", "euclidean", "chi", and "hellinger". Default: "manhattan}
}
\value{
A distance matrix.
}
\description{
Distance Matrix
}
\examples{

data(
  sequenceA,
  sequenceB
)

distance.matrix <- distance_matrix(
  a = sequenceA,
  b = sequenceB,
  method = "manhattan"
)

if(interactive()){
 plotMatrix(distance.matrix = distance.matrix)
}

}
