% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_matrix_cpp}
\alias{distance_matrix_cpp}
\title{Distance Matrix}
\usage{
distance_matrix_cpp(a, b, f)
}
\arguments{
\item{a}{(required, numeric matrix).}

\item{b}{(required, numeric matrix) of same ncol as `a`.}

\item{f}{(required, function) a function. One of:
\itemize{
  \item [distance_manhattan_cpp].
  \item [distance_euclidean_cpp].
  \item [distance_hellinger_cpp].
  \item [distance_chi_cpp]
}}
}
\value{
Matrix of distances between `a` (rows) and `b` (columns).
}
\description{
Computes the distance matrix between two matrices
\code{a} and \code{b} using one distance function.
}
\details{
Cannot handle NA values.
}
