% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_russelrao_cpp}
\alias{distance_russelrao_cpp}
\title{Russell-Rao Distance Between Two Binary Vectors}
\usage{
distance_russelrao_cpp(x, y)
}
\arguments{
\item{x}{(required, numeric). Binary vector of 1s and 0s.}

\item{y}{(required, numeric) Binary vector of 1s and 0s of same length as \code{x}.}
}
\value{
Russell-Rao distance between x and y.
}
\description{
Computes the Russell-Rao distance between two binary vectors.
}
\examples{
distance_russelrao_cpp(c(0, 1, 0, 1), c(1, 1, 0, 0))
}
