% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distances}
\alias{distances}
\title{Distance Methods}
\format{
data frame with 5 columns and 10 rows
}
\usage{
data(distances)
}
\description{
Data frame with the names, abbreviations, and expressions of the distance metrics implemented in the package.
}
\seealso{
Other psi_demo: 
\code{\link{distance}()},
\code{\link{psi}()},
\code{\link{psi_auto_distance}()},
\code{\link{psi_auto_sum}()},
\code{\link{psi_cost_matrix}()},
\code{\link{psi_cost_path}()},
\code{\link{psi_cost_path_ignore_blocks}()},
\code{\link{psi_cost_path_sum}()},
\code{\link{psi_dist_lock_step}()},
\code{\link{psi_dist_matrix}()}
}
\concept{psi_demo}
\keyword{datasets}
