% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia-package.R
\docType{package}
\name{distantia-package}
\alias{distantia-package}
\title{distantia: An Introduction to the distantia Package}
\description{
TODO: complete this section
The \code{distantia} package provides tools for analyzing and visualizing distances between
data points. It includes functions for calculating various distance metrics, performing
clustering analysis, and visualizing distance matrices.

The package includes the following main functions:
\itemize{
\item \code{\link{calculate_distance}}: Computes the distance between two sets of data points.
\item \code{\link{plot_distance_matrix}}: Visualizes the distance matrix as a heatmap.
\item \code{\link{cluster_data}}: Clusters data points based on distance metrics.
}
}
\section{Installation}{

To install the \code{distantia} package from CRAN, use:
\preformatted{install.packages("distantia")}
}

\section{Getting Started}{

After installation, load the package with:
\preformatted{library(distantia)}
For detailed documentation on individual functions, use \code{help(package = "distantia")}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/blasbenito/distantia/}
}

}
\author{
\strong{Maintainer}: Blas M. Benito \email{blasbenito@gmail.com} (\href{https://orcid.org/0000-0001-5105-7232}{ORCID}) [copyright holder]

}
\keyword{internal}
