% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia.R
\docType{package}
\name{distantia-package}
\alias{distantia}
\alias{distantia-package}
\title{distantia: Assessing Dissimilarity Between Multivariate Time Series}
\description{
Provides tools to assess the dissimilarity between multivariate time-series. It is based on the psi measure described  by Birks and Gordon (1985 <doi:10.1002/jqs.3390020110>), which computes dissimilarity between irregular time-series constrained by sample order. However, in this package the original idea has been extended to work with any kind of multivariate time-series, no matter whether they are regular, irregular, aligned or unaligned. Furthermore, the package allows to assess the significance of dissimilarity values by applying a restricted permutation test, allows to measure the contribution of individual variables to dissimilarity, and offers tools to transfer attributes (generally time or age, but other are possible) between sequences based on the similarity of their samples.
}
\details{
Details
}
\seealso{
Useful links:
\itemize{
  \item \url{https://blasbenito.github.io/distantia/}
}

}
