% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia.R
\name{distantia}
\alias{distantia}
\title{Psi Distance Between Time Series}
\usage{
distantia(
  a = NULL,
  b = NULL,
  method = c("euclidean", "manhattan"),
  diagonal = c(FALSE, TRUE),
  weighted = c(FALSE, TRUE),
  ignore_blocks = c(FALSE, TRUE),
  paired_samples = FALSE,
  repetitions = 100,
  permutation = "restricted_by_row",
  block_size = c(2, 3, 4),
  seed = 1
)
}
\arguments{
\item{a}{(required, data frame or matrix) a time series.}

\item{b}{(required, data frame or matrix) a time series.}

\item{method}{(optional, character string) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `methods`. Default: "euclidean".}

\item{diagonal}{(optional, logical). If TRUE, diagonals are included in the computation of the cost matrix. Default: FALSE.}

\item{weighted}{If TRUE, diagonal is set to TRUE, and diagonal cost is weighted by a factor of 1.414214. Default: FALSE.}

\item{ignore_blocks}{(optional, logical). If TRUE, blocks of consecutive path coordinates are trimmed to avoid inflating the psi distance. Default: FALSE.}

\item{paired_samples}{(optional, logical) If TRUE, time-series are compared row wise and no least-cost path is computed. Default: FALSE.}

\item{repetitions}{(optional, integer) number of permutations to compute the p-value (interpreted as the probability of finding a smaller dissimilarity on permuted versions of the sequences) of the psi distance. If 0, p-values are not computed. Otherwise, the minimum is 2. Default: 0}

\item{permutation}{(optional, character) permutation method. Valid values are listed below from higher to lower randomness:
\itemize{
  \item "free": unrestricted shuffling of rows and columns. Ignores block_size.
  \item "free_by_row": unrestricted shuffling of complete rows. Ignores block size.
  \item "restricted": restricted shuffling of rows and columns within blocks.
  \item "restricted_by_row": restricted shuffling of rows within blocks.
}}

\item{block_size}{(optional, integer vector) vector with block sizes for the restricted permutation test. A block of size 3 indicates that a row can only be permuted within a block of 3 adjacent rows. If several values are provided, one is selected at random separately for each sequence on each repetition. Only relevant when permutation methods are "restricted" or "restricted_by_row". Default: 3.}

\item{seed}{(optional, integer) initial random seed to use for replicability when computing p-values. Default: 1}
}
\value{
Data frame with names of the sequences, combinations of arguments method, diagonal, weighted, and trim blocks, psi distance, and if `p_value = TRUE`, mean and standard deviation of the null distribution for the psi distance, and the p_value, interpreted as the probability of finding psi distances smaller than the observed one in permuted versions of the sequences.
}
\description{
Psi Distance Between Time Series
}
\examples{

data(
  sequenceA,
  sequenceB
)

ab.psi <- distantia(
  a = sequenceA,
  b = sequenceB,
  method = "manhattan"
)

}
