% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia.R
\name{distantia}
\alias{distantia}
\title{Psi Distance Between Time Series}
\usage{
distantia(
  a = NULL,
  b = NULL,
  method = c("euclidean", "manhattan"),
  diagonal = c(FALSE, TRUE),
  weighted = c(FALSE, TRUE),
  trim_blocks = c(FALSE, TRUE),
  paired_samples = FALSE,
  p_value = FALSE,
  block_size = c(10),
  seed = 1,
  repetitions = 100
)
}
\arguments{
\item{a}{(required, data frame or matrix) a time series.}

\item{b}{(required, data frame or matrix) a time series.}

\item{method}{(optional, character string) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `methods`. Default: "euclidean".}

\item{diagonal}{(optional, logical). If TRUE, diagonals are included in the computation of the cost matrix. Default: FALSE.}

\item{weighted}{If TRUE, diagonal is set to TRUE, and diagonal cost is weighted by a factor of 1.414214. Default: FALSE.}

\item{trim_blocks}{(optional, logical). If TRUE, blocks of consecutive path coordinates are trimmed to avoid inflating the psi distance. Default: FALSE.}

\item{paired_samples}{(optional, logical) If TRUE, time-series are compared row wise and no least-cost path is computed. Default: FALSE.}

\item{p_value}{(optional, logical) If TRUE, the probability of finding a psi distance smaller than the observed one in permuted versions of the sequences/time-series is returned in the column "p_value". The observed psi distance is added to the distribution of permuted psi values to compute the p-value. Default: FALSE}

\item{block_size}{(optional, integer vector) vector with block sizes for the restricted permutation test. A block of size 3 indicates that a row can only be permuted within a block of 3 adjacent rows. Only relevant when `p_value = TRUE`. Default: c(2, 3, 4).}

\item{seed}{(optional, integer) initial random seed to use for replicability. Only relevant when `p_value = TRUE`. Default: 1}

\item{repetitions}{(optional, integer) number of permutations to compute the p-value of the psi distance. Only relevant when `p_value = TRUE`. Default: 100.}
}
\value{
Data frame with names of the sequences, combinations of arguments method, diagonal, weighted, and trim blocks, psi distance, and if `p_value = TRUE`, mean and standard deviation of the null distribution for the psi distance, and the p_value, interpreted as the probability of finding psi distances smaller than the observed one in permuted versions of the sequences.
}
\description{
Psi Distance Between Time Series
}
\examples{

data(
  sequenceA,
  sequenceB
)

ab.psi <- distantia(
  a = sequenceA,
  b = sequenceB,
  method = "manhattan"
)

}
