% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia.R
\name{distantia}
\alias{distantia}
\title{Psi Distance Between Time Series}
\usage{
distantia(
  x = NULL,
  distance = "euclidean",
  diagonal = FALSE,
  weighted = FALSE,
  ignore_blocks = FALSE,
  paired_samples = FALSE,
  repetitions = 0L,
  permutation = "restricted_by_row",
  block_size = 3,
  seed = 1
)
}
\arguments{
\item{x}{(required, list of matrices) list of input matrices generated with [tsl_prepare()].}

\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}

\item{diagonal}{(optional, logical vector). If TRUE, diagonals are included in the computation of the cost matrix. Default: FALSE.}

\item{weighted}{(optional, logical vector) If TRUE, diagonal is set to TRUE, and diagonal cost is weighted by a factor of 1.414214. Default: FALSE.}

\item{ignore_blocks}{(optional, logical vector). If TRUE, blocks of consecutive path coordinates are trimmed to avoid inflating the psi distance. Default: FALSE.}

\item{paired_samples}{(optional, logical vector) If TRUE, time-series are compared row wise and no least-cost path is computed. Default: FALSE.}

\item{repetitions}{(optional, integer vector) number of permutations to compute the p-value (interpreted as the probability of finding a smaller dissimilarity on permuted versions of the sequences) of the psi distance. If 0, p-values are not computed. Otherwise, the minimum is 2. Default: 0}

\item{permutation}{(optional, character vector) permutation method. Valid values are listed below from higher to lower randomness:
\itemize{
  \item "free": unrestricted shuffling of rows and columns. Ignores block_size.
  \item "free_by_row": unrestricted shuffling of complete rows. Ignores block size.
  \item "restricted": restricted shuffling of rows and columns within blocks.
  \item "restricted_by_row": restricted shuffling of rows within blocks.
}}

\item{block_size}{(optional, integer vector) vector with block sizes in rows for the restricted permutation test. A block of size 3 indicates that a row can only be permuted within a block of 3 adjacent rows. If several values are provided, one is selected at random separately for each sequence on each repetition. Only relevant when permutation methods are "restricted" or "restricted_by_row". Default: 3.}

\item{seed}{(optional, integer) initial random seed to use for replicability when computing p-values. Default: 1}
}
\value{
Data frame with the following columns:
\itemize(
  \item `name_a`: name of the sequence `a`.
  \item `name_bb`: name of the sequence `b`.
  \item `distance`: name of the distance metric.
  \item `diagonal`: value of the argument `diagonal`.
  \item `weighted`: value of the argument `weighted`.
  \item `ignore_blocks`: value of the argument `ignore_blocks`.
  \item `paired_samples`: value of the argument `paired_samples`. Only available if TRUE appears in the argument and the sequences have the same number of rows.
  \item `repetitions` (only if `repetitions > 0`): value of the argument `repetitions`.
  \item `permutation` (only if `repetitions > 0`): name of the permutation method used to compute p-values.
  \item `seed` (only if `repetitions > 0`): random seed used to in the permutations.
  \item `null_mean` (only if `repetitions > 0`): mean of the null distribution of psi values computed from the permutaitons.
  \item `null_sd` (only if `repetitions > 0`): standard deviation of the null distribution of psi values.
  \item `p_value`  (only if `repetitions > 0`): proportion of scores smaller or equal than `psi` in the null distribution.
  \item `psi`: psi dissimilarity of the sequences `a` and `b`.
)
}
\description{
Psi Distance Between Time Series
}
