% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia.R
\name{distantia}
\alias{distantia}
\title{Psi Distance Between Time Series}
\usage{
distantia(a, b, method, diagonal, weighted, trim_blocks, paired_samples)
}
\arguments{
\item{a}{(required, data frame or matrix) a time series.}

\item{b}{(required, data frame or matrix) a time series.}

\item{method}{(optional, character string) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `methods`. Default: "euclidean".}

\item{diagonal}{(optional, logical). If TRUE, diagonals are included in the computation of the cost matrix. Default: FALSE.}

\item{weighted}{If TRUE, diagonal is set to TRUE, and diagonal cost is weighted by a factor of 1.414214. Default: FALSE.}

\item{trim_blocks}{(optional, logical). If TRUE, blocks of consecutive path coordinates are trimmed to avoid inflating the psi distance. Default: FALSE.}

\item{paired_samples}{(optional, logical) If TRUE, time-series are compared row wise and no least-cost path is computed. Default: FALSE}
}
\value{
Psi distance
}
\description{
Psi Distance Between Time Series
}
\examples{

data(
  sequenceA,
  sequenceB
)

ab.psi <- distantia(
  a = sequenceA,
  b = sequenceB,
  method = "manhattan"
)

}
