% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_importance.R
\name{distantia_importance}
\alias{distantia_importance}
\title{Contribution of Individual Variables to Dissimilarity}
\usage{
distantia_importance(
  tsl = NULL,
  distance = "euclidean",
  diagonal = TRUE,
  weighted = TRUE,
  ignore_blocks = FALSE,
  lock_step = FALSE,
  robust = TRUE
)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset \code{distances}. Default: "euclidean".}

\item{diagonal}{(optional, logical vector). If TRUE, diagonals are included in the computation of the cost matrix. Default: TRUE}

\item{weighted}{(optional, logical vector) If TRUE, diagonal is set to TRUE, and diagonal cost is weighted by a factor of 1.414214. Default: TRUE}

\item{ignore_blocks}{(optional, logical vector). If TRUE, blocks of consecutive path coordinates are trimmed to avoid inflating the psi distance. Ignored if \code{diagonal = TRUE}. Default: FALSE.}

\item{lock_step}{(optional, logical vector) If TRUE, time series are compared row wise and no least-cost path optimization is performed. Requires time series in argument \code{tsl} to be fully aligned. Default: FALSE.}

\item{robust}{(required, logical). If TRUE (default), importance scores are computed using the least cost path of the complete time series as reference. Setting it to FALSE allows to replicate importance scores of the previous versions of this package. This option is irrelevant when \code{lock_step = TRUE}. Default: TRUE}
}
\value{
data frame:
\itemize{
\item \code{x}: name of the time series \code{x}.
\item \code{y}: name of the time series \code{y}.
\item \code{psi}: psi score of \code{x} and \code{y}.
\item \code{variable}: name of the individual variable.
\item \code{importance}: importance score of the variable.
\item \code{psi_only_with}: psi score of the variable.
\item \code{psi_without}: psi score without the variable.
\item \code{psi_difference}: difference between \code{psi_only_with} and \code{psi_without}.
\item \code{distance}: name of the distance metric.
\item \code{diagonal}: value of the argument \code{diagonal}.
\item \code{weighted}: value of the argument \code{weighted}.
\item \code{ignore_blocks}: value of the argument \code{ignore_blocks}.
\item \code{lock_step}: value of the argument \code{lock_step}.
\item \code{robust}: value of the argument \code{robust}.
}
}
\description{
This function measures the contribution of individual variables to the dissimilarity between pairs of time series to help answer the question \emph{what makes two time series more or less similar?}

Three key values are required to assess individual variable contributions:
\itemize{
\item \strong{psi}: dissimilarity when all variables are considered.
\item \strong{psi_only_with}: dissimilarity when using only the target variable.
\item \strong{psi_without}: dissimilarity when removing the target variable.
}

The values \code{psi_only_with} and \code{psi_without} can be computed in two different ways defined by the argument \code{robust}.
\itemize{
\item \code{robust = FALSE}: This method replicates the importance algorithm released with the first version of the package, and it is only recommended when the goal to compare new results with previous studies. It normalizes \code{psi_only_with} and \code{psi_without} using the least cost path obtained from the individual variable. As different variables may have different least cost paths for the same time series, normalization values may change from variable to variable, making individual importance scores harder to compare.
\item \code{robust = TRUE} (default, recommended): This a novel version of the importance algorithm that yields more stable and comparable solutions. It uses the least cost path of the complete time series to normalize \code{psi_only_with} and \code{psi_without}, making importance scores of separate variables fully comparable.
}

The individual importance score of each variable (column "importance" in the output data frame) is based on different expressions depending on the \code{robust} argument, even when \code{lock_step = TRUE}:
\itemize{
\item \code{robust = FALSE}: Importance is computed as \code{((psi - psi_without) * 100)/psi} and interpreted as "change in similarity when a variable is removed".
\item \code{robust = TRUE}: Importance is computed as \code{((psi_only_with - psi_without) * 100)/psi} and interpreted as "relative dissimilarity induced by the variable expressed as a percentage".
}

In either case, positive values indicate that the variable contributes to dissimilarity, while negative values indicate a net contribution to similarity.

This function allows computing dissimilarity between pairs of time series using different combinations of arguments at once. For example, when the argument \code{distance} is set to \code{c("euclidean", "manhattan")}, the output data frame will show two dissimilarity scores for each pair of time series, one based on euclidean distances, and another based on manhattan distances. The same happens for most other parameters.

This function supports progress bars generated by the \code{progressr} package. See examples.

This function also accepts a parallelization setup via \code{\link[future:plan]{future::plan()}}, but it might only be worth it for very long time series.
}
\examples{

#parallelization setup (not worth it for this data size)
future::plan(
 future::multisession,
 workers = 2 #set to parallelly::availableWorkers() - 1
)

#progress bar
if(interactive()){
  progressr::handlers(global = TRUE)
}

#three time series
#climate and ndvi in Fagus sylvatica stands in Spain, Germany, and Sweden
data("fagus_dynamics")

#load as tsl
#center and scale with same parameters
tsl <- tsl_initialize(
  x = fagus_dynamics,
  id_column = "site",
  time_column = "date"
) |>
  tsl_transform(
    f = f_scale
  )

if(interactive()){
  tsl_plot(
    tsl = tsl,
    guide_columns = 3
    )
}


#importance computed with dynamic time warping
#less sensitive to latitudinal or altitudinal differences
importance_dtw <- distantia_importance(
  tsl = tsl
)


#focus on important columns
importance_dtw[, c(
  "x",
  "y",
  "psi",
  "variable",
  "importance",
  "effect"
  )]

#Interpretation example:
#variable contributing the most to similarity between Germany and Sweden: temperature
#variable contributing the most to dissimilarity between Spain and Sweden: temperature

#importance computed with lock-step method
#more sensitive to latitudinal or altitudinal differences
importance_lock_step <- distantia_importance(
  tsl = tsl,
  lock_step = TRUE
)

importance_lock_step[, c(
  "x",
  "y",
  "psi",
  "variable",
  "importance",
  "effect"
  )]


#combinations of parameters
#---------------------------------
#most arguments accept vectors, and the results contain all argument combinations

importance_df <- distantia_importance(
  tsl = tsl,
  lock_step = c(TRUE, FALSE)
)

importance_df[, c(
  "x",
  "y",
  "psi",
  "variable",
  "importance",
  "effect",
  "lock_step"
)]

#disable parallelization
future::plan(
  future::sequential
)
}
\seealso{
Other dissimilarity_analysis: 
\code{\link{distantia}()},
\code{\link{distantia_aggregate}()},
\code{\link{distantia_boxplot}()},
\code{\link{distantia_cluster_hclust}()},
\code{\link{distantia_cluster_kmeans}()},
\code{\link{distantia_matrix}()},
\code{\link{distantia_plot}()},
\code{\link{distantia_to_sf}()}
}
\concept{dissimilarity_analysis}
