% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_kmeans.R
\name{distantia_kmeans}
\alias{distantia_kmeans}
\title{Kmeans Clustering of a Distantia Data Frame}
\usage{
distantia_kmeans(df = NULL, clusters = NULL, seed = 1)
}
\arguments{
\item{df}{(required, data frame) Output of [distantia()]. Default: NULL}

\item{clusters}{(required, integer) Number of groups to generate. If NULL (default), [cluster_kmeans_optimizer()] is used to find the number of clusters that maximizes the mean silhouette width (see [cluster_silhouette()]). Default: NULL}

\item{seed}{(optional, integer) Random seed to be used during the K-means computation. Default: 1}
}
\value{
List with the following objects:
\itemize{
  \item `df`: data frame with names of the time series, their group assignation, and their silhouette width scores.
  \item `d`: psi distance matrix.
  \item `k`: kmeans object.
  \item `silhouette_width`: mean silhouette width of the kmeans clustering.
  \item `silhouette_optimization`: only if `clusters = NULL`, data frame with optimization results from [cluster_kmeans_optimizer()].
}
}
\description{
Kmeans Clustering of a Distantia Data Frame
}
\examples{

tsl <- tsl_simulate(
  n = 10,
  time_range = c(
    "2010-01-01 12:00:25",
    "2024-12-31 11:15:45"
  )
)

df <- distantia(
  tsl = tsl
)

k <- distantia_kmeans(
  df = df
)

k$df

# #Optional: kmeans plot
# k_plot <- factoextra::fviz_cluster(
#   object = k$k,
#   data = k$d,
#   repel = TRUE
# )

}
\seealso{
[stats::kmeans()]
}
