% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_model_frame.R
\name{distantia_model_frame}
\alias{distantia_model_frame}
\title{Dissimilarity Model Frame}
\usage{
distantia_model_frame(
  df = NULL,
  predictors_df = NULL,
  predictors_list = NULL,
  predictors_scaled = FALSE,
  distance = "euclidean"
)
}
\arguments{
\item{df}{(required, data frame) Output of \code{\link[=distantia]{distantia()}} or \code{\link[=distantia_aggregate]{distantia_aggregate()}}. Default: NULL}

\item{predictors_df}{(required, data frame) data frame with numeric predictors to be added to the model frame. Must have a column with the names in \code{df$x} and \code{df$y}. If \code{sf} data frame, the predictor "distance" is added to the model frame. Default: NULL}

\item{predictors_list}{(optional, list) list defining new predictors as combinations of other predictors in \code{predictors_df}. For example, \code{predictors_list = list(a = c("b", "c"))} uses the columns \code{"b"} and \code{"c"} from \code{predictors_df} to generate the predictor \code{a} in the model frame. Default: NULL}

\item{predictors_scaled}{(optional, logical) if TRUE, all predictors are scaled and centered with \code{\link[=scale]{scale()}}. Default: FALSE}

\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset \link{distances}. Default: "euclidean".}
}
\value{
data frame: with attributes "response", "predictors" and "formula".
}
\description{
This function generates a model frame for statistical or machine learning analysis from these objects:

\itemize{
\item: Dissimilarity data frame generated by \code{\link[=distantia]{distantia()}}. The output model frame will have as many rows as this data frame.
\item: Data frame with static descriptors of the time series. These descriptors are converted to distances between pairs of time series via \code{\link[=distance_matrix]{distance_matrix()}}.
\item: List defining new predictors as combinations of other existing predictors. This feature allows grouping together predictors that have a common meaning. For example, \verb{predictors_list = list(temperature = c("temperature_mean", "temperature_min", "temperature_max")} generates a new predictor named "temperature", which results from computing the distances of the vector of temperature variables for each pair of time series. Predictors in one of such groups will be scaled before distance computation if their maximum standard deviations differ by a factor of 10 or more.
}

The resulting data frame contains the following columns:
\itemize{
\item \code{x} and \code{y}: names of the pair of time series represented in the row.
\item \code{psi}: dissimilarity between \code{x} and \code{y}.
\item predictors columns: representing the distance between the values of the given static predictor between \code{x} and \code{y}.
\item (optional) \code{distance}:If the static predictors data frame is an \code{sf} object, then this predictor is created via \code{sf::st_distance()}.
}

Statistical or machine learning analyses based on this data frame may help uncover drivers of dissimilarity. Model coefficients or importance scores generated from this model frame represent the effect of the distance between predictors on the dissimilarity between time seriess.

This function supports a parallelization setup via \code{\link[future:plan]{future::plan()}}, and progress bars provided by the package \href{https://CRAN.R-project.org/package=progressr}{progressr}.
}
\examples{

#covid prevalence in California counties
tsl <- tsl_initialize(
  x = covid_prevalence,
  name_column = "name",
  time_column = "time"
)

#dissimilarity analysis
df <- distantia(
  tsl = tsl,
  lock_step = TRUE
)

#combine several predictors
#into a new one
predictors_list <- list(
  economy = c(
    "poverty_percentage",
    "median_income",
    "domestic_product"
    )
)

#generate model frame
model_frame <- distantia_model_frame(
  df = df,
  predictors_df = covid_counties,
  predictors_list = predictors_list,
  predictors_scaled = TRUE
)

head(model_frame)

#names of response and predictors
#and an additive formula
#are stored as attributes
attributes(model_frame)$response
attributes(model_frame)$predictors
attributes(model_frame)$formula


#linear model
model <- lm(
  formula = attributes(model_frame)$formula,
  data = model_frame
)

summary(model)

}
\seealso{
Other dissimilarity_analysis_main: 
\code{\link{distantia}()},
\code{\link{momentum}()}
}
\concept{dissimilarity_analysis_main}
