% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_spatial_network.R
\name{distantia_spatial_network}
\alias{distantia_spatial_network}
\title{Convert Dissimilarity Analysis Data Frames to Spatial Dissimilarity Networks}
\usage{
distantia_spatial_network(df = NULL, sf = NULL, one_to_many = FALSE)
}
\arguments{
\item{df}{(required, data frame) Output of \code{\link[=distantia]{distantia()}} or \code{\link[=distantia_aggregate]{distantia_aggregate()}}. Default: NULL}

\item{sf}{(required, sf data frame) Points or polygons representing the location of the time series in argument 'df'. It must have a column with all time series names in \code{df$x} and \code{df$y}. Default: NULL}

\item{one_to_many}{(optional, logical) If TRUE, the resulting data frame is reorganized to help map one-to-many relationships using the original geometry in \code{sf} rather than network edges. Default: FALSE}
}
\value{
sf data frame (LINESTRING geometry)
}
\description{
Given an sf data frame with the coordinates of a time series list transforms a data frame resulting from \code{\link[=distantia]{distantia()}} to an sf data frame with lines connecting the time series locations. This can be used to visualize a geographic network of similarity/dissimilarity between locations. See example for further details.
}
\examples{
tsl <- distantia::tsl_initialize(
  x = distantia::covid_prevalence,
  name_column = "name",
  time_column = "time"
)

df_psi <- distantia::distantia_ls(
  tsl = tsl
)

#network many to many
sf_network <- distantia::distantia_spatial_network(
  df = df_psi,
  sf = distantia::covid_counties
)

#subset target counties
counties <- c("Los_Angeles", "San_Francisco", "Fresno", "San_Joaquin")

sf_network_subset <- sf_network[
  which(sf_network$x \%in\% counties & sf_network$y \%in\% counties), ]

#network map
# mapview::mapview(
#   distantia::covid_counties,
#   col.regions = NA,
#   color = "black",
#   label = "name",
#   legend = FALSE,
#   map.type = "OpenStreetMap"
# ) +
#   mapview::mapview(
#     sf_network_subset,
#     layer.name = "Psi",
#     label = "edge_name",
#     zcol = "psi",
#     lwd = 3
#   ) |>
#   suppressWarnings()

#one to many
sf_network <- distantia_spatial_network(
  df = df_psi,
  sf = distantia::covid_counties,
  one_to_many = TRUE
)

#subset one county
sf_network_subset <- sf_network[sf_network$x == "Los_Angeles", ]

#one to many map
#dissimilarity of all counties with Los Angeles
# mapview::mapview(
#   sf_network_subset,
#   layer.name = "Psi with LA",
#   label = "y",
#   zcol = "psi",
#   alpha.regions = 1
# ) |>
#   suppressWarnings()
}
\seealso{
Other distantia_support: 
\code{\link{distantia_aggregate}()},
\code{\link{distantia_boxplot}()},
\code{\link{distantia_cluster_hclust}()},
\code{\link{distantia_cluster_kmeans}()},
\code{\link{distantia_dtw_shift}()},
\code{\link{distantia_matrix}()},
\code{\link{distantia_model_frame}()},
\code{\link{distantia_stats}()},
\code{\link{utils_block_size}()},
\code{\link{utils_cluster_hclust_optimizer}()},
\code{\link{utils_cluster_kmeans_optimizer}()},
\code{\link{utils_cluster_silhouette}()}
}
\concept{distantia_support}
