% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distantia_stats.R
\name{distantia_stats}
\alias{distantia_stats}
\title{Stats of Dissimilarity Data Frame}
\usage{
distantia_stats(df = NULL)
}
\arguments{
\item{df}{(required, data frame) Output of \code{\link[=distantia]{distantia()}} or \code{\link[=momentum]{momentum()}}. Default: NULL}
}
\value{
data frame
}
\description{
Takes the output of \code{\link[=distantia]{distantia()}} to return a data frame with one row per time series with the stats of its dissimilarity scores with all other time series.
}
\examples{
tsl <- tsl_simulate(
  n = 5,
  irregular = FALSE
  )

df <- distantia(
  tsl = tsl,
  lock_step = TRUE
  )

df_stats <- distantia_stats(df = df)

df_stats
}
