% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eemian_coordinates}
\alias{eemian_coordinates}
\title{Coordinates of Nine Eemian Sites}
\format{
sf data frame with 4 columns and 9 rows.
}
\usage{
data(eemian_coordinates)
}
\description{
Coordinates of Nine Eemian Sites
}
\seealso{
Other example_data: 
\code{\link{albatross}},
\code{\link{cities_coordinates}},
\code{\link{covid_coordinates}},
\code{\link{covid_prevalence}},
\code{\link{eemian_pollen}},
\code{\link{fagus_coordinates}},
\code{\link{fagus_dynamics}},
\code{\link{mis}},
\code{\link{tsl_simulate}()}

Other example_data: 
\code{\link{albatross}},
\code{\link{cities_coordinates}},
\code{\link{covid_coordinates}},
\code{\link{covid_prevalence}},
\code{\link{eemian_pollen}},
\code{\link{fagus_coordinates}},
\code{\link{fagus_dynamics}},
\code{\link{mis}},
\code{\link{tsl_simulate}()}
}
\concept{example_data}
\keyword{datasets}
