% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_center}
\alias{f_center}
\title{Global Centering}
\usage{
f_center(x = NULL, center = TRUE, scale = FALSE, ...)
}
\arguments{
\item{x}{Data frame or matrix.}
}
\value{
Data frame or matrix
}
\description{
Centers using the global average of all zoo objects within a time series list. Please use [base::scale()] to scale each time series independently.
}
