% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_center}
\alias{f_center}
\title{Centering}
\usage{
f_center(x = NULL, center = TRUE, scale = FALSE)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{center}{(optional, logical or numeric vector) Triggers centering if TRUE. Default: TRUE}

\item{scale}{(optional, logical or numeric vector) Triggers scaling if TRUE. Default: FALSE}
}
\value{
zoo object
}
\description{
Wraps \code{\link[base:scale]{base::scale()}} for global centering with \code{\link[=tsl_transform]{tsl_transform()}}.
}
\examples{
x <- zoo_simulate()

y <- f_center(
  x = x
)

if(interactive()){
  zoo_plot(x)
  zoo_plot(y)
}
}
\keyword{transformation}
