% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_center}
\alias{f_center}
\title{Data Transformation: Data Centering by Column}
\usage{
f_center(x = NULL, center = TRUE, scale = FALSE)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{center}{(optional, logical or numeric vector) Triggers centering if TRUE. Default: TRUE}

\item{scale}{(optional, logical or numeric vector) Triggers scaling if TRUE. Default: FALSE}
}
\value{
zoo object
}
\description{
Wraps \code{\link[base:scale]{base::scale()}} for global centering with \code{\link[=tsl_transform]{tsl_transform()}}.
}
\examples{
x <- zoo_simulate()

y <- f_center(
  x = x
)

if(interactive()){
  zoo_plot(x)
  zoo_plot(y)
}
}
\seealso{
Other data_transformation: 
\code{\link{f_detrend_difference}()},
\code{\link{f_detrend_linear}()},
\code{\link{f_hellinger}()},
\code{\link{f_list}()},
\code{\link{f_pca}()},
\code{\link{f_percentage}()},
\code{\link{f_proportion}()},
\code{\link{f_rescale}()},
\code{\link{f_scale}()},
\code{\link{f_slope}()},
\code{\link{f_smooth_window}()},
\code{\link{f_trend_linear}()}
}
\concept{data_transformation}
