% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_detrend_difference}
\alias{f_detrend_difference}
\title{Differencing Detrending of Time Series}
\usage{
f_detrend_difference(x = NULL, lag = 1, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{lag}{(optional, integer)}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
Transformed zoo object.
}
\description{
Differencing detrending via [diff()]. Returns randomm fluctuations from sample to sample not related to the overall trend of the time series.
}
