% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_hellinger}
\alias{f_hellinger}
\title{Hellinger Transformation}
\usage{
f_hellinger(x, ...)
}
\arguments{
\item{x}{Data frame or matrix.}
}
\value{
Data frame or matrix
}
\description{
Hellinger Transformation
}
