% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_hellinger}
\alias{f_hellinger}
\title{Hellinger Transformation}
\usage{
f_hellinger(x = NULL, pseudozero = 1e-04, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{pseudozero}{(required, numeric) Small number above zero to replace zeroes with.}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
zoo object
}
\description{
Hellinger Transformation
}
\examples{
x <- zoo_simulate(
  cols = 5,
  data_range = c(0, 500)
  )

y <- f_hellinger(
  x = x
)

if(interactive()){
  zoo_plot(x)
  zoo_plot(y)
}
}
