% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_list}
\alias{f_list}
\title{Lists Available Transformation Functions}
\usage{
f_list()
}
\value{
character vector
}
\description{
Lists Available Transformation Functions
}
\examples{
f_list()
}
\seealso{
Other data_transformation: 
\code{\link{f_center}()},
\code{\link{f_detrend_difference}()},
\code{\link{f_detrend_linear}()},
\code{\link{f_hellinger}()},
\code{\link{f_pca}()},
\code{\link{f_percentage}()},
\code{\link{f_proportion}()},
\code{\link{f_rescale}()},
\code{\link{f_scale}()},
\code{\link{f_slope}()},
\code{\link{f_smooth_window}()},
\code{\link{f_trend_linear}()}
}
\concept{data_transformation}
