% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_pca}
\alias{f_pca}
\title{Principal Components of a Time Series}
\usage{
f_pca(x = NULL, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
zoo object
}
\description{
Uses \code{\link[stats:prcomp]{stats::prcomp()}} to compute the Principal Component Analysis of a time series and return the principal components instead of the original columns. Output columns are named "PC1", "PC2" and so on.
}
\examples{
x <- zoo_simulate(cols = 2)

y <- f_pca(
  x = x
)

if(interactive()){
  zoo_plot(x)
  zoo_plot(y)
}
}
\concept{transformations}
