% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_pca}
\alias{f_pca}
\title{Data Transformation: Principal Components of Zoo Time Series}
\usage{
f_pca(x = NULL, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
zoo object
}
\description{
Uses \code{\link[stats:prcomp]{stats::prcomp()}} to compute the Principal Component Analysis of a time series and return the principal components instead of the original columns. Output columns are named "PC1", "PC2" and so on.
}
\examples{
x <- zoo_simulate(cols = 2)

y <- f_pca(
  x = x
)

if(interactive()){
  zoo_plot(x)
  zoo_plot(y)
}
}
\seealso{
Other tsl_transformation: 
\code{\link{f_chi_squared}()},
\code{\link{f_clr}()},
\code{\link{f_detrend_difference}()},
\code{\link{f_detrend_linear}()},
\code{\link{f_hellinger}()},
\code{\link{f_list}()},
\code{\link{f_log}()},
\code{\link{f_percentage}()},
\code{\link{f_proportion}()},
\code{\link{f_proportion_sqrt}()},
\code{\link{f_rescale_global}()},
\code{\link{f_rescale_local}()},
\code{\link{f_scale_global}()},
\code{\link{f_scale_local}()},
\code{\link{f_slope}()},
\code{\link{f_trend_linear}()}
}
\concept{tsl_transformation}
