% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_percentage}
\alias{f_percentage}
\title{Transformation to Percentage}
\usage{
f_percentage(x = NULL, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
zoo object
}
\description{
Transformation to Percentage
}
\examples{
x <- zoo_simulate(cols = 2)

y <- f_percentage(
  x = x
)

if(interactive()){
  zoo_plot(x)
  zoo_plot(y)
}
}
\seealso{
Other data_processing: 
\code{\link{f_center}()},
\code{\link{f_detrend_difference}()},
\code{\link{f_detrend_linear}()},
\code{\link{f_hellinger}()},
\code{\link{f_list}()},
\code{\link{f_pca}()},
\code{\link{f_proportion}()},
\code{\link{f_rescale}()},
\code{\link{f_scale}()},
\code{\link{f_slope}()},
\code{\link{f_smooth_window}()},
\code{\link{f_trend_linear}()},
\code{\link{tsl_aggregate}()},
\code{\link{tsl_resample}()},
\code{\link{tsl_split}()},
\code{\link{tsl_subset}()},
\code{\link{tsl_to_df}()},
\code{\link{tsl_transform}()},
\code{\link{zoo_aggregate}()},
\code{\link{zoo_permute}()},
\code{\link{zoo_resample}()}
}
\concept{data_processing}
