% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_percentage}
\alias{f_percentage}
\title{Transformation to Percentage}
\usage{
f_percentage(x)
}
\arguments{
\item{x}{Data frame or matrix.}
}
\value{
Data frame or matrix
}
\description{
Transformation to Percentage
}
