% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_proportion}
\alias{f_proportion}
\title{Data Transformation: Convert Values to Proportions by Row}
\usage{
f_proportion(x = NULL, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
zoo object
}
\description{
Data Transformation: Convert Values to Proportions by Row
}
\examples{
x <- zoo_simulate(cols = 2)

y <- f_proportion(
  x = x
)

if(interactive()){
  zoo_plot(x)
  zoo_plot(y)
}
}
\seealso{
Other tsl_transformation: 
\code{\link{f_detrend_difference}()},
\code{\link{f_detrend_linear}()},
\code{\link{f_hellinger}()},
\code{\link{f_list}()},
\code{\link{f_pca}()},
\code{\link{f_percentage}()},
\code{\link{f_rescale}()},
\code{\link{f_scale_global}()},
\code{\link{f_scale_local}()},
\code{\link{f_slope}()},
\code{\link{f_smooth_window}()},
\code{\link{f_trend_linear}()}
}
\concept{tsl_transformation}
