% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_proportion}
\alias{f_proportion}
\title{Transformation to Proportions}
\usage{
f_proportion(x = NULL, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
zoo object
}
\description{
Transformation to Proportions
}
\examples{
x <- zoo_simulate(cols = 2)

y <- f_proportion(
  x = x
)

if(interactive()){
  zoo_plot(x)
  zoo_plot(y)
}

}
