% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_proportion}
\alias{f_proportion}
\title{Transformation to Proportions}
\usage{
f_proportion(x)
}
\arguments{
\item{x}{Data frame or matrix.}
}
\value{
Data frame or matrix
}
\description{
Transformation to Proportions
}
