% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_rescale}
\alias{f_rescale}
\title{Rescales Zoo Object to a New Range}
\usage{
f_rescale(x = NULL, new_min = 0, new_max = 1, old_min = NULL, old_max = NULL)
}
\arguments{
\item{x}{(required, zoo object) Numeric vector. Default: `NULL`}

\item{new_min}{(optional, numeric) New minimum value. Default: `0`}

\item{new_max}{(optional_numeric) New maximum value. Default: `1`}

\item{old_min}{(optional, numeric) Old minimum value. Default: `NULL`}

\item{old_max}{(optional_numeric) Old maximum value. Default: `NULL`}
}
\value{
Zoo object
}
\description{
Rescales Zoo Object to a New Range
}
