% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_scale}
\alias{f_scale}
\title{Centering and scaling}
\usage{
f_scale(x = NULL, center = TRUE, scale = TRUE, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{center}{(optional, logical or numeric vector) if center is TRUE then centering is done by subtracting the column means of x from their corresponding columns.}

\item{scale}{(optional, logical or numeric vector) if scale is TRUE, and center is TRUE, then the scaling is done by dividing each column by their standard deviation. If center is FALSE, the each column is divided by their root mean square.}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
zoo object
}
\description{
Wraps \code{\link[base:scale]{base::scale()}} for global centering and scaling with \code{\link[=tsl_transform]{tsl_transform()}}.
}
\examples{
x <- zoo_simulate()

y <- f_scale(
  x = x
)

if(interactive()){
  zoo_plot(x)
  zoo_plot(y)
}
}
