% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_scale}
\alias{f_scale}
\title{Global Centering and scaling}
\usage{
f_scale(x = NULL, center = TRUE, scale = TRUE, ...)
}
\arguments{
\item{x}{Data frame or matrix.}
}
\value{
Data frame or matrix
}
\description{
Centers and scales using the global average  and standard deviation of all zoo objects within a time series list. Please use [base::scale()] to center and scale each time series independently.
}
