% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_slope}
\alias{f_slope}
\title{Slope of Time Series via Linear Regression}
\usage{
f_slope(x = NULL)
}
\arguments{
\item{x}{(required, numeric vector) input vector. Default: NULL}
}
\value{
Slope of the numeric vector
}
\description{
Calculates the slope of a numeric vector using linear regression. If the length of the input vector is 1, it returns 0. If linear regression fails, it returns NA.

Useful for [tsl_transform()] or [tsl_aggregate()].
}
\examples{
# Numeric vector of any length
f_slope(x = cumsum(runif(10)))

# Numeric vector with length = 1
f_slope(x = 10)

# Numeric vector with length = 0
f_slope(x = numeric(0))

}
