% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_slope}
\alias{f_slope}
\title{Linear Slope of Time Series}
\usage{
f_slope(x = NULL, ...)
}
\arguments{
\item{x}{(required, numeric vector) input vector. Default: NULL}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
numeric value
}
\description{
Designed for usage in \code{\link[=tsl_aggregate]{tsl_aggregate()}}. Uses linear regression to compute and return the slope of a numeric vector. Returns 0 if the length of the input vector is 1 or the linear regression fails.
}
\examples{
# Numeric vector of any length
f_slope(x = cumsum(runif(10)))

# Numeric vector with length = 1
f_slope(x = 10)

# Numeric vector with length = 0
f_slope(x = numeric(0))
}
\seealso{
Other data_processing: 
\code{\link{f_center}()},
\code{\link{f_detrend_difference}()},
\code{\link{f_detrend_linear}()},
\code{\link{f_hellinger}()},
\code{\link{f_list}()},
\code{\link{f_pca}()},
\code{\link{f_percentage}()},
\code{\link{f_proportion}()},
\code{\link{f_rescale}()},
\code{\link{f_scale}()},
\code{\link{f_smooth_window}()},
\code{\link{f_trend_linear}()},
\code{\link{tsl_aggregate}()},
\code{\link{tsl_resample}()},
\code{\link{tsl_split}()},
\code{\link{tsl_subset}()},
\code{\link{tsl_to_df}()},
\code{\link{tsl_transform}()},
\code{\link{zoo_aggregate}()},
\code{\link{zoo_permute}()},
\code{\link{zoo_resample}()}
}
\concept{data_processing}
