% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_slope}
\alias{f_slope}
\title{Data Transformation: Slope of Linear Regression with Time}
\usage{
f_slope(x = NULL, ...)
}
\arguments{
\item{x}{(required, numeric vector) input vector. Default: NULL}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
numeric value
}
\description{
Designed for usage in \code{\link[=tsl_aggregate]{tsl_aggregate()}}. Uses linear regression to compute and return the slope of a numeric vector. Returns 0 if the length of the input vector is 1 or the linear regression fails.
}
\examples{
# Numeric vector of any length
f_slope(x = cumsum(runif(10)))

# Numeric vector with length = 1
f_slope(x = 10)

# Numeric vector with length = 0
f_slope(x = numeric(0))
}
\seealso{
Other tsl_transformation: 
\code{\link{f_detrend_difference}()},
\code{\link{f_detrend_linear}()},
\code{\link{f_hellinger}()},
\code{\link{f_list}()},
\code{\link{f_pca}()},
\code{\link{f_percentage}()},
\code{\link{f_proportion}()},
\code{\link{f_rescale_global}()},
\code{\link{f_rescale_local}()},
\code{\link{f_scale_global}()},
\code{\link{f_scale_local}()},
\code{\link{f_trend_linear}()}
}
\concept{tsl_transformation}
