% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_smooth}
\alias{f_smooth}
\title{Moving Window Smoothing}
\usage{
f_smooth(x = NULL, w = 3, f = mean, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{w}{(required, window width) width of the window to compute the rolling statistics of the time series.}

\item{f}{(required, function) name without quotes and parenthesis of a standard function to smooth a time series. Typical examples are `mean` (default), `max`, `mean`, `median`, and `sd`. Custom functions able to handle zoo objects or matrices are also allowed. Default: `mean`.}

\item{...}{(optional, additional arguments) additional arguments to `f` used as argument `...` in [zoo::rollapply()].}
}
\value{
Transformed zoo object.
}
\description{
Moving Window Smoothing
}
