% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_smooth}
\alias{f_smooth}
\title{Moving Window Smoothing}
\usage{
f_smooth(x = NULL, w = 3, f = mean, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{w}{(required, window width) width of the window to compute the rolling statistics of the time series.}

\item{f}{(required, function) name without quotes and parenthesis of a standard function to smooth a time series. Typical examples are \code{mean} (default), \code{max}, \code{mean}, \code{median}, and \code{sd}. Custom functions able to handle zoo objects or matrices are also allowed. Default: \code{mean}.}

\item{...}{(optional, additional arguments) additional arguments to \code{f} used as argument \code{...} in \code{\link[zoo:rollapply]{zoo::rollapply()}}.}
}
\value{
Transformed zoo object.
}
\description{
Moving Window Smoothing
}
