% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_smooth}
\alias{f_smooth}
\title{Rolling Mean Smoothing}
\usage{
f_smooth(x = NULL, w = 3, ...)
}
\arguments{
\item{x}{(required, zoo object).}

\item{w}{(required, window width) width of the window to compute the rolling mean of the time series.}
}
\value{
Data frame or matrix
}
\description{
Rolling Mean Smoothing
}
