% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{f_trend_linear}
\alias{f_trend_linear}
\title{Linear Trend of a Time Series}
\usage{
f_trend_linear(x = NULL, center = TRUE, ...)
}
\arguments{
\item{x}{(required, zoo object) Zoo time series object to transform.}

\item{center}{(required, logical) If TRUE, the output is centered at zero. If FALSE, it is centered at the data mean. Default: TRUE}

\item{...}{(optional, additional arguments) Ignored in this function.}
}
\value{
Transformed zoo object.
}
\description{
Fits a linear model on each column of a zoo object using time as a predictor, and predicts the outcome.
}
