% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fagus_coordinates}
\alias{fagus_coordinates}
\title{Site Coordinates of Fagus sylvatica Stands}
\format{
sf data frame with 3 rows and 4 columns
}
\usage{
data(fagus_coordinates)
}
\description{
Site Coordinates of Fagus sylvatica Stands
}
\seealso{
Other example_data: 
\code{\link{albatross}},
\code{\link{cities_coordinates}},
\code{\link{cities_temperature}},
\code{\link{covid_counties}},
\code{\link{covid_prevalence}},
\code{\link{eemian_coordinates}},
\code{\link{eemian_pollen}},
\code{\link{fagus_dynamics}},
\code{\link{honeycomb_climate}},
\code{\link{honeycomb_polygons}}
}
\concept{example_data}
\keyword{datasets}
