% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fagus_dynamics}
\alias{fagus_dynamics}
\title{Enhanced Vegetation Index and Climate of Three Fagus sylvatica stands}
\format{
data frame with 5 columns and 648 rows.
}
\usage{
data(fagus_dynamics)
}
\description{
A data frame with 648 rows representing enhanced vegetation index, rainfall and teperature in three stands of Fagus sylvatica in Spain, Germany, and Sweden.
}
\details{
Site coordinates for this dataset are in \link{fagus_coordinates}.
}
\examples{
data("fagus_dynamics")

#to time series list
fagus <- tsl_initialize(
  x = fagus_dynamics,
  id_column = "site",
  time_column = "date"
)

#time series plot
if(interactive()){

 tsl_plot(
  tsl = fagus,
  scale = TRUE
  )

}

}
\keyword{datasets}
