% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leastCostPath.R
\name{leastCostPath}
\alias{leastCostPath}
\title{Extracts the least cost from a least cost matrix.}
\usage{
leastCost(least.cost.matrix = NULL)
}
\arguments{
\item{least.cost.matrix}{numeric matrix or list of numeric matrices produced by \code{\link{leastCostMatrix}}.}
}
\value{
A list if \code{least.cost.matrix} is a list, or a numeric vector if \code{least.cost.matrix} is a matrix.
}
\description{
Extracts the minimum cost (cell in the lower right corner) of a least cost matrix.
}
\examples{

}
