% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mis}
\alias{mis}
\title{Pollen Counts of 12 Marine Isotope Stages}
\format{
data frame with 8 columns and 427 rows.
}
\usage{
data(mis)
}
\description{
A data frame with 427 rows representing pollen counts for 12 marine isotope stages and 6 pollen types
}
\examples{
#to time series list
tsl <- tsl_initialize(
  x = mis,
  id_column = "mis",
  time_column = "sample_order"
)

#time series plot
if(interactive()){

 tsl_plot(
  tsl = tsl_subset(
    tsl = tsl,
    names = 1:3
    ),
  guide_columns = 2
  )

}

}
\seealso{
Other example_data: 
\code{\link{albatross}},
\code{\link{cities_coordinates}},
\code{\link{cities_temperature}},
\code{\link{covid_coordinates}},
\code{\link{covid_prevalence}},
\code{\link{eemian_coordinates}},
\code{\link{eemian_pollen}},
\code{\link{fagus_coordinates}},
\code{\link{fagus_dynamics}},
\code{\link{tsl_simulate}()},
\code{\link{zoo_simulate}()}
}
\concept{example_data}
\keyword{datasets}
