% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Plot Distance or Cost Matrix and Least Cost Path}
\usage{
plot_matrix(
  m = NULL,
  color = NULL,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  text_cex = 1,
  path = NULL,
  path_width = 1,
  path_color = "black",
  guide = TRUE,
  subpanel = FALSE
)
}
\arguments{
\item{m}{(required, numeric matrix) distance or cost matrix generated by [psi_dist_matrix()] or [psi_cost_matrix()], but any numeric matrix will work. Default: NULL}

\item{color}{(optional, character vector) vector of colors. Uses the palette "Zissou 1" by default. Default: NULL}

\item{title}{(optional, character string) plot title. By default, names of the sequences used to compute the matrix `m`. Default: NULL}

\item{xlab}{(optional, character string) title of the x axis (matrix columns). By default, the name of one of the sequences used to compute the matrix `m`. Default: NULL}

\item{ylab}{(optional, character string) title of the y axis (matrix rows). By default, the name of one of the sequences used to compute the matrix `m`. Default: NULL}

\item{text_cex}{(optional, numeric) multiplicator of the text size for the plot labels and titles. Default: 1}

\item{path}{(optional, data frame) least cost path generated with [psi_cost_path()]. This data frame must have the attribute `type == "cost_path`, and must have been computed from the same sequences used to compute the matrix `m`. Default: NULL.}

\item{path_width}{(optional, numeric) width of the least-cost path. Default: 1}

\item{path_color}{(optional, character string) color of the least-cost path. Default: "black"}

\item{guide}{(optional, logical) if TRUE, a color guide for the matrix `m` is added by [utils_matrix_guide()].}

\item{subpanel}{(optional, logical) internal argument used when generating the multipanel plot produced by [plot_distantia()].}
}
\value{
A plot
}
\description{
This function is a simplified version of [fields::imagePlot()], by [Douglas Nychka](https://dnychka.github.io/). The original version is recommended in case more customization than the provided here is needed.
}
\examples{

library(distantia)

load("~/Dropbox/GITHUB/R_packages/distantia/data/sequencesMIS.RData")

#data frame with grouping column
###################################
data(sequencesMIS)

#prepare list of sequences
x <- tsl_prepare(
  x = sequencesMIS,
  id_column = "MIS",
  time_column = NULL,
  paired_samples = FALSE,
  pseudo_zero =  0.001,
  na_action = "to_zero"
)

#distance matrix of the first two sequences
dist_matrix <- psi_dist_matrix(
  x = x[[1]],
  y = x[[2]],
  distance = "euclidean"
)

#cost matrix
cost_matrix <- psi_cost_matrix(
  dist_matrix = dist_matrix
  )

#least cost path
path <- psi_cost_path(
  dist_matrix = dist_matrix,
  cost_matrix = cost_matrix
)

#plot cost matrix and least cost path
plot_matrix(
  m = cost_matrix,
  path = path,
  guide = TRUE
  )

}
