% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ab.R
\name{prepare_ab}
\alias{prepare_ab}
\title{Prepare Two Sequences for Dissimilarity Analysis}
\usage{
prepare_ab(a = NULL, b = NULL, distance = "euclidean", paired_samples = FALSE)
}
\arguments{
\item{a}{(required, data frame, matrix, or numeric vector) a time series.}

\item{b}{(required, data frame, matrix, or numeric vector) a time series.}

\item{distance}{(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}

\item{paired_samples}{(optional, logical vector) If TRUE, time-series are compared row wise and no least-cost path is computed. Default: FALSE.}
}
\value{
A list containing the prepared sequences'a' and 'b'.
}
\description{
Prepares two sequences 'a' and 'b' for dissimilarity analysis. Please make sure that 'a' and 'b' have matching column names, as the output will only contain their common columns.
}
\details{
This function checks and preprocesses the input data 'a' and 'b' for compatibility and consistency in preparation for dissimilarity analysis.
}
\examples{

ab <- prepare_ab(
 a = na.omit(sequenceA),
 b = na.omit(sequenceB),
 distance = "euclidean",
 paired_samples = FALSE
 )

 a <- ab$a
 b <- ab$b

}
