% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_zoo_list.R
\name{prepare_zoo_list}
\alias{prepare_zoo_list}
\title{Prepare List of Data Frames for Dissimilarity Analysis}
\usage{
prepare_zoo_list(
  x = NULL,
  time_column = NULL,
  pseudo_zero = NULL,
  na_action = NULL,
  paired_samples = FALSE
)
}
\arguments{
\item{x}{(required, list of data frames) A named list with data frames. Default: NULL.}

\item{time_column}{(optional if `paired_samples = FALSE`, and required otherwise, column name) Name of the column representing time, if any. Default: NULL.}

\item{pseudo_zero}{(optional, numeric) Value used to replace zeros in the data. Default: NULL.}

\item{na_action}{(optional, character string) Action to handle missing values. default: NULL.
\itemize{
  \item NULL: returns the input without changes.
  \item "omit": applies [na.omit()] to each sequence.
  \item "to_zero": replaces NA values with zero or `pseudo_zero`, if provided.
  \item "impute": not implemented yet.
}}

\item{paired_samples}{(optional, logical) If TRUE, all input sequences are subset to their common times according to the values in the `time_column`.}
}
\value{
A named list of data frames, matrices, or vectors.
}
\description{
Prepare List of Data Frames for Dissimilarity Analysis
}
\examples{
data(sequencesMIS)
x <- prepare_sequences(
  x = sequencesMIS,
  id_column = "MIS"
)
}
