% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_zoo_list.R
\name{prepare_zoo_list}
\alias{prepare_zoo_list}
\title{Prepare List of Data Frames for Dissimilarity Analysis}
\usage{
prepare_zoo_list(x = NULL, time_column = NULL, lock_step = FALSE)
}
\arguments{
\item{x}{(required, list of data frames) A named list with data frames. Default: NULL.}

\item{time_column}{(optional if `lock_step = FALSE`, and required otherwise, column name) Name of the column representing time, if any. Default: NULL.}

\item{lock_step}{(optional, logical) If TRUE, all input sequences are subset to their common times according to the values in the `time_column`.}
}
\value{
A named list of data frames, matrices, or vectors.
}
\description{
Prepare List of Data Frames for Dissimilarity Analysis
}
\examples{
data(sequencesMIS)
x <- tsl_prepare(
  x = sequencesMIS,
  id_column = "MIS"
)
}
