% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi}
\alias{psi}
\title{Normalized Dissimilarity Score}
\usage{
psi(path_sum = NULL, auto_sum = NULL, diagonal = FALSE)
}
\arguments{
\item{path_sum}{(required, numeric) Result of \code{\link[=path_sum]{path_sum()}} on the least cost path between the sequences to compare.}

\item{auto_sum}{(required, numeric) Result of \code{\link[=psi_auto_sum]{psi_auto_sum()}} on the sequences to compare. Default: NULL}

\item{diagonal}{(optional, logical) If the cost matrix and least cost path were computed using \code{diagonal = TRUE}, this argument should be \code{TRUE} as well. Used to correct the computation of \code{psi} when diagonals are used.}
}
\value{
Psi value.
}
\description{
Computes the dissimilarity metric \code{psi} (Birks and Gordon 1985).
}
\details{
The measure of dissimilarity \code{psi} is computed as: \code{least.cost - (autosum of sequences)) / autosum of sequences}. It has a lower limit at 0, while there is no upper limit.
}
\examples{


d <- psi_dist_matrix(
  x,
  y
)

m <- psi_cost_matrix(
  dist_matrix = d
)

path <- psi_cost_path(
  dist_matrix = d,
  cost_matrix = m
)

path_sum <- path_sum(path)

xy_sum <- psi_auto_sum(
  x = x,
  y = y,
  path = path
)

psi <- psi(
  path_sum = path_sum,
  auto_sum = xy_sum
)

}
\author{
Blas Benito \href{mailto:blasbenito@gmail.com}{blasbenito@gmail.com}

\itemize{
\item Birks, H.J.B.  and Gordon, A.D. (1985) Numerical Methods in Quaternary Pollen Analysis. Academic Press.
}
}
