% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_auto_sum.R
\name{psi_auto_sum}
\alias{psi_auto_sum}
\title{Auto Sum}
\usage{
psi_auto_sum(y = NULL, x = NULL, path = NULL, distance = "euclidean")
}
\arguments{
\item{y}{(required, data frame) a sequence. Default: NULL.}

\item{x}{(required, data frame) a sequence. Default: NULL.}

\item{path}{(required, data frame) dataframe produced by [psi_cost_path()]. Only required when [psi_cost_path_trim_blocks()] has been applied to `path`. Default: NULL.}

\item{distance}{(optional, character string) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset `distances`. Default: "euclidean".}
}
\value{
Named vector with the auto sums of `y` and `x`.
}
\description{
Computes the sum of distances between consecutive samples in two multivariate time-series under comparison. Required to compute the measure of dissimilarity \code{psi} (Birks and Gordon 1985).
}
\examples{
data(sequenceA, sequenceB)

y <- sequenceA |>
  na.omit() |>
  as.matrix()

x <- sequenceB |>
  na.omit() |>
  as.matrix()

xy_sum <- psi_auto_sum(
  y = y,
  x = x,
  distance = "manhattan"
)

xy_sum

}
