% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_auto_sum.R
\name{psi_auto_sum}
\alias{psi_auto_sum}
\title{Auto Sum}
\usage{
psi_auto_sum(y = NULL, x = NULL, path = NULL, distance = "euclidean")
}
\arguments{
\item{y}{(required, data frame) a zoo time series or numeric matrix. Default: NULL.}

\item{x}{(required, data frame) a zoo time series or numeric matrix. Default: NULL.}

\item{path}{(required, data frame) dataframe produced by \code{\link[=psi_cost_path]{psi_cost_path()}}. Only required when \code{\link[=psi_cost_path_trim_blocks]{psi_cost_path_trim_blocks()}} has been applied to \code{path}. Default: NULL.}

\item{distance}{(optional, character string) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset \code{distances}. Default: "euclidean".}
}
\value{
Named vector with the auto sums of \code{y} and \code{x}.
}
\description{
Computes the sum of distances between consecutive samples in two multivariate time-series under comparison. Required to compute the measure of dissimilarity \code{psi} (Birks and Gordon 1985).
}
\examples{
#simulate two time series
tsl <- tsl_simulate(
  n = 2
)

if(interactive()){
  tsl_plot(tsl = tsl)
}

#step by step computation of psi

#common distance method
dist_method <- "euclidean"

#distance matrix
d <- psi_dist_matrix(
  x = tsl[[1]],
  y = tsl[[2]],
  distance = dist_method
)

if(interactive()){
  utils_matrix_plot(m = d)
}

#cost matrix
m <- psi_cost_matrix(
  dist_matrix = d
)

if(interactive()){
  utils_matrix_plot(m = m)
}

#least cost path
path <- psi_cost_path(
  dist_matrix = d,
  cost_matrix = m
)

if(interactive()){
  utils_matrix_plot(
    m = m,
    path = path
    )
}

#sum of least cost path
path_sum <- psi_cost_path_sum(path = path)

#auto sum of the time series
xy_sum <- psi_auto_sum(
  x = tsl[[1]],
  y = tsl[[2]],
  path = path,
  distance = dist_method
)

#dissimilarity
psi <- psi(
  path_sum = path_sum,
  auto_sum = xy_sum
)

psi

#full computation in one line
distantia(
  tsl = tsl
)$psi

#dissimilarity plot
distantia_plot(
  tsl = tsl
)
}
