% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psi_auto_sum_cpp}
\alias{psi_auto_sum_cpp}
\title{Auto Sum of Two Sequences}
\usage{
psi_auto_sum_cpp(x, y, path, distance = "euclidean", ignore_blocks = FALSE)
}
\arguments{
\item{x}{(required, numeric matrix) of same number of columns as 'y'.}

\item{y}{(required, numeric matrix) of same number of columns as 'x'.}

\item{path}{(required, data frame) dataframe produced by [cost_path()].
Default: NULL}

\item{distance}{(optional, character string) name or abbreviation of the
distance method. Valid values are in the columns "names" and "abbreviation"
of the dataset `distances`. Default: "euclidean".}

\item{ignore_blocks}{(optional, logical). If TRUE, blocks of consecutive path
coordinates are trimmed to avoid inflating the psi distance. Default: FALSE.}
}
\value{
Auto sum of matrices y and x.
}
\description{
Cumulative sum times two of two sequences \code{y} and \code{x}
with the same number of columns and arbitrary numbers of rows to compute the
ABwithin component of the psi dissimilarity computation. This component is
used to normalize the least cost distance between the sequences.
}
