% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_cost_matrix.R
\name{psi_cost_matrix}
\alias{psi_cost_matrix}
\title{Cost Matrix}
\usage{
psi_cost_matrix(dist_matrix = NULL, diagonal = TRUE, weighted = TRUE)
}
\arguments{
\item{dist_matrix}{(required, numeric matrix). Distance matrix generated by \code{\link[=psi_dist_matrix]{psi_dist_matrix()}}. Default: NULL}

\item{diagonal}{(optional, logical). If \code{TRUE}, diagonals are included in the computation of the cost matrix Default: TRUE}

\item{weighted}{(optional, logical). Only relevant if \code{diagonal = TRUE}. If \code{TRUE}, weights diagonal cost by a factor of 1.414214. Default: TRUE}
}
\value{
numeric matrix
}
\description{
Demonstration function to compute a cost matrix from a distance matrix.
}
\examples{
#simulate two time series
tsl <- tsl_simulate(
  n = 2,
  seed = 1
)

if(interactive()){
  tsl_plot(tsl = tsl)
}

#distance matrix
dist_matrix <- psi_dist_matrix(
  x = tsl[[1]],
  y = tsl[[2]],
  distance = "euclidean"
)

#cost matrix
cost_matrix <- psi_cost_matrix(
  dist_matrix = dist_matrix
)

if(interactive()){
  utils_matrix_plot(
    m = cost_matrix
    )
}
}
\seealso{
Other psi_demo: 
\code{\link{distance}()},
\code{\link{distances}},
\code{\link{psi}()},
\code{\link{psi_auto_distance}()},
\code{\link{psi_auto_sum}()},
\code{\link{psi_cost_path}()},
\code{\link{psi_cost_path_ignore_blocks}()},
\code{\link{psi_cost_path_sum}()},
\code{\link{psi_dist_lock_step}()},
\code{\link{psi_dist_matrix}()}
}
\concept{psi_demo}
