% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_cost_matrix.R
\name{psi_cost_matrix}
\alias{psi_cost_matrix}
\title{Cost Matrix}
\usage{
psi_cost_matrix(dist_matrix = NULL, diagonal = FALSE, weighted = FALSE)
}
\arguments{
\item{dist_matrix}{(required, numeric matrix). Distance matrix.}

\item{diagonal}{(optional, logical). If `TRUE`, diagonals are included in the computation of the cost matrix Default: FALSE.}

\item{weighted}{(optional, logical). Only relevant if `diagonal = TRUE`. If `TRUE`, weights diagonal cost by a factor of 1.414214.}
}
\value{
A cost matrix.
}
\description{
Computes the cost matrix from a distance matrix.
}
\examples{

distance.matrix <- psi_dist_matrix(
  x = x,
  y = y,
  distance = "manhattan"
)

cost.matrix <- psi_cost_matrix(
  dist_matrix = distance.matrix,
  diagonal = FALSE
  )

if(interactive()){
 plotMatrix(distance.matrix = cost.matrix)
}


}
