% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_cost_path_ignore_blocks.R
\name{psi_cost_path_ignore_blocks}
\alias{psi_cost_path_ignore_blocks}
\title{Trims Blocks from a Least Cost Path}
\usage{
psi_cost_path_ignore_blocks(path = NULL)
}
\arguments{
\item{path}{(required, data frame) least cost path produced by \code{\link[=psi_cost_path]{psi_cost_path()}}. Default: NULL}
}
\value{
data frame
}
\description{
In orthogonal least cost paths (when diagonals are ignored), long straight segments (a.k.a "blocks") may appear in highly dissimilar sections of the time series. Blocks inflate psi dissimilarity scores, making pairs of time series seem more dissimilar than they actually are. This demonstration function identifies and removes blocks, resulting in fairer dissimilarity analyses.
}
\examples{
#simulate two time series
tsl <- tsl_simulate(
  n = 2,
  seed = 1
)

if(interactive()){
  tsl_plot(tsl = tsl)
}

#distance matrix
dist_matrix <- psi_dist_matrix(
  x = tsl[[1]],
  y = tsl[[2]],
  distance = "euclidean"
)

#orthogonal least cost matrix
cost_matrix <- psi_cost_matrix(
  dist_matrix = dist_matrix,
  diagonal = FALSE
)

#orthogonal least cost path
cost_path <- psi_cost_path(
  dist_matrix = dist_matrix,
  cost_matrix = cost_matrix,
  diagonal = FALSE
)

nrow(cost_path)

#notice several blocks in the lower left corner
if(interactive()){
  utils_matrix_plot(
    m = cost_matrix,
    path = cost_path
  )
}

#removing blocks
cost_path_trimmed <- psi_cost_path_ignore_blocks(
  path = cost_path
)

#resulting data frame is much smaller
#rows with repeated x or y coordinates removed
nrow(cost_path_trimmed)

#but the path shape is preserved
if(interactive()){
  utils_matrix_plot(
    m = cost_matrix,
    path = cost_path_trimmed
  )
}

#psi score with blocks
distantia(
  tsl = tsl,
  diagonal = FALSE,
  ignore_blocks = FALSE
)$psi

#psi score without blocks
distantia(
  tsl = tsl,
  diagonal = FALSE,
  ignore_blocks = TRUE
)$psi
}
\seealso{
Other psi_demo: 
\code{\link{psi_auto_distance}()},
\code{\link{psi_auto_sum}()},
\code{\link{psi_cost_matrix}()},
\code{\link{psi_cost_path}()},
\code{\link{psi_cost_path_sum}()},
\code{\link{psi_dist_matrix}()}
}
\concept{psi_demo}
