% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_cost_path_sum.R
\name{psi_cost_path_sum}
\alias{psi_cost_path_sum}
\title{Sum of Distances in Least-cost Path}
\usage{
psi_cost_path_sum(path = NULL)
}
\arguments{
\item{path}{(required, data frame) dataframe produced by \code{\link[=psi_cost_path]{psi_cost_path()}}. Default: NULL}
}
\value{
Sum of distances.
}
\description{
Sum of Distances in Least-cost Path
}
\examples{


d <- distance_matrix(x, y, distance = "euclidean")

m <- cost_matrix(dist_matrix = d)

path <- psi_cost_path(
  dist_matrix = d,
  cost_matrix = m
  )

cost_path_sum(path = path)

}
