% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_cost_path_sum.R
\name{psi_cost_path_sum}
\alias{psi_cost_path_sum}
\title{Sum of Distances in Least Cost Path}
\usage{
psi_cost_path_sum(path = NULL)
}
\arguments{
\item{path}{(required, data frame) least cost path produced by \code{\link[=psi_cost_path]{psi_cost_path()}}. Default: NULL}
}
\value{
numeric value
}
\description{
Demonstration function to sum the distances of a least cost path.
}
\examples{
#simulate two time series
tsl <- tsl_simulate(
  n = 2,
  seed = 1
)

if(interactive()){
  tsl_plot(tsl = tsl)
}

#distance matrix
dist_matrix <- psi_dist_matrix(
  x = tsl[[1]],
  y = tsl[[2]],
  distance = "euclidean"
)

#cost matrix
cost_matrix <- psi_cost_matrix(
  dist_matrix = dist_matrix
)

#least cost path
cost_path <- psi_cost_path(
  dist_matrix = dist_matrix,
  cost_matrix = cost_matrix
)

if(interactive()){
  utils_matrix_plot(
    m = cost_matrix,
    path = path
    )
}

#sum of distances of the least cost path
psi_cost_path_sum(
  path = cost_path
  )
}
\seealso{
Other psi_demo: 
\code{\link{psi_auto_distance}()},
\code{\link{psi_auto_sum}()},
\code{\link{psi_cost_matrix}()},
\code{\link{psi_cost_path}()},
\code{\link{psi_cost_path_ignore_blocks}()},
\code{\link{psi_dist_matrix}()}
}
\concept{psi_demo}
