% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psi_formula_cpp}
\alias{psi_formula_cpp}
\title{Psi Dissimilarity Metric}
\usage{
psi_formula_cpp(path, auto_sum, diagonal = TRUE)
}
\arguments{
\item{path}{(required, data frame) dataframe produced by \code{\link[=cost_path_cpp]{cost_path_cpp()}}.
Default: NULL}

\item{auto_sum}{(required, numeric) auto sum of both sequences,
result of \code{\link[=psi_auto_sum_cpp]{psi_auto_sum_cpp()}}.}

\item{diagonal}{(optional, logical). If TRUE, diagonals are included in the
computation of the cost matrix. Default: FALSE.}
}
\value{
Numeric, psi dissimilarity
}
\description{
Computes the psi dissimilarity score between two sequences from
their least cost path and their auto sums.
}
