% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psi_formula_cpp}
\alias{psi_formula_cpp}
\title{Psi Dissimilarity Metric}
\usage{
psi_formula_cpp(path, auto_sum, diagonal = FALSE)
}
\arguments{
\item{path}{(required, data frame) dataframe produced by [cost_path()].
Default: NULL}

\item{diagonal}{(optional, logical). If TRUE, diagonals are included in the
computation of the cost matrix. Default: FALSE.}

\item{autosum}{(required, numeric) auto sum of both sequences,
result of psi_auto_sum_cpp().}
}
\value{
Numeric, psi dissimilarity
}
\description{
Computes the psi dissimilarity score between two sequences from
their least cost path and their auto sums.
}
