% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_aggregate.R
\name{tsl_aggregate}
\alias{tsl_aggregate}
\title{Aggregates Time Series List}
\usage{
tsl_aggregate(tsl = NULL, breaks = NULL, f = mean, quiet = TRUE, ...)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{stat}{(required, function) name without quotes and parenthesis of a standard function to smooth a time series. Typical examples are `mean` (default), `max`,`min`, `median`, and `sd`. Default: `mean`.}
}
\value{
time series list
}
\description{
Aggregates Time Series List
}
