% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_aggregate.R
\name{tsl_aggregate}
\alias{tsl_aggregate}
\title{Aggregates Time Series List}
\usage{
tsl_aggregate(tsl = NULL, breaks = NULL, f = mean, ...)
}
\arguments{
\item{tsl}{(required, list of zoo objects) List of time series. Default: NULL}

\item{breaks}{(required, numeric, numeric vector, or keyword) definition of the aggregation groups. There are several options:
\itemize{
  \item keyword: Only when time in tsl is either a date "YYYY-MM-DD" or a datetime "YYYY-MM-DD hh-mm-ss". Valid options are "year", "quarter", "month", and "week" for date, and, "day", "hour", "minute", and "second" for datetime.
}}

\item{stat}{(required, function) name without quotes and parenthesis of a standard function to smooth a time series. Typical examples are `mean` (default), `max`,`min`, `median`, and `sd`. Default: `mean`.}
}
\value{
time series list
}
\description{
Aggregates Time Series List
}
