% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_colnames.R
\name{tsl_colnames}
\alias{tsl_colnames}
\title{Column Names of Zoo Objects in Time Series List}
\usage{
tsl_colnames(
  tsl = NULL,
  names = c("all", "shared", "exclusive"),
  test_valid = TRUE
)
}
\arguments{
\item{tsl}{(required, time series list) Individual time series or time series list created with [tsl_initialize]. Default: NULL}

\item{names}{(optional, character string) Three different sets of column names can be requested:
\itemize{
  \item "all" (default): list with the column names in each zoo object in `tsl`.
  \item "shared": character vector with the shared column names in `tsl`.
  \item "exclusive": list with names of exclusive columns in each zoo object in `tsl`.
}}

\item{test_valid}{(optional, logical) If FALSE, the validity test performed by [tsl_is_valid()] is ignored. Useful to improve performance when the user knows that `tsl` is valid. Default: TRUE}
}
\value{
character vector of common column names if `common = TRUE`, or list with character vector of column names otherwise.
}
\description{
Column Names of Zoo Objects in Time Series List
}
