% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_colnames.R
\name{tsl_colnames_set}
\alias{tsl_colnames_set}
\title{Renames Columns of Zoo Objects in a Time Series List}
\usage{
tsl_colnames_set(tsl = NULL, names = NULL)
}
\arguments{
\item{tsl}{(required, time series list) Individual time series or time series list created with [tsl_initialize]. Default: NULL}

\item{names}{(required, list) Named list. List names should match old column names in `tsl`, and each named item should contain a character string with the new name. For example, `colnames = list(old_name = "new_name")` changes the name of the column "old_name" to "new_name".}
}
\description{
Renames Columns of Zoo Objects in a Time Series List
}
