% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_colnames.R
\name{tsl_colnames_set}
\alias{tsl_colnames_set}
\title{Renames Columns of Zoo Objects in a Time Series List}
\usage{
tsl_colnames_set(tsl = NULL, new_names = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{new_names}{(required, list) Named list. List names should match old column names in \code{tsl}, and each named item should contain a character string with the new name. For example, \code{colnames = list(old_name = "new_name")} changes the name of the column "old_name" to "new_name".}
}
\value{
time series list
}
\description{
Renames Columns of Zoo Objects in a Time Series List
}
\examples{
#generate example data
tsl <- tsl_simulate(cols = 3)

#list all column names
tsl_colnames_get(
  tsl = tsl,
  names = "all"
)

#rename columns
tsl <- tsl_colnames_set(
  tsl = tsl,
  new_names = list(
    a = "new_name_1",
    b = "new_name_1",
    c = "new_name_3"
  )
)

#check result
tsl_colnames_get(
  tsl = tsl,
  names = "all"
)
}
\seealso{
Other data_preparation: 
\code{\link{tsl_colnames_clean}()},
\code{\link{tsl_diagnose}()},
\code{\link{tsl_handle_NA}()},
\code{\link{tsl_initialize}()},
\code{\link{tsl_names_clean}()},
\code{\link{tsl_names_set}()},
\code{\link{tsl_repair}()},
\code{\link{zoo_to_tsl}()}
}
\concept{data_preparation}
