% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_count_NA.R
\name{tsl_count_NA}
\alias{tsl_count_NA}
\title{Counts NA and Inf Cases in Time Series}
\usage{
tsl_count_NA(tsl = NULL, verbose = TRUE)
}
\arguments{
\item{tsl}{(required, time series list) Individual time series or time series list created with [tsl_initialize]. Default: NULL}

\item{verbose}{(optional, logical) If FALSE, all messages are suppressed. Default: TRUE}
}
\value{
Integer, sum of NA cases
}
\description{
Counts NA and Inf Cases in Time Series
}
