% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_count_NA.R
\name{tsl_count_NA}
\alias{tsl_count_NA}
\title{Counts NA Cases in Time Series}
\usage{
tsl_count_NA(tsl = NULL, test_valid = TRUE, verbose = TRUE)
}
\arguments{
\item{tsl}{(required, time series list) Individual time series or time series list created with [tsl_initialize]. Default: NULL}

\item{test_valid}{(optional, logical) If FALSE, the validity test performed by [tsl_is_valid()] is ignored. Useful to improve performance when the user knows that `tsl` is valid. Default: TRUE}

\item{verbose}{(optional, logical) If FALSE, all messages are suppressed. Default: TRUE}
}
\value{
Integer, sum of NA cases
}
\description{
Counts NA Cases in Time Series
}
