% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_count_NA.R
\name{tsl_count_NA}
\alias{tsl_count_NA}
\title{Counts NA Cases in Time Series Lists}
\usage{
tsl_count_NA(tsl = NULL, quiet = FALSE)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{quiet}{(optional, logical) If TRUE, all messages are suppressed. Default: FALSE}
}
\value{
list
}
\description{
Converts Inf, -Inf, and NaN to NA (via \code{\link[=tsl_Inf_to_NA]{tsl_Inf_to_NA()}} and \code{\link[=tsl_NaN_to_NA]{tsl_NaN_to_NA()}}), and counts the total number of NA cases in each time series.
}
\examples{
#tsl with no NA cases
tsl <- tsl_simulate()

tsl_count_NA(tsl = tsl)

#tsl with NA cases
tsl <- tsl_simulate(
  na_fraction = 0.3
)

tsl_count_NA(tsl = tsl)

#tsl with variety of empty cases
tsl <- tsl_simulate()
tsl[[1]][1, 1] <- Inf
tsl[[1]][2, 1] <- -Inf
tsl[[1]][3, 1] <- NaN
tsl[[1]][4, 1] <- NaN

tsl_count_NA(tsl = tsl)
}
\seealso{
Other data_exploration: 
\code{\link{tsl_colnames_get}()},
\code{\link{tsl_names_get}()},
\code{\link{tsl_ncol}()},
\code{\link{tsl_nrow}()},
\code{\link{tsl_plot}()},
\code{\link{tsl_stats}()},
\code{\link{tsl_time}()},
\code{\link{zoo_plot}()},
\code{\link{zoo_time}()}
}
\concept{data_exploration}
