% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_diagnose.R
\name{tsl_diagnose}
\alias{tsl_diagnose}
\title{Validate Time Series List}
\usage{
tsl_diagnose(tsl = NULL, full = FALSE)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{full}{(optional, logical) If TRUE, the function runs checks looking at all the individual values of the time series list. This makes the test slower, but more comprehensive. Default: FALSE}
}
\value{
time series list
}
\description{
Validity assessment of time series lists. The required features for a valid time series list are:
\itemize{
\item Argument \code{tsl} is a list.
\item List \code{tsl} has unique names.
\item All elements in \code{tsl} are \code{zoo} objects.
\item All \code{zoo} objects have the attribute "name".
\item All \code{zoo} objects have unique values in the attribute "name".
\item Names of the list slots and the \code{zoo} objects are the same.
\item All \code{zoo} objects have at least one shared column name.
\item All columns in the \code{zoo} objects are numeric.
\item All \code{zoo} objects have zero NA cases.
}

This function analyzes a time series list and returns it without modification, but issues a warning if any condition is not met, and an explanation on how to solve the issue.
}
\examples{
#tsl with NA cases
tsl <- tsl_simulate(
  na_fraction = 0.25
)

#removing list names
names(tsl) <- NULL

#run validation test
tsl <- tsl_diagnose(
  tsl = tsl,
  full = TRUE #to examine individual data values
)
#two issues identified!

#addressing names issue
tsl <- tsl_names_set(
  tsl = tsl
)

#addressing NA issue
tsl <- tsl_handle_NA(
  tsl = tsl
)

#validation test again
tsl <- tsl_diagnose(
  tsl = tsl,
  full = TRUE
)
}
\seealso{
Other tsl_integrity: 
\code{\link{tsl_handle_NA}()},
\code{\link{tsl_initialize}()},
\code{\link{tsl_repair}()}
}
\concept{tsl_integrity}
