% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_handle_NA.R
\name{tsl_handle_NA}
\alias{tsl_handle_NA}
\alias{tsl_Inf_to_NA}
\alias{tsl_NaN_to_NA}
\title{Handle NA and Inf data in Time Series}
\usage{
tsl_handle_NA(
  tsl = NULL,
  na_action = "none",
  na_fill = 0.0001,
  f = zoo::na.approx,
  verbose = TRUE
)

tsl_Inf_to_NA(tsl = NULL)

tsl_NaN_to_NA(tsl = NULL)
}
\arguments{
\item{tsl}{(required, list of zoo objects) List of time series. Default: NULL}

\item{na_action}{(required, character) Action to handle NA data in \code{x}. Current options are:
\itemize{
\item "none" (default): No transformation is applied. If this option is selected and the data has NA cases, the function returns a warning.
\item "impute" : NA cases are interpolated against time using the function introduced in the argument \code{f}.
\item "omit": rows with NA cases are removed.
\item "fill": NA cases are replaced with the value in \code{na_fill}.
}}

\item{na_fill}{(optional, numeric) Only relevant when \code{na_action = "fill"}, defines the value used to replace NAs with. Ideally, a small number different from zero (pseudo-zero). Default: 0.0001}

\item{f}{(optional, function) Only relevant when \code{na_action = "impute"}. Function to impute missing cases. Recommended options are \code{\link[zoo:na.approx]{zoo::na.approx()}} (default), and \code{\link[zoo:na.approx]{zoo::na.spline()}}, but any function able to fill NAs in a matrix or zoo object should work. Default: zoo::na.approx}

\item{verbose}{(optional, logical) If FALSE, all messages are suppressed. Default: TRUE}
}
\value{
time series list
}
\description{
Handle NA and Inf data in Time Series
}
\examples{
TODO: complete example
}
