% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_initialize.R
\name{tsl_initialize}
\alias{tsl_initialize}
\alias{tsl_init}
\title{Prepare Data for Dissimilarity Analysis}
\usage{
tsl_initialize(
  x = NULL,
  id_column = NULL,
  time_column = NULL,
  paired_samples = FALSE,
  verbose = TRUE
)

tsl_init(
  x = NULL,
  id_column = NULL,
  time_column = NULL,
  paired_samples = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{(required, list or data frame) A named list with ordered sequences, or a long data frame with a grouping column. Default: NULL.}

\item{id_column}{(optional, column name) Column name used for splitting a 'x' data frame into a list.}

\item{time_column}{(optional if `paired_samples = FALSE`, and required otherwise, column name) Name of the column representing time, if any. Default: NULL.}

\item{paired_samples}{(optional, logical) If TRUE, all input sequences are subset to their common times according to the values in the `time_column`.}

\item{verbose}{(optional, logical) If FALSE, all messages are suppressed. Default: TRUE}
}
\value{
A named list of matrices.
}
\description{
This function performs the following steps:
- Converts a data frame to a list by 'id_column'.
- Checks the validity of input x.
- Orders x by time if a time column is PROVIDED
- Handles missing values according to the specified action.
- Handles zeros by replacing them with a pseudo-zero value, if provided.
- Transforms the data if a transformation function is provided.
}
\examples{
data(sequencesMIS)
x <- tsl_prepare(
  x = sequencesMIS,
  id_column = "MIS"
)
}
