% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_is_valid.R
\name{tsl_is_valid}
\alias{tsl_is_valid}
\title{Checks Structure Validity of a Time Series List}
\usage{
tsl_is_valid(tsl = NULL, tsl_test = FALSE)
}
\arguments{
\item{tsl}{(required, time series list) Individual time series or time series list created with [tsl_initialize]. Default: NULL}

\item{tsl_test}{(optional, logical) If TRUE, a validity test on the argument `tsl` is performed by [tsl_is_valid()]. It might be useful to set it to TRUE if something goes wrong while executing this function. Default: FALSE}
}
\value{
time series list or error.
}
\description{
Checks that the input argument is a valid time series list. The required features for a valid time series list are:
\itemize{
  \item Argument `tsl` is list.
  \item List `tsl` has unique names.
  \item List `tsl` has more than one element.
  \item All elements in `tsl` are "zoo" objects.
  \item All "zoo" objects have an attribute "name".
  \item All "zoo" objects have unique attributes "name".
  \item All "zoo" objects have at least one common column name.
  \item All "zoo" objects have zero NA cases.
}
}
