% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_names.R
\name{tsl_names}
\alias{tsl_names}
\title{Get Names of a Time Series List}
\usage{
tsl_names(tsl = NULL, test_valid = TRUE)
}
\arguments{
\item{tsl}{(required, time series list) Individual time series or time series list created with [tsl_initialize]. Default: NULL}

\item{test_valid}{(optional, logical) If FALSE, the validity test performed by [tsl_is_valid()] is ignored. Useful to improve performance when the user knows that `tsl` is valid. Default: TRUE}
}
\value{
Character vector of names.
}
\description{
A list of zoo objects created with [tsl_initialize()] has two sets of names that should ideally be the same: the list names, accessed with `names(tsl)`, and the names of the individual zoo objects (stored in their attribute "name"), accessed with `tsl_names(x)`.
}
