% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_names.R
\name{tsl_names_set}
\alias{tsl_names_set}
\title{Set Names of a Time Series List}
\usage{
tsl_names_set(tsl = NULL, names = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{names}{(optional, character vector) names to set. Must be of the same length of \code{x}. If NULL, and the list \code{x} has names, then the names of the zoo objects inside of the list are taken from the names of the list elements.}
}
\value{
time series list
}
\description{
Sets the names of a time series list and the internal names of the zoo objects inside, stored in their attribute "name".
}
\examples{
#tsl with NA cases
tsl <- tsl_simulate(
  na_fraction = 0.25
)

#removing list names
names(tsl) <- NULL

#check validity
tsl <- tsl_validate(
  tsl = tsl
)

#fix names issue
tsl <- tsl_names_set(
  tsl = tsl
)

#check validity again
tsl <- tsl_validate(
  tsl = tsl
)

#value of attribute "valid"
attributes(x = tsl)$valid

#fix NA issues
tsl <- tsl_handle_NA(
  tsl = tsl
)

#check validity again
#no errors, warnings, or messages
tsl <- tsl_validate(
  tsl = tsl
)

#value of attribute "valid"
attributes(x = tsl)$valid
}
