% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_names.R
\name{tsl_names_test}
\alias{tsl_names_test}
\title{Tests Naming Issues in Time Series Lists}
\usage{
tsl_names_test(tsl = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}
}
\value{
logical
}
\description{
Tests Naming Issues in Time Series Lists
}
\examples{
#TODO example missing, use the one in tsl_diagnose()
}
