% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_nrow.R
\name{tsl_nrow}
\alias{tsl_nrow}
\title{Number of Rows in a Time Series Lists}
\usage{
tsl_nrow(tsl = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}
}
\value{
list
}
\description{
Number of Rows in a Time Series Lists
}
\examples{
#simulate zoo time series
tsl <- tsl_simulate(
  rows = 150
  )

#count rows
tsl_nrow(
  tsl = tsl
)
}
\seealso{
Other tsl_info: 
\code{\link{tsl_colnames_get}()},
\code{\link{tsl_count_NA}()},
\code{\link{tsl_names_get}()},
\code{\link{tsl_ncol}()},
\code{\link{tsl_stats}()},
\code{\link{tsl_time}()}
}
\concept{tsl_info}
