% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_plot.R
\name{tsl_plot}
\alias{tsl_plot}
\title{Plot Multivariate Sequences}
\usage{
tsl_plot(
  tsl = NULL,
  center = FALSE,
  scale = FALSE,
  columns = 1,
  xlim = NULL,
  ylim = NULL,
  color = NULL,
  width = 1,
  text_cex = 1,
  guide = TRUE,
  guide_columns = 1,
  guide_cex = 0.8
)
}
\arguments{
\item{tsl}{(required, matrix) List of sequences generated by \code{\link[=tsl_initialize]{tsl_initialize()}}. Default: NULL}

\item{center}{(optional, logical) If TRUE, the columns of \code{x} are centered via \code{\link[=scale]{scale()}}. Default: FALSE}

\item{scale}{(optional, logical) If TRUE, the columns of \code{x} are scaled via \code{\link[=scale]{scale()}}. Default: FALSE}

\item{columns}{(optional, integer) Number of columns to organize the sequence plots into. Default: 1}

\item{xlim}{(optional, numeric vector) Numeric vector with the limits of the x axis. Applies to all sequences. Default: NULL}

\item{ylim}{(optional, numeric vector) Numeric vector with the limits of the x axis. Applies to all sequences. Default: NULL}

\item{color}{(optional, character vector) vector of colors for the distance or cost matrix. If NULL, uses an appropriate palette generated with \code{\link[grDevices:palette]{grDevices::palette.colors()}}. Default: NULL}

\item{width}{(optional, numeric vector) Widths of the sequence curves. Default: 1}

\item{text_cex}{(optional, numeric) Multiplicator of the text size. Default: 1}

\item{guide}{(optional, logical) If TRUE, plots a legend. Default: TRUE}

\item{guide_columns}{(optional, integer) Columns of the line guide. Default: 1.}

\item{guide_cex}{(optional, numeric) Size of the guide's text and separation between the guide's rows. Default: 0.7.}
}
\value{
A plot.
}
\description{
Designed to help quickly explore a set of sequences.
}
