% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_repair.R
\name{tsl_repair}
\alias{tsl_repair}
\title{Repair a Time Series List}
\usage{
tsl_repair(tsl = NULL, full = FALSE)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{full}{(optional, logical) If TRUE, the function also subsets shared numeric columns across zoo objects and interpolates NA cases. Default: FALSE}
}
\value{
time series list
}
\description{
Validity assessment of time series lists. The required features for a valid time series list are:
\itemize{
\item Argument \code{tsl} is a list.
\item List \code{tsl} has unique names.
\item All elements in \code{tsl} are \code{zoo} objects.
\item All \code{zoo} objects have the attribute "name".
\item All \code{zoo} objects have unique values in the attribute "name".
\item Names of the list slots and the \code{zoo} objects are the same.
\item All \code{zoo} objects have at least one shared column name.
\item All columns in the \code{zoo} objects are numeric.
\item All \code{zoo} objects have zero NA cases.
}

This function takes a time series list, and tries to make it comply with the rules listed above. Finally, it runs \code{\link[=tsl_diagnose]{tsl_diagnose()}} and informs the user if there are any issues left to fix.
}
\examples{
#TODO missing example here
}
\seealso{
Other tsl_integrity: 
\code{\link{tsl_diagnose}()},
\code{\link{tsl_handle_NA}()},
\code{\link{tsl_initialize}()}
}
\concept{tsl_integrity}
