% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_split.R
\name{tsl_split}
\alias{tsl_split}
\title{Splits Multivariate Time Series List into Univariate}
\usage{
tsl_split(tsl = NULL, sep = "__")
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{sep}{(required, character string) separator between the time series name and the column name.}
}
\value{
Time series list
}
\description{
This function takes a time series list with multivariate zoo objects to generate a new one with univariate zoo objects. A time series list with the the zoo objects "A" and "B", each with the columns "a", "b", and "c", becomes a time series list with the zoo objects "A__a", "A__b", "A__c", "B__a", "B__b", and "B__c". The only column of each new zoo object is named "x".
}
\examples{

tsl <- tsl_simulate(
  n = 2,
  time_range = c(
    "2010-01-01",
    "2024-12-31"
  ),
  cols = 3
)

names(tsl)

if(interactive()){
  tsl_plot(tsl = tsl)
}

tsl <- tsl_split(tsl = tsl)

names(tsl)

if(interactive()){
  tsl_plot(tsl = tsl)
}

lapply(tsl, colnames)

}
\seealso{
Other tsl_manipulation: 
\code{\link{tsl_aggregate}()},
\code{\link{tsl_colnames_clean}()},
\code{\link{tsl_colnames_set}()},
\code{\link{tsl_names_clean}()},
\code{\link{tsl_names_set}()},
\code{\link{tsl_resample}()},
\code{\link{tsl_subset}()},
\code{\link{tsl_to_df}()},
\code{\link{tsl_transform}()}
}
\concept{tsl_manipulation}
