% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_subset.R
\name{tsl_subset}
\alias{tsl_subset}
\title{Subset Time Series List}
\usage{
tsl_subset(
  tsl = NULL,
  names = NULL,
  colnames = NULL,
  time = NULL,
  test_valid = TRUE
)
}
\arguments{
\item{tsl}{(required, matrix) List of sequences generated by [tsl_initialize()]. Default: NULL}

\item{names}{(optional, character or numeric vector) Character vector of sequence names or numeric vector of sequence indices in `tsl`. If NULL, all sequences are kept. Default: NULL}

\item{colnames}{(optional, character vector) Column names of the sequences in `tsl`. If NULL, all variables are returned. Default: NULL}

\item{time}{(optional, numeric vector) numeric vector with range of times to keep from the sequences in `tsl`. Rows of sequences in `tsl` outside such range are removed. If NULL, all rows in `tsl` are preserved. Default: NULL}
}
\value{
#' @return A named list of matrices.
}
\description{
Manipulates the dimension of a time series list.
}
