% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_time_range.R
\name{tsl_time_range}
\alias{tsl_time_range}
\title{Time Range of Time Series List}
\usage{
tsl_time_range(tsl = NULL)
}
\arguments{
\item{tsl}{(required, list of zoo objects) List of time series. Default: NULL}
}
\value{
A list if each time series has a different range, and a vector of length two otherwise.
}
\description{
Time Range of Time Series List
}
