% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_time.R
\name{tsl_time_summary}
\alias{tsl_time_summary}
\title{Time Summary of a Time Series List}
\usage{
tsl_time_summary(tsl = NULL, keywords = FALSE)
}
\arguments{
\item{tsl}{(required, list of zoo objects) Time series list. Default: NULL}

\item{keywords}{(optional, logical) If TRUE, the output list contains the slot keywords with valid aggregation keywords. Default: FALSE}
}
\value{
List with class, range, units, and resolution of the time in `tsl`.
}
\description{
Time Summary of a Time Series List
}
