% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsl_transform.R
\name{tsl_transform}
\alias{tsl_transform}
\title{Title}
\usage{
tsl_transform(tsl = NULL, tsl_test = FALSE, f = NULL, ...)
}
\arguments{
\item{tsl}{(required, list of zoo objects) List of time series. Default: NULL}

\item{tsl_test}{(optional, logical) If TRUE, a validity test on the argument `tsl` is performed by [tsl_is_valid()]. It might be useful to set it to TRUE if something goes wrong while executing this function. Default: FALSE}

\item{f}{(required, transformation function) name of a function taking a matrix as input. Currently, the following options are implemented:
\itemize{
  \item f_proportion: proportion computed by row.
  \item f_percentage: percentage computed by row.
  \item f_hellinger: Hellinger transformation computed by row
  \item f_scale: Centering and scaling computed by column using the overall mean and standard deviation across all zoo objects within `tsl`.
}}

\item{...}{(optional, arguments of `f`) Optional arguments for the transformation function.}
}
\value{
time series list
}
\description{
Title
}
