% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_block_size.R
\name{utils_block_size}
\alias{utils_block_size}
\title{Logics to Handle Argument block_size in distantia()}
\usage{
utils_block_size(tsl = NULL, block_size = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{block_size}{(optional, integer) Row block sizes for the restricted permutation test. Only relevant when permutation methods are "restricted" or "restricted_by_row". A block of size \code{n} indicates that a row can only be permuted within a block of \code{n} adjacent rows. If NULL, defaults to the rounded one tenth of the shortest time_series in \code{tsl}. Default: NULL.}
}
\value{
integer
}
\description{
Logics to Handle Argument block_size in distantia()
}
\seealso{
Other internal_dissimilarity_analysis: 
\code{\link{utils_cluster_hclust_optimizer}()},
\code{\link{utils_cluster_kmeans_optimizer}()},
\code{\link{utils_cluster_silhouette}()},
\code{\link{utils_df_to_matrix}()},
\code{\link{utils_distantia_df_split}()},
\code{\link{utils_importance_df_to_wide}()},
\code{\link{utils_tsl_pairs}()}
}
\concept{internal_dissimilarity_analysis}
