% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_arguments.R
\name{utils_check_matrix_args}
\alias{utils_check_matrix_args}
\title{Checks Input Matrix}
\usage{
utils_check_matrix_args(m = NULL, arg_name = "m")
}
\arguments{
\item{m}{(required, matrix) distance or cost matrix resulting from \code{\link[=psi_dist_matrix]{psi_dist_matrix()}} or \code{\link[=psi_cost_matrix]{psi_cost_matrix()}}. Default: NULL}

\item{arg_name}{(optional, character string) name of the argument being checked. Default: NULL}
}
\value{
matrix
}
\description{
Checks Input Matrix
}
\seealso{
Other internal: 
\code{\link{utils_check_distance_args}()},
\code{\link{utils_check_distantia_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_check_path_args}()},
\code{\link{utils_check_zoo_args}()},
\code{\link{utils_digits}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()}
}
\concept{internal}
