% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_arguments.R
\name{utils_check_path_args}
\alias{utils_check_path_args}
\title{Checks Least Cost Path}
\usage{
utils_check_path_args(path = NULL, arg_name = "path")
}
\arguments{
\item{path}{(required, data frame) least cost path generated with \code{\link[=psi_cost_path]{psi_cost_path()}}. This data frame must have the attribute \verb{type == "cost_path}, and must have been computed from the same time series used to compute the matrix \code{m}. Default: NULL.}

\item{arg_name}{(optional, character string) name of the argument being checked. Default: NULL}
}
\value{
data frame
}
\description{
Checks Least Cost Path
}
\seealso{
Other internal: 
\code{\link{utils_check_distance_args}()},
\code{\link{utils_check_distantia_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_check_matrix_args}()},
\code{\link{utils_check_zoo_args}()},
\code{\link{utils_digits}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()}
}
\concept{internal}
