% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_arguments.R
\name{utils_check_zoo_args}
\alias{utils_check_zoo_args}
\title{Checks Argument x}
\usage{
utils_check_zoo_args(x = NULL, arg_name = "x")
}
\arguments{
\item{x}{(required, list of matrices) list of input matrices generated with \code{\link[=tsl_initialize]{tsl_initialize()}}. Default: NULL}

\item{arg_name}{(optional, character string) name of the argument being checked. Default: NULL}
}
\value{
zoo object
}
\description{
Checks Argument x
}
