% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clean_names.R
\name{utils_clean_names}
\alias{utils_clean_names}
\title{Clean Names}
\usage{
utils_clean_names(
  x = NULL,
  separator = "_",
  capitalize_first = FALSE,
  capitalize_all = FALSE,
  length = NULL,
  suffix = NULL,
  prefix = NULL
)
}
\arguments{
\item{x}{(required, character vector) Names to be cleaned. Default: NULL}

\item{separator}{(optional, character string) Separator when replacing spaces and dots. Also used to separate \code{suffix} and \code{prefix} from the main word. Default: "_".}

\item{capitalize_first}{(optional, logical) Indicates whether to capitalize the first letter of each name Default: FALSE.}

\item{capitalize_all}{(optional, logical) Indicates whether to capitalize all letters of each name Default: FALSE.}

\item{length}{(optional, integer) Minimum length of abbreviated names. Names are abbreviated via \code{\link[=abbreviate]{abbreviate()}}. Default: NULL.}

\item{suffix}{(optional, character string) String to append to the cleaned names. Default: NULL.}

\item{prefix}{(optional, character string)  String to prepend to the cleaned names. Default: NULL.}
}
\value{
Character vector.
}
\description{
Clean and format character vectors for use as column names or variable names.
}
\examples{

utils_clean_names(names = c("Column Name 1", "column.name.2"), capitalize_first = TRUE)

utils_clean_names(names = c("Column Name 1", "column.name.2"), capitalize_all = TRUE)
}
