% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coerce_time_class.R
\name{utils_coerce_time_class}
\alias{utils_coerce_time_class}
\title{Converts Between Classes Date and POSIXct}
\usage{
utils_coerce_time_class(x = NULL, to = "Date")
}
\arguments{
\item{x}{(required, vector of class Date or POSIXct) time vector to convert. Default: NULL}

\item{to}{(required, class name) either Date, POSIXct, or numeric. Default: Date}
}
\value{
time vector
}
\description{
Converts Between Classes Date and POSIXct
}
