% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coerce_time_class.R
\name{utils_coerce_time_class}
\alias{utils_coerce_time_class}
\title{Coerces Vector to a Given Time Class}
\usage{
utils_coerce_time_class(x = NULL, to = "Date")
}
\arguments{
\item{x}{(required, vector of class Date or POSIXct) time vector to convert. Default: NULL}

\item{to}{(required, class name) either Date, POSIXct, or numeric. Default: Date}
}
\value{
time vector
}
\description{
Coerces Vector to a Given Time Class
}
\examples{
x <- utils_coerce_time_class(
  x = c("2024-01-01", "2024-02-01"),
  to = "Date"
)

x
class(x)

x <- utils_coerce_time_class(
  x = c("2024-01-01", "2024-02-01"),
  to = "POSIXct"
)

x
class(x)

x <- utils_coerce_time_class(
  x = c("2024-01-01", "2024-02-01"),
  to = "numeric"
)

x
class(x)
}
\concept{time_handling}
\keyword{internal}
