% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coerce_time_class.R
\name{utils_coerce_time_class}
\alias{utils_coerce_time_class}
\title{Coerces Vector to a Given Time Class}
\usage{
utils_coerce_time_class(x = NULL, to = "Date")
}
\arguments{
\item{x}{(required, vector of class Date or POSIXct) time vector to convert. Default: NULL}

\item{to}{(required, class name) either Date, POSIXct, or numeric. Default: Date}
}
\value{
time vector
}
\description{
Coerces Vector to a Given Time Class
}
\examples{
x <- utils_coerce_time_class(
  x = c("2024-01-01", "2024-02-01"),
  to = "Date"
)

x
class(x)

x <- utils_coerce_time_class(
  x = c("2024-01-01", "2024-02-01"),
  to = "POSIXct"
)

x
class(x)

x <- utils_coerce_time_class(
  x = c("2024-01-01", "2024-02-01"),
  to = "numeric"
)

x
class(x)
}
\seealso{
Other internal: 
\code{\link{distantia-package}},
\code{\link{utils_block_size}()},
\code{\link{utils_check_distantia_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_clean_names}()},
\code{\link{utils_cluster_hclust_optimizer}()},
\code{\link{utils_cluster_kmeans_optimizer}()},
\code{\link{utils_cluster_silhouette}()},
\code{\link{utils_color_breaks}()},
\code{\link{utils_color_continuous_default}()},
\code{\link{utils_color_discrete_default}()},
\code{\link{utils_df_to_matrix}()},
\code{\link{utils_digits}()},
\code{\link{utils_distantia_df_split}()},
\code{\link{utils_global_scaling_params}()},
\code{\link{utils_importance_df_to_wide}()},
\code{\link{utils_is_time}()},
\code{\link{utils_line_color}()},
\code{\link{utils_line_guide}()},
\code{\link{utils_matrix_guide}()},
\code{\link{utils_matrix_plot}()},
\code{\link{utils_new_time}()},
\code{\link{utils_optimize_loess}()},
\code{\link{utils_optimize_spline}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()},
\code{\link{utils_rescale_vector}()},
\code{\link{utils_time_keywords}()},
\code{\link{utils_time_keywords_dictionary}()},
\code{\link{utils_time_keywords_translate}()},
\code{\link{utils_time_units}()},
\code{\link{utils_tsl_pairs}()}
}
\concept{internal}
