% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color_breaks.R
\name{utils_color_breaks}
\alias{utils_color_breaks}
\title{Auto Breaks for Matrix Plotting Functions}
\usage{
utils_color_breaks(m = NULL, n = 100)
}
\arguments{
\item{m}{(required, numeric matrix) distance or cost matrix generated by \code{\link[=psi_dist_matrix]{psi_dist_matrix()}} or \code{\link[=psi_cost_matrix]{psi_cost_matrix()}}, but any numeric matrix will work. Default: NULL}

\item{n}{(required, integer) number of colors to compute the breaks for. Default: 100}
}
\value{
Numeric vector
}
\description{
Auto Breaks for Matrix Plotting Functions
}
\keyword{internal}
\keyword{plotting}
