% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color_breaks.R
\name{utils_color_breaks}
\alias{utils_color_breaks}
\title{Auto Breaks for Matrix Plotting Functions}
\usage{
utils_color_breaks(m = NULL, n = 100)
}
\arguments{
\item{m}{(required, numeric matrix) distance or cost matrix generated by \code{\link[=psi_distance_matrix]{psi_distance_matrix()}} or \code{\link[=psi_cost_matrix]{psi_cost_matrix()}}, but any numeric matrix will work. Default: NULL}

\item{n}{(required, integer) number of colors to compute the breaks for. Default: 100}
}
\value{
numeric vector
}
\description{
Auto Breaks for Matrix Plotting Functions
}
\seealso{
Other internal_plotting: 
\code{\link{color_continuous}()},
\code{\link{color_discrete}()},
\code{\link{utils_line_color}()},
\code{\link{utils_line_guide}()},
\code{\link{utils_matrix_guide}()},
\code{\link{utils_matrix_plot}()}
}
\concept{internal_plotting}
