% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color_discrete_default.R
\name{utils_color_discrete_default}
\alias{utils_color_discrete_default}
\title{Default Discrete Color Palettes}
\usage{
utils_color_discrete_default(n = NULL)
}
\arguments{
\item{n}{(required, integer) number of colors (up to 36) to generate. Default: NULL}
}
\value{
Color palette
}
\description{
Uses the function \code{\link[grDevices:palette]{grDevices::palette.colors()}} to generate discrete color palettes using the following rules:
\itemize{
\item \code{n <= 9}: "Okabe-Ito".
\item \code{n == 10}: "Tableau 10"
\item \code{n > 10 && n <= 12}: "Paired"
\item \code{n > 12 && n <= 26}: "Alphabet"
\item \code{n > 26 && n <= 36}: "Polychrome 36"
}
}
\examples{

utils_color_discrete_default(n = 9)

}
