% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df_to_matrix.R
\name{utils_df_to_matrix}
\alias{utils_df_to_matrix}
\title{Data Frame to Matrix}
\usage{
utils_df_to_matrix(df = NULL, x = "x", y = "y", value = "psi")
}
\arguments{
\item{df}{(required, data frame) data frame typically resulting from \code{\link[=distantia]{distantia()}}, but others are accepted as long as the columns \code{x}, \code{y}, and \code{value} are available. Default: NULL}

\item{x}{(required, character string) Name of the column with the names of the matrix columns. Default: "x"}

\item{y}{(required, character string) Name of the column with the names of the matrix rows. Default: "y"}

\item{value}{(required, character string) Name of the column with values.}
}
\value{
Matrix of class dist
}
\description{
Internal function to convert a data frame into a matrix.
}
\examples{
tsl <- tsl_simulate(
  n = 3,
  time_range = c(
    "2010-01-01 12:00:25",
    "2024-12-31 11:15:45"
  )
)

df <- distantia(
  tsl = tsl
)

m <- utils_df_to_matrix(
  df = df
)

m
}
\keyword{internal}
