% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df_to_matrix.R
\name{utils_df_to_matrix}
\alias{utils_df_to_matrix}
\title{Data Frame to Matrix}
\usage{
utils_df_to_matrix(df = NULL, x = "x", y = "y", value = "psi")
}
\arguments{
\item{df}{(required, data frame) data frame typically resulting from \code{\link[=distantia]{distantia()}}, but others are accepted as long as the columns \code{x}, \code{y}, and \code{value} are available. Default: NULL}

\item{x}{(required, character string) Name of the column with the names of the matrix columns. Default: "x"}

\item{y}{(required, character string) Name of the column with the names of the matrix rows. Default: "y"}

\item{value}{(required, character string) Name of the column with values.}
}
\value{
dist matrix
}
\description{
Internal function to convert a data frame into a matrix.
}
\examples{
tsl <- tsl_simulate(
  n = 3,
  time_range = c(
    "2010-01-01 12:00:25",
    "2024-12-31 11:15:45"
  )
)

df <- distantia(
  tsl = tsl
)

m <- utils_df_to_matrix(
  df = df
)

m
}
\seealso{
Other internal: 
\code{\link{distantia-package}},
\code{\link{utils_block_size}()},
\code{\link{utils_check_distantia_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_clean_names}()},
\code{\link{utils_cluster_hclust_optimizer}()},
\code{\link{utils_cluster_kmeans_optimizer}()},
\code{\link{utils_cluster_silhouette}()},
\code{\link{utils_coerce_time_class}()},
\code{\link{utils_color_breaks}()},
\code{\link{utils_color_continuous_default}()},
\code{\link{utils_color_discrete_default}()},
\code{\link{utils_digits}()},
\code{\link{utils_distantia_df_split}()},
\code{\link{utils_global_scaling_params}()},
\code{\link{utils_importance_df_to_wide}()},
\code{\link{utils_is_time}()},
\code{\link{utils_line_color}()},
\code{\link{utils_line_guide}()},
\code{\link{utils_matrix_guide}()},
\code{\link{utils_matrix_plot}()},
\code{\link{utils_new_time}()},
\code{\link{utils_optimize_loess}()},
\code{\link{utils_optimize_spline}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()},
\code{\link{utils_rescale_vector}()},
\code{\link{utils_time_keywords}()},
\code{\link{utils_time_keywords_dictionary}()},
\code{\link{utils_time_keywords_translate}()},
\code{\link{utils_time_units}()},
\code{\link{utils_tsl_pairs}()}
}
\concept{internal}
