% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_digits.R
\name{utils_digits}
\alias{utils_digits}
\title{Number of Decimal Places}
\usage{
utils_digits(x = NULL)
}
\arguments{
\item{x}{(required, numeric) Default: NULL}
}
\value{
integer
}
\description{
Number of Decimal Places
}
\examples{
utils_digits(x = 0.234)
}
\seealso{
Other internal: 
\code{\link{utils_boxplot_common}()},
\code{\link{utils_check_args_distantia}()},
\code{\link{utils_check_args_matrix}()},
\code{\link{utils_check_args_path}()},
\code{\link{utils_check_args_tsl}()},
\code{\link{utils_check_args_zoo}()},
\code{\link{utils_check_distance_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_clean_names}()},
\code{\link{utils_distantia_df_split}()},
\code{\link{utils_distantia_df_to_matrix}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()},
\code{\link{utils_tsl_pairs}()}
}
\concept{internal}
