% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_distantia_df_split.R
\name{utils_distantia_df_split}
\alias{utils_distantia_df_split}
\title{Splits Distantia Data Frame in Groups}
\usage{
utils_distantia_df_split(df = NULL)
}
\arguments{
\item{df}{(required, data frame) Output of \code{\link[=distantia]{distantia()}} or \code{\link[=distantia_aggregate]{distantia_aggregate()}}. Default: NULL}
}
\value{
list
}
\description{
Internal function to split a distantia data frame by groups of the arguments 'distance', 'diagonal', 'weighted', 'ignore_blocks', and 'lock_step'.
}
\examples{
#three time series
#climate and ndvi in Fagus sylvatica stands in Spain, Germany, and Sweden
tsl <- tsl_initialize(
  x = fagus_dynamics,
  id_column = "site",
  time_column = "date"
)

#dissimilarity analysis with four combinations of parameters
df <- distantia(
  tsl = tsl,
  distance = c(
    "euclidean",
    "manhattan"
    ),
  lock_step = c(
    TRUE,
    FALSE
  )
)

#split by combinations of parameters
df_split <- utils_distantia_df_split(
  df = df
)

#print output
df_split

#class and length of the output
class(df_split)
length(df_split)
}
\seealso{
Other internal_dissimilarity_analysis: 
\code{\link{utils_block_size}()},
\code{\link{utils_cluster_hclust_optimizer}()},
\code{\link{utils_cluster_kmeans_optimizer}()},
\code{\link{utils_cluster_silhouette}()},
\code{\link{utils_df_to_matrix}()},
\code{\link{utils_importance_df_to_wide}()},
\code{\link{utils_tsl_pairs}()}
}
\concept{internal_dissimilarity_analysis}
