% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_distantia_df_split.R
\name{utils_distantia_df_split}
\alias{utils_distantia_df_split}
\title{Splits Distantia Data Frame in Groups}
\usage{
utils_distantia_df_split(df = NULL)
}
\arguments{
\item{df}{(required, data frame) Output of \code{\link[=distantia]{distantia()}} or \code{\link[=distantia_aggregate]{distantia_aggregate()}}. Default: NULL}
}
\value{
List of Distantia Data Frames
}
\description{
Internal function to split a distantia data frame by groups of the arguments 'distance', 'diagonal', 'weighted', 'ignore_blocks', and 'lock_step'.
}
