% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_global_scaling_params.R
\name{utils_global_scaling_params}
\alias{utils_global_scaling_params}
\title{Global Scaling Parameters of Time Series Lists}
\usage{
utils_global_scaling_params(tsl = NULL, f = NULL, ...)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{f}{(required, function) function name. Usually, one of \code{\link[=f_center]{f_center()}}, \code{\link[=f_scale]{f_scale()}}, or \code{\link[base:scale]{base::scale()}}. Default: NULL}

\item{...}{(optional, arguments of \code{f}) Optional arguments for the transformation function. Used only when \code{f = scale} for arguments \code{center} and \code{scale}.}
}
\value{
Named list with global mean and standard deviation
}
\description{
Internal function to compute global scaling parameters (mean and standard deviation) for time series lists. Used within \code{\link[=tsl_transform]{tsl_transform()}} when the scaling functions \code{\link[=f_center]{f_center()}}, \code{\link[=f_scale]{f_scale()}}, or \code{\link[base:scale]{base::scale()}} are used as inputs for the argument \code{f}.

Warning: this function removes exclusive columns from the data. See function \code{\link[=tsl_subset]{tsl_subset()}}.
}
\seealso{
Other internal: 
\code{\link{distantia-package}},
\code{\link{utils_block_size}()},
\code{\link{utils_check_distantia_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_clean_names}()},
\code{\link{utils_cluster_hclust_optimizer}()},
\code{\link{utils_cluster_kmeans_optimizer}()},
\code{\link{utils_cluster_silhouette}()},
\code{\link{utils_coerce_time_class}()},
\code{\link{utils_color_breaks}()},
\code{\link{utils_color_continuous_default}()},
\code{\link{utils_color_discrete_default}()},
\code{\link{utils_df_to_matrix}()},
\code{\link{utils_digits}()},
\code{\link{utils_distantia_df_split}()},
\code{\link{utils_importance_df_to_wide}()},
\code{\link{utils_is_time}()},
\code{\link{utils_line_color}()},
\code{\link{utils_line_guide}()},
\code{\link{utils_matrix_guide}()},
\code{\link{utils_matrix_plot}()},
\code{\link{utils_new_time}()},
\code{\link{utils_optimize_loess}()},
\code{\link{utils_optimize_spline}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()},
\code{\link{utils_rescale_vector}()},
\code{\link{utils_time_keywords}()},
\code{\link{utils_time_keywords_dictionary}()},
\code{\link{utils_time_keywords_translate}()},
\code{\link{utils_time_units}()},
\code{\link{utils_tsl_pairs}()}
}
\concept{internal}
\concept{transformation}
