% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_global_scaling_params.R
\name{utils_global_scaling_params}
\alias{utils_global_scaling_params}
\title{Global Scaling Parameters of Time Series Lists}
\usage{
utils_global_scaling_params(tsl = NULL, f = NULL, ...)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{f}{(required, function) function name. Usually, one of \code{\link[=f_center]{f_center()}}, \code{\link[=f_scale]{f_scale()}}, or \code{\link[base:scale]{base::scale()}}. Default: NULL}

\item{...}{(optional, arguments of \code{f}) Optional arguments for the transformation function. Used only when \code{f = scale} for arguments \code{center} and \code{scale}.}
}
\value{
Named list with global mean and standard deviation
}
\description{
Internal function to compute global scaling parameters (mean and standard deviation) for time series lists. Used within \code{\link[=tsl_transform]{tsl_transform()}} when the scaling functions \code{\link[=f_center]{f_center()}}, \code{\link[=f_scale]{f_scale()}}, or \code{\link[base:scale]{base::scale()}} are used as inputs for the argument \code{f}.

Warning: this function removes exclusive columns from the data. See function \code{\link[=tsl_remove_exclusive_cols]{tsl_remove_exclusive_cols()}}.
}
