% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_importance_df_to_wide.R
\name{utils_importance_df_to_wide}
\alias{utils_importance_df_to_wide}
\title{Importance Data Frame to Wide Format}
\usage{
utils_importance_df_to_wide(df = NULL, sep = "__")
}
\arguments{
\item{df}{(required, data frame) Output of \code{\link[=distantia_importance]{distantia_importance()}}. Default: NULL}

\item{sep}{(required, character string) Separator between the name of the importance metric and the time series variable. Default: "__".}
}
\value{
data frame
}
\description{
Takes as input a data frame returned by \code{\link[=distantia_importance]{distantia_importance()}} to return a data frame with one pair of time series per row, and the following columns:
\itemize{
\item \code{most_similar}: name of the variable with the highest contribution to similarity (most negative value in the \code{importance} column) for each pair of time series.
\itemize \code{most_dissimilar}: name of the variable with the highest contribution to dissimilarity (most positive value in the \code{importance} column) for each pair of time series.
\itemize \code{importance__variable_name}: contribution to similarity (negative values) or dissimilarity (positive values) of the given variable.
\itemize \code{psi_only_with__variable_name}: dissimilarity of the two time series when only using the given variable.
\itemize \code{psi_without__variable_name}: dissimilarity of the two time series when removing the given variable.
}
}
\examples{
#prepare time series
data("fagus_dynamics")

tsl <- tsl_initialize(
  x = fagus_dynamics,
  id_column = "site",
  time_column = "date"
) |>
  tsl_transform(
    f = f_scale
  )

#importance data frame
df <- distantia_importance(
  tsl = tsl
)
df

#to wide format
df_wide <- utils_importance_df_to_wide(
  df = df
)
df_wide

}
