% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_importance_df_to_wide.R
\name{utils_importance_df_to_wide}
\alias{utils_importance_df_to_wide}
\title{Data Frame with Contribution of Individual Variables to Dissimilarity to Wide Format}
\usage{
utils_importance_df_to_wide(df = NULL, sep = "__")
}
\arguments{
\item{df}{(required, data frame) Output of \code{\link[=momentum]{momentum()}}. Default: NULL}

\item{sep}{(required, character string) Separator between the name of the importance metric and the time series variable. Default: "__".}
}
\value{
data frame
}
\description{
Takes as input a data frame returned by \code{\link[=momentum]{momentum()}} to return a data frame with one pair of time series per row, and the following columns:
\itemize{
\item \code{most_similar}: name of the variable with the highest contribution to similarity (most negative value in the \code{importance} column) for each pair of time series.
\item \code{most_dissimilar}: name of the variable with the highest contribution to dissimilarity (most positive value in the \code{importance} column) for each pair of time series.
\item \code{importance__variable_name}: contribution to similarity (negative values) or dissimilarity (positive values) of the given variable.
\item \code{psi_only_with__variable_name}: dissimilarity of the two time series when only using the given variable.
\item \code{psi_without__variable_name}: dissimilarity of the two time series when removing the given variable.
}
}
\examples{
#prepare time series
data("fagus_dynamics")

tsl <- tsl_initialize(
  x = fagus_dynamics,
  name_column = "name",
  time_column = "time"
) |>
  tsl_transform(
    f = f_scale_global
  )

#importance data frame
df <- momentum(
  tsl = tsl
)
df

#to wide format
df_wide <- utils_importance_df_to_wide(
  df = df
)
df_wide

}
\seealso{
Other internal_dissimilarity_analysis: 
\code{\link{utils_block_size}()},
\code{\link{utils_cluster_hclust_optimizer}()},
\code{\link{utils_cluster_kmeans_optimizer}()},
\code{\link{utils_cluster_silhouette}()}
}
\concept{internal_dissimilarity_analysis}
