% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_is_time.R
\name{utils_is_time}
\alias{utils_is_time}
\title{Title}
\usage{
utils_is_time(x = NULL)
}
\arguments{
\item{x}{(required, vector) Vector to test. If the class of the vector elements is 'numeric', 'POSIXct', or 'Date', the function returns TRUE. Default: NULL.}
}
\value{
Logical
}
\description{
Title
}
