% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_line_color.R
\name{utils_line_color}
\alias{utils_line_color}
\title{Handles Line Colors for Sequence Plots}
\usage{
utils_line_color(x = NULL, line_color = NULL)
}
\arguments{
\item{x}{(required, sequence) zoo object or time series list. Default: NULL}

\item{line_color}{(optional, character vector) vector of colors for the time series columns. Selected palette depends on the number of columns to plot. Default: NULL}
}
\value{
color vector
}
\description{
This is an internal function, but can be used to better understand how line colors are handled within other plotting functions.
}
\seealso{
Other internal_plotting: 
\code{\link{color_continuous}()},
\code{\link{color_discrete}()},
\code{\link{utils_color_breaks}()},
\code{\link{utils_line_guide}()},
\code{\link{utils_matrix_guide}()},
\code{\link{utils_matrix_plot}()}
}
\concept{internal_plotting}
