% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_line_color.R
\name{utils_line_color}
\alias{utils_line_color}
\title{Handles Line Colors for Sequence Plots}
\usage{
utils_line_color(x = NULL, color = NULL)
}
\arguments{
\item{x}{(required, sequence) a single sequence or set of sequences generated via [tsl_initialize()]. Default: NULL}

\item{color}{(optional, character vector) vector of colors for the sequence columns. Selected palette depends on the number of columns to plot. Default: NULL}
}
\value{
Named vector of colors
}
\description{
This is an internal function, but can be used to better understand how line colors are handled within other plotting functions.
}
\keyword{internal}
