% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrix_guide.R
\name{utils_matrix_guide}
\alias{utils_matrix_guide}
\title{Color Guide for Matrix Plot}
\usage{
utils_matrix_guide(
  m = NULL,
  color = NULL,
  breaks = NULL,
  title = NULL,
  text_cex = 1
)
}
\arguments{
\item{m}{(required, numeric matrix) distance or cost matrix generated by [psi_dist_matrix()] or [psi_cost_matrix()], but any numeric matrix will work. Default: NULL}

\item{color}{(optional, character vector) vector of colors. Default: NULL}

\item{breaks}{(optional, numeric vector) vector of breaks for the color guide. Default: NULL}

\item{title}{(optional, character string) guide title. Default: NULL}

\item{text_cex}{(optional, numeric) multiplier for the text size. Default: 1}
}
\value{
Plot
}
\description{
Plots a color legend for a distance or cost matrix for multi-panel plots or external image editors.
}
\examples{
data(sequencesMIS)

#prepare list of sequences
x <- tsl_prepare(
  x = sequencesMIS,
  id_column = "MIS"
)

#distance matrix of the first two sequences
dist_matrix <- psi_dist_matrix(
  x = x[["MIS-3"]],
  y = x[["MIS-5"]]
)

plot_guide(m = dist_matrix)

}
