% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_optimize_gam.R
\name{utils_optimize_gam}
\alias{utils_optimize_gam}
\title{Optimization of Univariate GAM Models}
\usage{
utils_optimize_gam(x = NULL, y = NULL)
}
\arguments{
\item{x}{(required, numeric vector) predictor, usually a numeric version of a time vector.}

\item{y}{(required, numeric vector) response, usually a time series.}
}
\value{
GAM model.
}
\description{
Internal function used in [zoo_resample()]. It finds the `k` parameter of a univariate GAM formula `y ~ s(x, k = ?)` that minimizes the root mean squared error (rmse) between observations and predictions, and returns a model fitted with such `k`.
}
\examples{

#zoo time series
ts <- zoo_simulate(
  cols = 1,
  rows = 30
)

#response
y <- as.numeric(ts)
#y <- ts[, 1] #for multivariate time series

#predictor
x <- as.numeric(zoo::index(ts))

#optimize k
m <- utils_optimize_gam(
  x = x,
  y = y
)

summary(m)

#plot observation
plot(
  x = zoo::index(ts),
  y = as.numeric(y),
  col = "forestgreen",
  type = "l",
  lwd = 2
  )

#plot prediction
lines(
  x = zoo::index(ts),
  y = stats::predict(m, type = "response"),
  col = "red4"
  )

}
