% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_optimize_gam.R
\name{utils_optimize_gam}
\alias{utils_optimize_gam}
\title{Optimization of Univariate GAM Models}
\usage{
utils_optimize_gam(x = NULL, y = NULL)
}
\arguments{
\item{x}{(required, numeric vector) predictor, usually a numeric version of a time vector.}

\item{y}{(required, numeric vector) response, usually a time series.}
}
\value{
GAM model.
}
\description{
Internal function used in [zoo_resample()]. It finds the `k` parameter of a univariate GAM formula `y ~ s(x, k = ?)` that minimizes the root mean squared error (rmse) between observations and predictions, and returns a model fitted with such `k`.

This function supports progress bars generated by the `progressr` package. See examples.

This function also accepts a parallelization setup via [future::plan()], but it might only be worth it for very long time series.
}
\examples{

#zoo time series
x <- zoo_simulate(
  cols = 1,
  rows = 30
)

#response
y <- x[, 1]

#predictor
x <- as.numeric(zoo::index(x))

#parallelization
#only recommended for very long time series
# future::plan(multisession)

#progress bar
progressr::handlers(global = TRUE)

#optimize k
best.gam <- utils_optimize_gam(
  x = x,
  y = y
)

summary(best.gam)
}
