% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prepare_df.R
\name{utils_prepare_df}
\alias{utils_prepare_df}
\title{Convert Data Frame to a List of Data Frames}
\usage{
utils_prepare_df(x = NULL, id_column = NULL, time_column = NULL)
}
\arguments{
\item{x}{(required, data frame) Input data frame. Default: NULL.}

\item{id_column}{(optional, column name) Column name used to split \code{x} to a list of data frames. If omitted, the split is done by column instead. Default: NULL}

\item{time_column}{(optional, column name) Name of the column representing time, if any. Default: NULL.}
}
\value{
List of data frames
}
\description{
Convert Data Frame to a List of Data Frames
}
\seealso{
Other internal: 
\code{\link{utils_check_distance_args}()},
\code{\link{utils_check_distantia_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_check_matrix_args}()},
\code{\link{utils_check_path_args}()},
\code{\link{utils_check_zoo_args}()},
\code{\link{utils_digits}()},
\code{\link{utils_prepare_matrix}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()}
}
\concept{internal}
