% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prepare_matrix.R
\name{utils_prepare_matrix}
\alias{utils_prepare_matrix}
\title{Convert Matrix to Data Frame}
\usage{
utils_prepare_matrix(x = NULL)
}
\arguments{
\item{x}{(required, matrix) Default: NULL}
}
\value{
A data frame
}
\description{
Convert Matrix to Data Frame
}
\seealso{
Other internal: 
\code{\link{utils_check_distance_args}()},
\code{\link{utils_check_distantia_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_check_matrix_args}()},
\code{\link{utils_check_path_args}()},
\code{\link{utils_check_zoo_args}()},
\code{\link{utils_digits}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()}
}
\concept{internal}
