% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prepare_matrix_list.R
\name{utils_prepare_matrix_list}
\alias{utils_prepare_matrix_list}
\title{Convert List of Matrices to List of Data Frames}
\usage{
utils_prepare_matrix_list(x = NULL)
}
\arguments{
\item{x}{(required, list of matrices) Default: NULL}
}
\value{
List of Data Frames
}
\description{
Convert List of Matrices to List of Data Frames
}
\seealso{
Other internal: 
\code{\link{utils_check_distance_args}()},
\code{\link{utils_check_distantia_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_check_matrix_args}()},
\code{\link{utils_check_path_args}()},
\code{\link{utils_check_zoo_args}()},
\code{\link{utils_digits}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()}
}
\concept{internal}
