% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prepare_matrix_list.R
\name{utils_prepare_matrix_list}
\alias{utils_prepare_matrix_list}
\title{List of Matrices to List of Data Frames}
\usage{
utils_prepare_matrix_list(x = NULL)
}
\arguments{
\item{x}{(required, list of matrices) Default: NULL}
}
\value{
List of Data Frames
}
\description{
List of Matrices to List of Data Frames
}
