% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prepare_time.R
\name{utils_prepare_time}
\alias{utils_prepare_time}
\title{Handles Time Column in List of Data Frames}
\usage{
utils_prepare_time(x = NULL, time_column = NULL, lock_step = NULL)
}
\arguments{
\item{x}{(required, named list of data frames). List with named data frames. Default: NULL.}

\item{time_column}{(optional if \code{lock_step = FALSE}, and required otherwise, column name) Name of numeric column representing time. Default: NULL.}

\item{lock_step}{(optional, logical) If TRUE, all input sequences are subset to their common times according to the values in the \code{time_column}. Default: FALSE.}
}
\value{
List of data frames
}
\description{
Handles Time Column in List of Data Frames
}
\seealso{
Other tsl_preparation: 
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()}
}
\concept{tsl_preparation}
\keyword{internal}
\keyword{time_handling}
