% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prepare_vector_list.R
\name{utils_prepare_vector_list}
\alias{utils_prepare_vector_list}
\title{List of Vectors to List of Data Frames}
\usage{
utils_prepare_vector_list(x = NULL)
}
\arguments{
\item{x}{(required, list of vectors) Default: NULL}
}
\value{
List of data frames
}
\description{
List of Vectors to List of Data Frames
}
\seealso{
Other internal_data_preparation: 
\code{\link{utils_clean_names}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_zoo_list}()}
}
\concept{internal_data_preparation}
