% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prepare_zoo_list.R
\name{utils_prepare_zoo_list}
\alias{utils_prepare_zoo_list}
\title{Convert List of Data Frames to List of Zoo Objects}
\usage{
utils_prepare_zoo_list(x = NULL, time_column = NULL, lock_step = FALSE)
}
\arguments{
\item{x}{(required, list of data frames) A named list with data frames. Default: NULL.}

\item{time_column}{(optional if \code{lock_step = FALSE}, and required otherwise, column name) Name of the column representing time, if any. Default: NULL.}

\item{lock_step}{(optional, logical) If TRUE, all input sequences are subset to their common times according to the values in the \code{time_column}.}
}
\value{
A named list of data frames, matrices, or vectors.
}
\description{
Convert List of Data Frames to List of Zoo Objects
}
\examples{
data(mis)
x <- tsl_initialize(
  x = mis,
  id_column = "mis"
)
}
\seealso{
Other internal: 
\code{\link{utils_check_distance_args}()},
\code{\link{utils_check_distantia_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_check_matrix_args}()},
\code{\link{utils_check_path_args}()},
\code{\link{utils_check_zoo_args}()},
\code{\link{utils_digits}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()}
}
\concept{internal}
