% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_keywords.R
\name{utils_time_keywords}
\alias{utils_time_keywords}
\title{Valid Aggregation Keywords}
\usage{
utils_time_keywords(tsl = NULL)
}
\arguments{
\item{tsl}{(required, time series list) Default: NULL}
}
\value{
Character string, aggregation keyword, or "none".
}
\description{
Internal function to obtain valid aggregation keywords from a zoo object or a time series list.
}
\examples{
TODO: complete example
}
