% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_keywords_dictionary.R
\name{utils_time_keywords_dictionary}
\alias{utils_time_keywords_dictionary}
\title{Dictionary of Time Keywords}
\usage{
utils_time_keywords_dictionary()
}
\value{
data frame
}
\description{
Called by \code{\link[=utils_time_keywords_translate]{utils_time_keywords_translate()}} to generate a data frame that helps translate misnamed or abbreviated time keywords, like "day", "daily", or "d", into correct ones such as "days".
}
\examples{
df <- utils_time_keywords_dictionary()
}
\seealso{
Other internal_time_handling: 
\code{\link{utils_as_time}()},
\code{\link{utils_coerce_time_class}()},
\code{\link{utils_is_time}()},
\code{\link{utils_new_time}()},
\code{\link{utils_time_keywords}()},
\code{\link{utils_time_keywords_translate}()},
\code{\link{utils_time_units}()}
}
\concept{internal_time_handling}
