% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_keywords_dictionary.R
\name{utils_time_keywords_dictionary}
\alias{utils_time_keywords_dictionary}
\title{Dictionary of Time Keywords}
\usage{
utils_time_keywords_dictionary()
}
\value{
Data frame.
}
\description{
Data frame to help translate misnamed or abbreviated time keywords, like "day", "daily", or "d", into correct ones such as "days".
}
\examples{
TODO: complete example
}
