% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_keywords_dictionary.R
\name{utils_time_keywords_dictionary}
\alias{utils_time_keywords_dictionary}
\title{Dictionary of Time Keywords}
\usage{
utils_time_keywords_dictionary()
}
\value{
data frame
}
\description{
Called by \code{\link[=utils_time_keywords_translate]{utils_time_keywords_translate()}} to generate a data frame that helps translate misnamed or abbreviated time keywords, like "day", "daily", or "d", into correct ones such as "days".
}
\examples{
df <- utils_time_keywords_dictionary()
}
