% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_keywords_translate.R
\name{utils_time_keywords_translate}
\alias{utils_time_keywords_translate}
\title{Translates The User's Time Keywords Into Valid Ones}
\usage{
utils_time_keywords_translate(keyword = NULL)
}
\arguments{
\item{keyword}{(optional, character string) A time keyword such as "day".}
}
\value{
Time keyword.
}
\description{
Internal function to translate misnamed or abbreviated keywords into valid ones. Uses [utils_time_keywords_dictionary()] as reference dictionary.
}
