% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_units.R
\name{utils_time_units}
\alias{utils_time_units}
\title{Data Frame with Supported Time Units}
\usage{
utils_time_units(all_columns = FALSE, class = NULL)
}
\arguments{
\item{all_columns}{(optional, logical) If TRUE, all columns are returned. Default: FALSE}

\item{class}{(optional, class name). Used to filter rows and columns. Accepted values are "numeric", "Date", and "POSIXct". Default: NULL}
}
\value{
data frame
}
\description{
Returns a data frame with the names of the supported time units, the classes that can handle each time unit, and a the threshold used to identify what time units can be used when aggregating a time series.
}
\examples{

df <- utils_time_units()
head(df)
}
\seealso{
Other internal: 
\code{\link{distantia-package}},
\code{\link{utils_block_size}()},
\code{\link{utils_check_distantia_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_clean_names}()},
\code{\link{utils_cluster_hclust_optimizer}()},
\code{\link{utils_cluster_kmeans_optimizer}()},
\code{\link{utils_cluster_silhouette}()},
\code{\link{utils_coerce_time_class}()},
\code{\link{utils_color_breaks}()},
\code{\link{utils_color_continuous_default}()},
\code{\link{utils_color_discrete_default}()},
\code{\link{utils_df_to_matrix}()},
\code{\link{utils_digits}()},
\code{\link{utils_distantia_df_split}()},
\code{\link{utils_global_scaling_params}()},
\code{\link{utils_importance_df_to_wide}()},
\code{\link{utils_is_time}()},
\code{\link{utils_line_color}()},
\code{\link{utils_line_guide}()},
\code{\link{utils_matrix_guide}()},
\code{\link{utils_matrix_plot}()},
\code{\link{utils_new_time}()},
\code{\link{utils_optimize_loess}()},
\code{\link{utils_optimize_spline}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()},
\code{\link{utils_rescale_vector}()},
\code{\link{utils_time_keywords}()},
\code{\link{utils_time_keywords_dictionary}()},
\code{\link{utils_time_keywords_translate}()},
\code{\link{utils_tsl_pairs}()}
}
\concept{internal}
