% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tsl_pairs.R
\name{utils_tsl_pairs}
\alias{utils_tsl_pairs}
\title{Data Frame of Time Series Pairs}
\usage{
utils_tsl_pairs(tsl = NULL, args_list = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{args_list}{(required, list) arguments to combine with the pairs of time series. Default: NULL}
}
\value{
data frame
}
\description{
Internal function used in \code{\link[=distantia]{distantia()}} and \code{\link[=distantia_importance]{distantia_importance()}} to generate a data frame with combinations of time series and function arguments.
}
\keyword{internal}
