% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tsl_pairs.R
\name{utils_tsl_pairs}
\alias{utils_tsl_pairs}
\title{Data Frame with Pairs of Time Series in Time Series Lists}
\usage{
utils_tsl_pairs(tsl = NULL, args_list = NULL)
}
\arguments{
\item{tsl}{(required, list) Time series list. Default: NULL}

\item{args_list}{(required, list) arguments to combine with the pairs of time series. Default: NULL}
}
\value{
data frame
}
\description{
Internal function used in \code{\link[=distantia]{distantia()}} and \code{\link[=momentum]{momentum()}} to generate a data frame with combinations of time series and function arguments.
}
\seealso{
Other internal: 
\code{\link{utils_boxplot_common}()},
\code{\link{utils_check_args_distantia}()},
\code{\link{utils_check_args_matrix}()},
\code{\link{utils_check_args_path}()},
\code{\link{utils_check_args_tsl}()},
\code{\link{utils_check_args_zoo}()},
\code{\link{utils_check_distance_args}()},
\code{\link{utils_check_list_class}()},
\code{\link{utils_clean_names}()},
\code{\link{utils_digits}()},
\code{\link{utils_distantia_df_split}()},
\code{\link{utils_prepare_df}()},
\code{\link{utils_prepare_matrix}()},
\code{\link{utils_prepare_matrix_list}()},
\code{\link{utils_prepare_time}()},
\code{\link{utils_prepare_vector_list}()},
\code{\link{utils_prepare_zoo_list}()}
}
\concept{internal}
