% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_aggregate.R, R/zoo_resample.R
\name{zoo_aggregate}
\alias{zoo_aggregate}
\title{Aggregates Zoo Time Series}
\usage{
zoo_aggregate(x = NULL, new_time = NULL)

zoo_aggregate(x = NULL, new_time = NULL)
}
\arguments{
\item{x}{(required, zoo object) Time series to resample Default: NULL}

\item{new_time}{(required, zoo object, or vector of the classes numeric, Date, or POSIXct) New time of the resampled zoo object. Must be of the same class of the time in `x`. If one of them is "numeric" and the other is not, an error is returned. If a zoo object is provided, its time is used as a template to resample `x`. If NULL, irregular time series are predicted into a regular version of their own time, and regular time series are returned as is. Default: NULL}

\item{breaks}{(required, vector) Vector to define time breaks for aggregation. Must be in the same time units as `zoo::index(x)`.}

\item{stat}{(required, function) name without quotes and parenthesis of a standard function to smooth a time series. Typical examples are `mean` (default), `max`,`min`, `median`, and `sd`. Default: `mean`.}
}
\value{
zoo object

zoo object
}
\description{
Given a zoo object and a time vector, this function uses Generalized Additive Models (GAM) to resample a zoo object to a new time. This operation is useful to align time series with different frequencies, or to transform irregular time series into regular.

The new time vector must be of the same class as the time of the zoo object (either  numeric, Date, or POSIXct). This new time vector can be either created from scratch, or can be extracted from another zoo object with [zoo::index()].

Each time series within the zoo object is modelled as a function of the original time of the zoo object with [mgcv::gam()], and then predicted to te new time.

Please use this operation with care, as there are limits to the amount of resampling that can be done without distorting the data. The safest option is to keep the distance between new time points within the same magnitude of the distance between the old time points.
}
