% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_time.R
\name{zoo_time}
\alias{zoo_time}
\title{Time Features of a Zoo Object}
\usage{
zoo_time(x = NULL, keywords = FALSE)
}
\arguments{
\item{x}{(required, zoo object) Default: NULL.}

\item{keywords}{(optional, logical) If TRUE, the output list contains the slot keywords with valid aggregation keywords. Default: FALSE}
}
\value{
List with time features of the zoo object.
}
\description{
Time Features of a Zoo Object
}
