% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_to_tsl.R
\name{zoo_to_tsl}
\alias{zoo_to_tsl}
\title{Wraps Zoo Object in Time Series List}
\usage{
zoo_to_tsl(x = NULL)
}
\arguments{
\item{x}{(required, zoo object) zoo time series. Default: NULL}
}
\value{
time series list of length one.
}
\description{
Internal function to wrap a zoo object into a time series list in these contexts where the objective is not to compare two time series.
}
